% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_ambient}
\alias{add_ambient}
\title{Mark ambient outliers in combined dataset}
\usage{
add_ambient(scseqs, combined)
}
\arguments{
\item{scseqs}{the original scseqs}

\item{combined}{the combined scseqs}
}
\value{
\code{combined} with \code{out_ambient} column added to \code{meta.features} slot of \code{SCT} assay.
}
\description{
A gene is marked as an ambient outlier if it is an ambient outlier in at least one of the test datasets.
}
