% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_centered}
\alias{add_centered}
\title{Add median centered sample data to scale.data}
\usage{
add_centered(scseqs, combined)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects used during sample integration to generate \code{combined}}

\item{combined}{Integrated \code{Seurat} object from \code{\link{integrate_scseq}}}
}
\value{
\code{combined} with median centered sample data in \code{scale.data} slot of \code{SCT} assay
}
\description{
For each sample, median centers each gene. Resulting matrices
are merged and added to the \code{'scale.data'} slot of the \code{'SCT'} assay in \code{combined}.
}
\details{
This is helpful to remove the effect of high background gene expression seen in one sample but not another
which will dominate cross-sample differential gene expression analyses.
}
\keyword{internal}
