% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{add_cluster_numbers}
\alias{add_cluster_numbers}
\title{Add cluster numbers to annotation}
\usage{
add_cluster_numbers(annot, pad_left = FALSE)
}
\arguments{
\item{annot}{Character vector of cluster names}

\item{pad_left}{Should cluster numbers be padded on the left? Used to align
numbers for violin plot.}
}
\value{
\code{annot} with cluster numbers pre-pended to non-numeric values.
}
\description{
Adds cluster number to non-numeric cluster names.
E.g. if \code{annot = c('Monocytes', '2')} then
\code{annot = c('1: Monocytes', '2')} is returned
}
\examples{

annot <- c('Monocytes', '2')
add_cluster_numbers(annot)

}
