% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{add_doublet_score}
\alias{add_doublet_score}
\title{Calculate doublet score for each cell}
\usage{
add_doublet_score(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object with \code{hvg} column in \code{rowData}
and \code{npcs} in \code{metadata} slot}
}
\value{
\code{scseq} with column \code{doublet_score} added to \code{colData}.
}
\description{
Calculate doublet score for each cell
}
\keyword{internal}
