% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_es.R
\name{add_es}
\alias{add_es}
\title{Add metaMA effectsize values to top table.}
\usage{
add_es(diff_exprs, cols = c("dprime", "vardprime"))
}
\arguments{
\item{diff_exprs}{Result from call to \code{\link{diff_expr}}.}

\item{cols}{Columns from \code{\link[metaMA]{effectsize}} result to add to
top table.}
}
\value{
diff_exprs with specified columns added to top_table.
}
\description{
Adds moderated unbiased standardised effect sizes (dprimes) to top table
from differential expression analysis.
}
\examples{
# location of previously saved analysis
data_dir <- system.file('extdata', 'IBD', package = 'drugseqr')

# load previous analysis for eset
anal <- readRDS(file.path(data_dir, 'diff_exprs_symbol.rds'))

# add dprime and vardprime to top tables
anal <- add_es(anal)
}
