% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{add_hvgs}
\alias{add_hvgs}
\title{Get top highly variable genes for subsequent dimensionality reduction}
\usage{
add_hvgs(sce, hvgs = NULL)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object}

\item{hvgs}{Character vector of genes to use. Used to specify a priori
set of genes to use for clustering.}
}
\description{
Runs after \code{preprocess_scseq}
}
