% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{add_norms}
\alias{add_norms}
\title{Add edgeR normalization factors to pdata.}
\usage{
add_norms(quants, pdata)
}
\arguments{
\item{quants}{\code{DGEList} with RNA-seq counts.}

\item{pdata}{\code{data.frame} with sample annotations.}
}
\value{
\code{pdata} with \code{lib.size} and \code{norm.factors} columns added.
}
\description{
Add edgeR normalization factors to pdata.
}
\keyword{internal}
