% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_pcan}
\alias{add_pcan}
\title{Add PCA reducedDim with selected number of PCs}
\usage{
add_pcan(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} with PCA \code{reducedDim} and npcs metadata.}
}
\value{
\code{sce} with PCAN reducedDim
}
\description{
Add PCA reducedDim with selected number of PCs
}
