% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{add_project_scseqs}
\alias{add_project_scseqs}
\title{Adds project name to meta.data of scseq}
\usage{
add_project_scseqs(scseqs)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects.}
}
\value{
\code{scseqs} with \code{meta.data$project} column equal to \code{project.name} slot.
}
\description{
Used to keep track of which sample is which for integrated datasets. This is used for generating pseudo-bulk
counts for each sample.
}
\keyword{internal}
