% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_qc_genes}
\alias{add_qc_genes}
\title{Add Mitochondrial and Ribosomal RNA gene Names to SingelCellExperiment}
\usage{
add_qc_genes(sce, species)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{species}{Character indicating species. Either \code{'Homo sapiens'},
or \code{'Mus musculus'}.}
}
\description{
Add Mitochondrial and Ribosomal RNA gene Names to SingelCellExperiment
}
