% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_redn}
\alias{add_redn}
\title{Add reducedDim with selected number of PCs}
\usage{
add_redn(sce, dimred = "PCA")
}
\arguments{
\item{sce}{\code{SingleCellExperiment} with \code{reducedDim} and npcs metadata.}
}
\value{
\code{sce} with REDN reducedDim
}
\description{
Add reducedDim with selected number of PCs
}
