% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_scaled}
\alias{add_scaled}
\title{Add scale.data to integrated data}
\usage{
add_scaled(scseqs, combined)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects used during sample integration to generate \code{combined}}

\item{combined}{Integrated \code{Seurat} object from \code{\link{integrate_scseq}}}
}
\value{
\code{combined} with \code{scale.data} slot of \code{SCT} assay
}
\description{
For each sample merges the \code{'scale.data'} slot of the \code{'SCT'} assay in \code{combined}.
}
\keyword{internal}
