% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_scseq_qc_metrics}
\alias{add_scseq_qc_metrics}
\title{Calculate QC metrics for SingleCellExperiment}
\usage{
add_scseq_qc_metrics(sce)
}
\arguments{
\item{sce}{}
}
\description{
Calculate QC metrics for SingleCellExperiment
}
