% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{add_vsd}
\alias{add_vsd}
\title{Add VST normalized assay data element to expression set}
\usage{
add_vsd(eset)
}
\arguments{
\item{eset}{ExpressionSet with group column in \code{pData(eset)}}
}
\value{
\code{eset} with \code{'vsd'} \code{assayDataElement} added.
}
\description{
Add VST normalized assay data element to expression set
}
