% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{advancedOptions}
\alias{advancedOptions}
\title{Logic for advanced options for selectedDrugStudy}
\usage{
advancedOptions(input, output, session, cmap_res, l1000_res, drug_study,
  show_advanced)
}
\description{
Logic for advanced options for selectedDrugStudy
}
\keyword{internal}
