% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{advancedOptionsInput}
\alias{advancedOptionsInput}
\title{advanced options input for drugs page}
\usage{
advancedOptionsInput(id)
}
\description{
advanced options input for drugs page
}
\keyword{internal}
