% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utils.R
\name{aggregate.Matrix}
\alias{aggregate.Matrix}
\title{Compute summary statistics of a Matrix}
\usage{
\method{aggregate}{Matrix}(x, groupings = NULL, form = NULL, fun = "sum", ...)
}
\description{
Compute summary statistics of a Matrix
}
