% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{ambient.omit}
\alias{ambient.omit}
\title{Omit ambient genes from analysis or top table}
\usage{
ambient.omit(scseq, markers, cluster_markers)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{anal}{Result of \code{diff_expr_scseq}}

\item{top_table}{Usually from \code{anal$top_table}}
}
\value{
One of \code{anal} or \code{top_table} with entries corresponding to ambient genes omited.
}
\description{
Must provide one of anal or top_table
}
\keyword{internal}
