% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-ui.R
\name{analysisTypeToggle}
\alias{analysisTypeToggle}
\title{Input form to control/test/all groups for integrated datasets}
\usage{
analysisTypeToggle(ns)
}
\description{
Input form to control/test/all groups for integrated datasets
}
\keyword{internal}
