% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{annot_query_res}
\alias{annot_query_res}
\title{Annotate query result}
\usage{
annot_query_res(query_res, drug_annot)
}
\arguments{
\item{query_res}{Numeric vector of correlation values with names as HGNC symbols}

\item{drug_annot}{result of \link{get_drugs_table}}
}
\value{
Annotated query table
}
\description{
Annotate query result
}
\keyword{internal}
