% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_pdata.R
\name{append_pdata}
\alias{append_pdata}
\title{Join CMAP02/L1000 pdata and query results.}
\usage{
append_pdata(query_res, study)
}
\arguments{
\item{query_res}{Named numeric vector. Result of previous call to \code{\link{query_drugs}}}

\item{study}{Character vector. Either \code{'CMAP02'} or \code{'L1000'}.}
}
\value{
\code{data.frame} with correlation column showing results from \code{query_res} and other columns giving information about each treatment.
}
\description{
Join CMAP02/L1000 pdata and query results.
}
\examples{

# load CMAP02 data
cmap_path <- system.file('extdata', 'cmap_es_ind.rds', package = 'drugseqr', mustWork = TRUE)
cmap_es <- readRDS(cmap_path)

# load previous analysis
anal_path <- file.path('data-raw/example-data', 'diff_expr_symbol.rds')
anal <- readRDS(anal_path)

# get effect size values
dprimes <- get_dprimes(anal)

# run query
res <- query_drugs(dprimes, cmap_es)

# append drug pdata
res <- append_pdata(res, 'CMAP02')

}
