% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\title{Coerce Seurat to SingleCellExperiment}
\usage{
as.SingleCellExperiment(x, assay = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
This exists because of bug satijalab/seurat#1626
}
