% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{ascpR}
\alias{ascpR}
\title{Utility function to run aspera}
\usage{
ascpR(ascpCMD, files, destDir = getwd())
}
\arguments{
\item{ascpCMD}{aspera command as string.}

\item{files}{Urls to aspera files to download.}

\item{destDir}{Path to directory to download \code{files} into.}
}
\description{
Utility function to run aspera
}
