% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{boxPlotlyCells}
\alias{boxPlotlyCells}
\title{Generate boxplotly for cell type deconvolution plots.}
\usage{
boxPlotlyCells(df, boxgap, boxgroupgap, plot_fname, ytitle, xtitle)
}
\arguments{
\item{df}{\code{data.frame} with columns: \itemize{
 \item x Factor for x-labels.
 \item y Numeric column used for y-values.
 \item text Character column used for hoverinfo.
 \item name Character column used for legend names of \code{x} values.
 \item color Factor column used to generate ordered colors for boxplots for each \code{'x'} value.
}}

\item{boxgap}{Used for plotly layout.}

\item{boxgroupgap}{Used for plotly layout.}

\item{plot_fname}{Name to save plot as.}

\item{ytitle}{Y axis title.}

\item{xtitle}{X axis title.}
}
\value{
plotly
}
\description{
Generate boxplotly for cell type deconvolution plots.
}
\keyword{internal}
