% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{build_ensdb}
\alias{build_ensdb}
\title{Build ensembldb annotation package.}
\usage{
build_ensdb(species = "Homo sapiens", version = "94")
}
\arguments{
\item{species}{Character vector indicating species. Should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{version}{EnsemblDB version. Should be same as used in \code{\link{build_index}}.}
}
\description{
Build ensembldb annotation package.
}
\examples{

# build ensembldb annotation package for human
build_ensdb()

}
