% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{build_gencode_index}
\alias{build_gencode_index}
\title{Download ensembl transcriptome and build index for salmon quantification}
\usage{
build_gencode_index(indices_dir, species = "human", release = "29",
  command = "salmon")
}
\arguments{
\item{indices_dir}{directory to place indices in.}

\item{species}{The species. Default is \code{homo_sapiens.}}

\item{release}{gencode release. Default is \code{29} (matches ensembl release 94 for \code{\link{build_ensdb_index}})).}
}
\description{
This index is used for single cell RNA-seq quantification. See \code{\link{build_ensdb_index}} for bulk RNA-seq equivalent.
}
\examples{
# build salmon alevin index for humans
indices_dir <- 'data-raw/indices'
build_gencode_index(indices_dir)

}
