% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{build_kallisto_index}
\alias{build_kallisto_index}
\title{Download ensembl transcriptome and build index for kalliso quantification}
\usage{
build_kallisto_index(indices_dir, species = "homo_sapiens",
  release = "94")
}
\arguments{
\item{indices_dir}{Directory to put kallisto indices.}

\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for AnnotationHub -
needs to match with \code{\link{build_ensdb}}).}
}
\description{
This index is used for RNA seq quantification.
}
\examples{
# build kallist index for humans
indices_dir <- 'data-raw/indices/kallisto'
build_ensdb_index(indices_dir)

}
