% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{bulkAnal}
\alias{bulkAnal}
\title{Logic for bulk data analyses}
\usage{
bulkAnal(
  input,
  output,
  session,
  data_dir,
  dataset_name,
  dataset_dir,
  explore_eset
)
}
\description{
Logic for bulk data analyses
}
\keyword{internal}
