% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{bulkAnalInput}
\alias{bulkAnalInput}
\title{Bulk Differential expression analysis input}
\usage{
bulkAnalInput(
  id,
  with_dl = TRUE,
  run_title = "Model all groups and selected surrogate variables"
)
}
\description{
Bulk Differential expression analysis input
}
\keyword{internal}
