% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkAnnot}
\alias{bulkAnnot}
\title{Logic for downloading and uploading bulk annotation}
\usage{
bulkAnnot(input, output, session, dataset_name, annot)
}
\description{
Logic for downloading and uploading bulk annotation
}
\keyword{internal}
