% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkCellsPlotly}
\alias{bulkCellsPlotly}
\title{Logic for Bulk Data cell type deconvolution plotly}
\usage{
bulkCellsPlotly(input, output, session, dtangle_est, pdata, dataset_name)
}
\description{
Logic for Bulk Data cell type deconvolution plotly
}
\keyword{internal}
