% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkDataset}
\alias{bulkDataset}
\title{Logic for selected dataset part of bulkFrom}
\usage{
bulkDataset(
  input,
  output,
  session,
  sc_dir,
  bulk_dir,
  data_dir,
  new_dataset,
  explore_eset
)
}
\description{
Logic for selected dataset part of bulkFrom
}
\keyword{internal}
