% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-ui.R
\name{bulkDatasetInput}
\alias{bulkDatasetInput}
\title{Dataset selection input for bulkFormInput}
\usage{
bulkDatasetInput(id)
}
\description{
Dataset selection input for bulkFormInput
}
\keyword{internal}
