% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkEndType}
\alias{bulkEndType}
\title{Logic for end type selection is bulkFormQuant}
\usage{
bulkEndType(input, output, session, fastq_dir, is.sc)
}
\description{
Logic for end type selection is bulkFormQuant
}
\keyword{internal}
