% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-ui.R
\name{bulkEndTypeInput}
\alias{bulkEndTypeInput}
\title{Dataset end-type input for bulkFormQuantInput}
\usage{
bulkEndTypeInput(id)
}
\description{
Dataset end-type input for bulkFormQuantInput
}
\keyword{internal}
