% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkExploreTable}
\alias{bulkExploreTable}
\title{Logic for differential expression analysis table}
\usage{
bulkExploreTable(
  input,
  output,
  session,
  eset,
  up_annot,
  data_dir,
  dataset_dir,
  dataset_name,
  svobj_r,
  numsv_r
)
}
\description{
Logic for differential expression analysis table
}
\keyword{internal}
