% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkForm}
\alias{bulkForm}
\title{Logic for Bulk Data form}
\usage{
bulkForm(
  input,
  output,
  session,
  data_dir,
  sc_dir,
  bulk_dir,
  msg_quant,
  new_anal,
  explore_eset,
  enable_sva,
  pdata
)
}
\description{
Logic for Bulk Data form
}
\keyword{internal}
