% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkFormAnal}
\alias{bulkFormAnal}
\title{Logic for differential expression analysis part of bulkForm}
\usage{
bulkFormAnal(
  input,
  output,
  session,
  data_dir,
  dataset_name,
  dataset_dir,
  explore_eset,
  numsv_r,
  svobj_r,
  enable_sva
)
}
\description{
Logic for differential expression analysis part of bulkForm
}
\keyword{internal}
