% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-ui.R
\name{bulkFormAnalInput}
\alias{bulkFormAnalInput}
\title{Differential expression analysis inputs for bulkFormInput}
\usage{
bulkFormAnalInput(id)
}
\description{
Differential expression analysis inputs for bulkFormInput
}
\keyword{internal}
