% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkGenePlotly}
\alias{bulkGenePlotly}
\title{Logic for Bulk Data gene plotly}
\usage{
bulkGenePlotly(input, output, session, eset, explore_genes, dataset_name)
}
\description{
Logic for Bulk Data gene plotly
}
\keyword{internal}
