% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkMDSplotly}
\alias{bulkMDSplotly}
\title{Logic for Bulk Data MDS plotly}
\usage{
bulkMDSplotly(
  input,
  output,
  session,
  explore_eset,
  dataset_name,
  numsv,
  mds,
  group_colors,
  adjusted
)
}
\description{
Logic for Bulk Data MDS plotly
}
\keyword{internal}
