% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkPage}
\alias{bulkPage}
\title{Logic for Bulk Data Tab}
\usage{
bulkPage(
  input,
  output,
  session,
  project_dir,
  sc_dir,
  bulk_dir,
  tx2gene_dir,
  indices_dir,
  gs_dir,
  add_bulk,
  remove_bulk
)
}
\arguments{
\item{input, output, session}{standard shiny module boilerplate}

\item{project_dir}{path to folder with project files}

\item{sc_dir}{sub folder of \code{project_dir} where single-cell data is stored}

\item{bulk_dir}{sub folder of \code{project_dir} where bulk data is stored}

\item{tx2gene_dir}{Path to directory containing transcript to gene maps
produced by \link[dseqr.data]{load_tx2gene}.}

\item{indices_dir}{Path to directory containing \code{kallisto} indices and whitelists.}

\item{gs_dir}{Path to directory where gene to Gene Ontology maps produced by \link{get_genego} are saved.}

\item{add_bulk}{reactive that triggers modal to upload a bulk dataset}

\item{remove_bulk}{reactive that triggers modal for deleting bulk datasets}
}
\value{
list with reactive \code{new_dataset} that is triggered with a new
  bulk dataset is added.
}
\description{
to be called with \link[shiny]{callModule}
}
