% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkQuantTable}
\alias{bulkQuantTable}
\title{Logic for dataset quantification table}
\usage{
bulkQuantTable(input, output, session, fastq_dir, labels, paired)
}
\description{
Logic for dataset quantification table
}
\keyword{internal}
