% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{calcx}
\alias{calcx}
\title{Calculate x position for correlation plot}
\usage{
calcx(cor, range = c(-1, 1), width = 180, pad = 0.1)
}
\arguments{
\item{cor}{Numeric vector of correlation values.}

\item{range}{Numeric vector of length 2 specifying maximum and minimum values of \code{cor}.}

\item{width}{Plot width to scale correlation values to.}

\item{pad}{Numeric value that is respectively, subtracted and added to values in \code{range}. Make it so that circles and
correlation text values don't get cut off.}
}
\value{
Numeric vector giving x position for correlation plot in Drugs tab.
}
\description{
Calculate x position for correlation plot
}
