% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{cca_integrate}
\alias{cca_integrate}
\title{Integrate with CCA from Seurat}
\usage{
cca_integrate(scseqs, genes)
}
\arguments{
\item{scseqs}{Seurat objects}

\item{genes}{Character vector of genes to integrate on}
}
\value{
Seurat object
}
\description{
Uses SCT method
}
\keyword{internal}
