% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{center_data}
\alias{center_data}
\title{Median center matrix or data.frame}
\usage{
center_data(data, dim = 1)
}
\arguments{
\item{data}{matrix or data.frame}

\item{dim}{Dimension to center. Either 1 for rows (default) or 2 for columns.}
}
\value{
\code{data} median centered along \code{dim}
}
\description{
Median center matrix or data.frame
}
\keyword{internal}
