% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{check_bulk_changed}
\alias{check_bulk_changed}
\title{Check if newly uploaded pdata is the same as previously uploaded}
\usage{
check_bulk_changed(prev, pdata)
}
\value{
\code{TRUE} is \code{pdata} column 'Group name' or 'Pair' is different
from \code{prev}
}
\description{
Check if newly uploaded pdata is the same as previously uploaded
}
\keyword{internal}
