% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pug.R
\name{check_gras}
\alias{check_gras}
\title{Check Pubchem PUG-View for GRAS notice}
\usage{
check_gras(pug_view)
}
\arguments{
\item{pug_view}{List from reading PUG-View JSON data with \code{\link[rjson]{fromJSON}}}
}
\value{
Boolean indicating if \code{pug_view} has a GRAS notice.
}
\description{
Check Pubchem PUG-View for GRAS notice
}
