% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{check_has_scseq}
\alias{check_has_scseq}
\title{Check if folder has single-cell files}
\usage{
check_has_scseq(dataset_names, sc_dir)
}
\arguments{
\item{dataset_names}{Character vector of folder names to check}

\item{sc_dir}{Path to directory with \code{dataset_names} folders}
}
\value{
Boolean vector with \code{length(dataset_names)} indicating folders that
have single-cell files
}
\description{
Check if folder has single-cell files
}
