% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{check_is_cellranger}
\alias{check_is_cellranger}
\title{Determine if the selected folder has CellRanger files}
\usage{
check_is_cellranger(data_dir)
}
\arguments{
\item{data_dir}{path to directory to check.}
}
\value{
\code{TRUE} if CellRanger files detected, otherwise \code{FALSE}.
}
\description{
Determine if the selected folder has CellRanger files
}
\keyword{internal}
