% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{check_is_market_matrix}
\alias{check_is_market_matrix}
\title{Determine if the selected folder has cellranger market matrix format files}
\usage{
check_is_market_matrix(data_dir)
}
\arguments{
\item{data_dir}{path to directory to check.}
}
\value{
\code{TRUE} if market matrix files detected, otherwise \code{FALSE}.
}
\description{
Determine if the selected folder has cellranger market matrix format files
}
\keyword{internal}
