% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kb_scseq.R
\name{clean_kb_scseq}
\alias{clean_kb_scseq}
\title{Clean up kb output directory}
\usage{
clean_kb_scseq(data_dir, sample_name = NULL)
}
\arguments{
\item{data_dir}{Path containing folder 'bus_output'}

\item{sample_name}{Optional sample name to append to cellranger files}
}
\value{
Called for side effects
}
\description{
Clean up kb output directory
}
