% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{clean_y}
\alias{clean_y}
\title{Adjusts expression data for surrogate variables.}
\usage{
clean_y(y, mod, mod.clean)
}
\arguments{
\item{y}{Expression data of eset.}

\item{mod}{Full model matrix supplied to \code{sva}.}

\item{svs}{Surrogate variables returned by \code{sva} (svobj$sv).}
}
\value{
Expression data with effects of svs removed.
}
\description{
Factors out effect of surrogate variables discovered during surrogate variable
analysis.
}
