% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{clusterComparison}
\alias{clusterComparison}
\title{Logic for cluster comparison input}
\usage{
clusterComparison(
  input,
  output,
  session,
  selected_anal,
  scseq,
  markers,
  annot_path,
  sc_dir,
  ref_preds
)
}
\description{
Logic for cluster comparison input
}
\keyword{internal}
