% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{clusterComparisonInput}
\alias{clusterComparisonInput}
\title{Input form and buttons to select a cluster or contrast and rename a cluster}
\usage{
clusterComparisonInput(id)

clusterComparisonInput(id)
}
\description{
Input form and buttons to select a cluster or contrast and rename a cluster

Input form and buttons to select a cluster or contrast and rename a cluster
}
\keyword{internal}
