% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{collapse_sorted}
\alias{collapse_sorted}
\title{Used to generate file names for single cell analyses.}
\usage{
collapse_sorted(x, collapse = ",")
}
\arguments{
\item{x}{Character vector of selected cluster numbers}
}
\value{
String with sorted cluster numbers comma collapsed.
}
\description{
Used to generate file names for single cell analyses.
}
\keyword{internal}
