% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{comparisonType}
\alias{comparisonType}
\title{Logic for comparison type toggle for integrated analyses}
\usage{
comparisonType(input, output, session, scseq)
}
\description{
Logic for comparison type toggle for integrated analyses
}
\keyword{internal}
