% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{construct_eset}
\alias{construct_eset}
\title{Construct expression set}
\usage{
construct_eset(quants, fdata, annot, txi.deseq = NULL)
}
\arguments{
\item{quants}{\code{DGEList} with RNA-seq counts.}

\item{fdata}{\code{data.table} returned from \code{\link{setup_fdata}}.}

\item{annot}{Character vector with ensembldb package name. e.g. \code{'EnsDb.Hsapiens.v94'}. Returned from \code{\link{get_ensdb_package}}.}
}
\value{
\code{ExpressionSet} with input arguments are in corresponding atributes.
}
\description{
Construct expression set
}
\keyword{internal}
