% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{construct_pbulk_eset}
\alias{construct_pbulk_eset}
\title{Run pseudo bulk limma fit}
\usage{
construct_pbulk_eset(summed, species = "Homo sapiens", release = "94")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object}
}
\description{
Run pseudo bulk limma fit
}
