% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbulk_qcsets.R
\name{construct_pbulk_qcsets}
\alias{construct_pbulk_qcsets}
\title{Construct Pseudobulk QC expression sets}
\usage{
construct_pbulk_qcsets(pbulk_esets)
}
\arguments{
\item{pbulk_esets}{List of pseudobulk expression sets}
}
\value{
List of pseodobulk qc expression sets
}
\description{
Used for testing if significant changes in e.g. sum of mtRNA or ribosomal counts
}
\keyword{internal}
