% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{construct_pbulk_subsets}
\alias{construct_pbulk_subsets}
\title{Run pseudo bulk limma fit}
\usage{
construct_pbulk_subsets(summed, method = "TMMwsp", ...)
}
\arguments{
\item{summed}{Pseudobulk \code{DietSCE}}

\item{...}{additional arguments to \link[edgeR]{filterByExpr}}
}
\value{
Normalized \code{ExpressionSet}
}
\description{
Run pseudo bulk limma fit
}
\keyword{internal}
