% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{customQueryForm}
\alias{customQueryForm}
\title{Logic for custom query form on Drugs page}
\usage{
customQueryForm(
  input,
  output,
  session,
  show_custom,
  is_custom,
  anal_name,
  new_custom,
  data_dir
)
}
\description{
Logic for custom query form on Drugs page
}
\keyword{internal}
