% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{customQueryFormInput}
\alias{customQueryFormInput}
\title{Input form for custom query on Drugs page}
\usage{
customQueryFormInput(id)
}
\value{
\code{shiny.tag}
}
\description{
Input form for custom query on Drugs page
}
\keyword{internal}
