% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{datasets_to_list}
\alias{datasets_to_list}
\title{Convert single-cell datasets data.frame to list}
\usage{
datasets_to_list(datasets)
}
\arguments{
\item{datasets}{data.frame of datasets from \link{get_sc_dataset_choices}}
}
\value{
list
}
\description{
Used to allow client side updates of single-cell dataset choices. This
is needed to prevent de-selection of current dataset after subset, intgration,
and loading.
}
\keyword{internal}
