% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{decide_ambient}
\alias{decide_ambient}
\title{Get ambient genes to exclude}
\usage{
decide_ambient(ambient, top_table, cluster_markers)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}
}
\value{
Character vector of HGNC symbols representing ambient genes to exclude
}
\description{
Excludes upregulated genes if they are ambient in the test group.
Excludes downregulated genes if they are ambient in the control group.
}
\keyword{internal}
