% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-ui.R
\name{deconvolutionFormInput}
\alias{deconvolutionFormInput}
\title{Input form for single-cell deconvolution}
\usage{
deconvolutionFormInput(id)
}
\description{
Input form for single-cell deconvolution
}
\keyword{internal}
