% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_pdata.R
\name{destructure_title}
\alias{destructure_title}
\title{Split CMAP02/L1000 title into multiple columns.}
\usage{
destructure_title(pdata, drop = TRUE, ...)
}
\arguments{
\item{pdata}{\code{data.frame} with \code{'title'} column. \code{'title'} column
should contain 4 underscore seperated parts (e.g. \code{'10-DEBC_A375_20um_24h'}).}

\item{drop}{Should the \code{'title'} column be droped? Default is \code{TRUE}.}

\item{...}{Additional arguments to \code{\link[tibble]{add_column}} (e.g. \code{.before} or \code{.after}).}
}
\value{
\code{pdata} without \code{'title'} column and 4 new columns: \code{'Compound'}, \code{'Cell Line'}, \code{'Dose'},
  and \code{'Duration'}.
}
\description{
Function used by \code{\link{append_pdata}}.
}
