% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kb_scseq.R
\name{detect_10x_chemistry}
\alias{detect_10x_chemistry}
\title{Auto Detect 10x Chemistry}
\usage{
detect_10x_chemistry(
  index_path,
  data_dir,
  fqs,
  techs = c("10XV2", "10XV3"),
  threads = 1
)
}
\arguments{
\item{index_path}{Path to kallisto index.}

\item{data_dir}{Path to folder with 10x fastq.gz files.}

\item{fqs}{Vector of fastq file names.}

\item{techs}{10x chemistries to check. Passed to kallisto -x argument.}

\item{threads}{Number of threads to use. Default is 1.}

\item{bus_args}{Additional arguments to kallisto bus.}
}
\value{
one of \code{techs} corresponding to chemistry with most reads that agree with whitelist.
}
\description{
Creates samples of 10,000 reads per lane and then determins the number of reads
with barcodes in agreement with the chemistry whitelist.
}
\keyword{internal}
