% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_qc.R
\name{detect_cells}
\alias{detect_cells}
\title{Determine non-empty droplets}
\usage{
detect_cells(counts, qcgenes)
}
\arguments{
\item{counts}{dgCMatrix of counts}
}
\value{
Indices of columns of counts corresponding to non-empty droplets
}
\description{
Determine non-empty droplets
}
\keyword{internal}
