% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pairs.R
\name{detect_paired}
\alias{detect_paired}
\title{Detect if experiment is pair-ended.}
\usage{
detect_paired(fastq_id1s)
}
\arguments{
\item{fastq_id1s}{Character vector of first sequence identifiers from fastq.gz files. Returned from \code{\link{get_fastq_id1s}}.}
}
\value{
boolean indicating if experiement is pair-ended (\code{TRUE}) or single-ended (\code{FALSE}).
}
\description{
Detect if experiment is pair-ended.
}
\keyword{internal}
