% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{detect_sc_method}
\alias{detect_sc_method}
\title{Determine Method used for Single Cell RNA Seq}
\usage{
detect_sc_method(data_dir)
}
\arguments{
\item{data_dir}{Path to folder with fastq.gz scRNA-seq files}
}
\value{
Character vector of detected method.
}
\description{
Currently just looks for 10X data.
}
