% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DietSCE.R
\docType{class}
\name{DietSCE-class}
\alias{DietSCE-class}
\title{An lighter-weight SingleCellExperiment S4 class}
\value{

}
\description{
An lighter-weight SingleCellExperiment S4 class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{list.}

\item{\code{colData}}{data.frame.}

\item{\code{reducedDims}}{list.}
}}

