% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_anal}
\alias{diff_anal}
\title{Run limma analysis.}
\usage{
diff_anal(
  eset,
  anal_name,
  exprs_sva,
  modsv,
  data_dir,
  annot = "SYMBOL",
  rna_seq = TRUE
)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}. Replicate features and
non-selected samples removed by \code{iqr_replicates}.}

\item{exprs_sva}{Expression data with surrogate variables removed. Created by
\code{iqr_replicates}}

\item{modsv}{Model matrix generated by \code{diff_setup}. With
and without surrogate variables.}

\item{data_dir}{String, path to directory with raw data.}

\item{annot}{String, either "ENTREZID" or "SYMBOL" for probe or gene level
analysis respectively. If "ENTREZID", appends "_entrezid.rds" to save name.}

\item{rna_seq}{Is the analysis for RNA-seq? Auto inferred in \code{diff_expr}.}
}
\value{
List, final result of \code{diff_expr}. Used for subsequent
  meta-analysis.
}
\description{
Runs limma differential expression analysis on all contrasts selected by
\code{add_contrast}. Analysis performed with and without surrogate
variables discovered by \code{diff_setup}. Also prints MDS plot and saves
results.
}
\seealso{
\code{\link{diff_expr}}.
}
