% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_expr}
\alias{diff_expr}
\title{Differential expression analysis of eset.}
\usage{
diff_expr(
  eset,
  data_dir,
  anal_name,
  contrast = "test-ctrl",
  annot = "SYMBOL",
  svobj = list(sv = NULL),
  num_svs = 0,
  prev_anal = NULL
)
}
\arguments{
\item{eset}{Annotated eset created by \code{\link{load_seq}}.}

\item{data_dir}{String specifying directory of GSE folders.}

\item{annot}{String, column name in fData. For duplicated
values in this column, the row with the highest interquartile range
across selected samples will be kept. Appropriate values are \code{"SYMBOL"} (default - for gene level analysis)
or \code{"ENTREZID_HS"} (for probe level analysis).}

\item{prev_anal}{Previous result of \code{\link{diff_expr}}. If present, previous group
selections will be reused.}
}
\value{
List of named lists, one for each GSE. Each named list contains:
  \item{pdata}{data.frame with phenotype data for selected samples.
     Columns \code{treatment} ('ctrl' or 'test'), \code{group}, and \code{pairs} are
     added based on user selections.}
  \item{top_tables}{List with results of \code{\link[limma]{topTable}} call (one per
     contrast). These results account for the effects of nuissance variables
     discovered by surrogate variable analysis.}
  \item{ebayes_sv}{Results of call to \code{\link[limma]{eBayes}} with surrogate
     variables included in the model matrix.}
  \item{annot}{Value of \code{annot} variable.}
}
\description{
After selecting control and test samples for a contrast, surrogate variable
analysis (\code{\link[sva]{sva}}) and differential expression analysis is performed.
}
\details{
To specify a contrast, first select rows for control samples, and click the \emph{Add Control Rows}
button. Repeat for test samples. After control and test samples have been added
click the \emph{Done} button. If you make a mistake, click the \emph{Reset} button.

Analysis results are saved in the corresponding in \code{data_dir} as "diff_expr.rds". If analyses
needs to be repeated, previous results can be reloaded with \code{\link[base]{readRDS}}
and supplied to the \code{prev_anal} parameter. In this case, previous selections will be reused.
}
\examples{

data_dir <- system.file('extdata', 'IBD', package='drugseqr', mustWork = TRUE)
eset <- load_seq(data_dir)
anal <- diff_expr(eset, data_dir, anal_name = 'IBD')
}
