% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{diff_expr_pbulk}
\alias{diff_expr_pbulk}
\title{Run pseudo bulk differential expression  analysis}
\usage{
diff_expr_pbulk(scseq, data_dir, clusters_name)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{data_dir}{Path to folder to save analysis in.}

\item{clusters_name}{String with comma seperated selected clusters.}
}
\description{
Run pseudo bulk differential expression  analysis
}
