% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{diff_expr_scseq}
\alias{diff_expr_scseq}
\title{Run limma differential expression between test and control groups in single cell RNA-Seq dataset}
\usage{
diff_expr_scseq(scseq, clusters, exclude_ambient = TRUE)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{clusters}{Character vector of clusters to include in analysis}

\item{exclude_ambient}{Boolean indicating if ambient genes should be excluded. Default is \code{TRUE}.}
}
\value{
Named list with slots: \itemize{
 \item top_table result of limma::topTable.
 \item ebayes_sv results of \code{\link{fit_ebayes_scseq}}.
 \item pdata \code{data.frame} with column \code{'group'} indicating test and control samples.
}
}
\description{
Run limma differential expression between test and control groups in single cell RNA-Seq dataset
}
\seealso{
\code{\link{get_ambient}}
}
\keyword{internal}
