% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_path.R
\name{diff_path}
\alias{diff_path}
\title{Differential expression of KEGG pathways.}
\usage{
diff_path(eset, prev_anal, data_dir, anal_name, rna_seq = TRUE)
}
\arguments{
\item{eset}{Expression set.}

\item{prev_anal}{Previous result of \code{\link{diff_expr}}.}
}
\description{
Performs PADOG pathway analysis using KEGG database.
PADOG outperforms other pathway analysis algorithms at prioritizing expected pathways (see references).
}
\examples{
eset <- readRDS('data-raw/patient_data/sjia/bulk/mono/eset.rds')
prev_anal <- readRDS('data-raw/patient_data/sjia/bulk/mono/diff_expr_symbol_ferritin.rds')
data_dir <- 'data-raw/patient_data/sjia/bulk/mono'
anal_name <- 'ferritin'

path_anal <- diff_path(eset, prev_anal, data_dir, anal_name)

}
\references{
Tarca AL, Bhatti G, Romero R. A Comparison of Gene Set Analysis Methods
   in Terms of Sensitivity, Prioritization and Specificity. Chen L, ed. PLoS ONE.
   2013;8(11):e79217. doi:10.1371/journal.pone.0079217.

   Dong X, Hao Y, Wang X, Tian W. LEGO: a novel method for gene set over-representation
   analysis by incorporating network-based gene weights. Scientific Reports.
   2016;6:18871. doi:10.1038/srep18871.
}
