% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{diff_path_scseq}
\alias{diff_path_scseq}
\title{Run PADOG pathway analysis on a single cell RNA-Seq dataset}
\usage{
diff_path_scseq(scseq, prev_anal, data_dir, anal_name, clusters)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{prev_anal}{Result of call to \code{diff_expr_scseq}}

\item{data_dir}{Directory to folders with single cell analyses}

\item{anal_name}{Name of folder in \code{data_dir} to save results to}

\item{clusters}{Character vector of clusters in \code{scseq$seurat_clusters} to include}
}
\value{
result of \code{\link[PADOG]{padog}}
}
\description{
Analysis is performed comparing test to control cells for all included clusters.
}
\keyword{internal}
