% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{dir_exists}
\alias{dir_exists}
\title{Safe dir.exists}
\usage{
dir_exists(x)
}
\arguments{
\item{x}{directory path}
}
\value{
a logical vector of TRUE or FALSE values
}
\description{
Safe dir.exists
}
