% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{disableAll}
\alias{disableAll}
\title{Disable multiple ids}
\usage{
disableAll(ids, asis = rep(FALSE, length(ids)))
}
\arguments{
\item{ids}{Character vector of ids to disable}
}
\description{
Disable multiple ids
}
\keyword{internal}
