% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{dl_10x_whitelists}
\alias{dl_10x_whitelists}
\title{Download 10X whitelists for kallisto bustools correction}
\usage{
dl_10x_whitelists(indices_dir)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices.}
}
\description{
Download 10X whitelists for kallisto bustools correction
}
\keyword{internal}
