% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_drug_es.R
\name{dl_drug_es}
\alias{dl_drug_es}
\title{Download drug effect size data.}
\usage{
dl_drug_es(files = c("cmap_es_ind.rds", "l1000_drugs_es.rds",
  "l1000_genes_es.rds"), check = FALSE)
}
\arguments{
\item{files}{Character vector of file names to download.}

\item{check}{Check that existing drug effect size data is loadable? Default is FALSE.}
}
\description{
Download drug effect size data.
}
