% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{dl_pert_pvals}
\alias{dl_pert_pvals}
\title{Download CMAP02/L1000 pert pvals from S3}
\usage{
dl_pert_pvals(sig_path, pert_type)
}
\arguments{
\item{sig_path}{Path to download file to.}

\item{pert_type}{One of \code{'cmap'}, \code{'l1000_drugs'}, or \code{'l1000_genes'}.}
}
\description{
Download CMAP02/L1000 pert pvals from S3
}
\examples{
sig_path <- file.path('data-raw/drug_es/signatures', 'BRD-K45319408_PC3_5um_24h.rds')
dl_pert_pvals(sig_path, pert_type = 'cmap')

}
