% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{dl_pert_result}
\alias{dl_pert_result}
\title{Download CMAP02/L1000 pert query result from S3}
\usage{
dl_pert_result(res_path)
}
\arguments{
\item{res_path}{Path to download file to.}
}
\description{
Download CMAP02/L1000 pert query result from S3
}
