% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-docs-ui.R
\name{docsPageUI}
\alias{docsPageUI}
\title{UI for docs page (add new sections here)}
\usage{
docsPageUI(id, tab, active)
}
\arguments{
\item{id}{Identification string that is names-paced using \link[shiny]{NS}.}

\item{tab}{Name to appear on tab}

\item{active}{Name of current active \code{tab}}
}
\value{
shiny.tag with html for docs tab
}
\description{
UI for docs page (add new sections here)
}
