% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kb_scseq.R
\name{download_kb_index}
\alias{download_kb_index}
\title{Download pre-built index using kb-python}
\usage{
download_kb_index(indices_dir, species = c("human", "mouse"))
}
\arguments{
\item{indices_dir}{directory with indices. Output files will be saved in
indices_dir/kb_version/species/}

\item{species}{either 'human' or 'mouse'}
}
\value{
Called for side effects
}
\description{
Downloads index.idx and t2g.txt.
}
