% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{downloadablePlot}
\alias{downloadablePlot}
\title{Logic for plot with downloadable data}
\usage{
downloadablePlot(
  input,
  output,
  session,
  plot_fun,
  fname_fun,
  data_fun,
  downloadable
)
}
\description{
Logic for plot with downloadable data
}
\keyword{internal}
