% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadablePlotly.R
\name{downloadablePlotly}
\alias{downloadablePlotly}
\title{Logic for plotly with download data button in modebar}
\usage{
downloadablePlotly(
  input,
  output,
  session,
  plotly_fun,
  fname_fun = function() {     NULL },
  data_fun = function() {     NULL },
  title = "Download plot data"
)
}
\arguments{
\item{plotly_fun}{function to generate base plotly}

\item{fname_fun}{function that returns filename for download}

\item{data_fun}{function that saves file for download}
}
\description{
Logic for plotly with download data button in modebar
}
\keyword{internal}
