% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadablePlotly.R
\name{downloadablePlotlyUI}
\alias{downloadablePlotlyUI}
\title{Plotly with hidden download link}
\usage{
downloadablePlotlyUI(id, height = "auto")
}
\description{
Plotly with hidden download link
}
\keyword{internal}
