% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{downloadablePlotUI}
\alias{downloadablePlotUI}
\title{UI for plot with downloadable data}
\usage{
downloadablePlotUI(id, height)

downloadablePlotUI(id, height)
}
\description{
UI for plot with downloadable data

UI for plot with downloadable data
}
\keyword{internal}
