% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{downsample_clusters}
\alias{downsample_clusters}
\title{Downsample SingleCellExperiment clusters}
\usage{
downsample_clusters(scseq, max.cells = 200)
}
\arguments{
\item{scseq}{SingleCellExperiment}

\item{max.cells}{maximum number of cells in each \code{scseq$cluster}}
}
\value{
scseq
}
\description{
Gets a maximum number of cells in each cluster. Used for generate mini
datasets for faster label transfer.
}
\keyword{internal}
