% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{downsample_group}
\alias{downsample_group}
\title{Downsample cells within each cluster across a group}
\usage{
downsample_group(scseq, group = "orig.ident")
}
\arguments{
\item{scseq}{SingleCellExperiment}
}
\value{
\code{scseq} with equal number
}
\description{
Used for sample comparison plots to show equal number of test and control
cells for each cluster
}
\keyword{internal}
