% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{downsample_scseq}
\alias{downsample_scseq}
\title{Downsample very large scseq objects for integration}
\usage{
downsample_scseq(scseq, max.cells = 1000, seed = 0L)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{max.cells}{Maximum number of cells to keep. Default is 10000.}

\item{seed}{Integer used for reproducibility.}
}
\value{
\code{scseq} with maximum \code{max.cells} cells.
}
\description{
Used by \code{load_scseqs_for_integration}.
}
\keyword{internal}
