% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{dprimesPlotly}
\alias{dprimesPlotly}
\title{Generate plotly of dprimes values for Drugs and Pathways tab}
\usage{
dprimesPlotly(path_df)
}
\arguments{
\item{path_df}{result of \link{get_path_df}.}
}
\value{

}
\description{
Generate plotly of dprimes values for Drugs and Pathways tab
}
\keyword{internal}
