% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdownButton.R
\name{dropdownButton}
\alias{dropdownButton}
\title{Modified shinyWidgets dropdownButton}
\usage{
dropdownButton(
  ...,
  circle = TRUE,
  status = "default",
  size = "default",
  icon = NULL,
  label = NULL,
  tooltip = FALSE,
  right = FALSE,
  up = FALSE,
  width = NULL,
  margin = "10px",
  inline = FALSE,
  inputId = NULL,
  title = NULL
)
}
\arguments{
\item{...}{List of tag to be displayed into the dropdown menu.}

\item{circle}{Logical. Use a circle button}

\item{status}{Add a class to the buttons, you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.
Or use an arbitrary strings to add a custom class, e.g. : with \code{status = 'myClass'}, buttons will have class \code{btn-myClass}.}

\item{size}{Size of the button : default, lg, sm, xs.}

\item{icon}{An icon to appear on the button.}

\item{label}{Label to appear on the button. If circle = TRUE and tooltip = TRUE, label is used in tooltip.}

\item{tooltip}{Put a tooltip on the button, you can customize tooltip with \code{tooltipOptions}.}

\item{right}{Logical. The dropdown menu starts on the right.}

\item{up}{Logical. Display the dropdown menu above.}

\item{width}{Width of the dropdown menu content.}

\item{margin}{Value of the dropdown margin-right and margin-left menu content.}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()}) for the output.}

\item{inputId}{Optional, id for the button, the button act like an \code{actionButton},
and you can use the id to toggle the dropdown menu server-side with \code{\link[shinyWidgets]{toggleDropdownButton}}.}

\item{title}{title attribute for button}
}
\description{
Button is wrapped in input-group-btn class to allow inline with forms. Also includes title parameter.
}
\keyword{internal}
