% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{dropdownMenuButton}
\alias{dropdownMenuButton}
\title{Dropdown menu button for dsLabelRowsUI}
\usage{
dropdownMenuButton(id, label)
}
\description{
Dropdown menu button for dsLabelRowsUI
}
\keyword{internal}
