% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{drugsGenesPlotlyOutput}
\alias{drugsGenesPlotlyOutput}
\title{UI for query/drug genes plotly}
\usage{
drugsGenesPlotlyOutput(id)
}
\description{
UI for query/drug genes plotly
}
\keyword{internal}
