% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{drugsTable}
\alias{drugsTable}
\title{Logic for drug table}
\usage{
drugsTable(input, output, session, cmap_res, l1000_res, drug_study, cells,
  show_clinical)
}
\description{
Logic for drug table
}
\keyword{internal}
