% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{drugsTableOutput}
\alias{drugsTableOutput}
\title{Output table for Drugs Page}
\usage{
drugsTableOutput(id)
}
\value{
\code{shiny.tag}
}
\description{
Output table for Drugs Page
}
\keyword{internal}
