% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsAnalTable}
\alias{dsAnalTable}
\title{Logic for differential expression analysis table}
\usage{
dsAnalTable(input, output, session, fastq_dir, labels, data_dir,
  dataset_dir, anal_name)
}
\description{
Logic for differential expression analysis table
}
\keyword{internal}
