% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsCellsPlotly}
\alias{dsCellsPlotly}
\title{Logic for Dataset Cell Type deconvolution plotly}
\usage{
dsCellsPlotly(input, output, session, dtangle_est, pdata)
}
\description{
Logic for Dataset Cell Type deconvolution plotly
}
\keyword{internal}
