% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsEndType}
\alias{dsEndType}
\title{Logic for end type selection is dsFormQuant}
\usage{
dsEndType(input, output, session, fastq_dir)
}
\description{
Logic for end type selection is dsFormQuant
}
\keyword{internal}
