% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-ui.R
\name{dsFormAnalInput}
\alias{dsFormAnalInput}
\title{Differential expression analysis inputs for dsFormInput}
\usage{
dsFormAnalInput(id)
}
\description{
Differential expression analysis inputs for dsFormInput
}
\keyword{internal}
