% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-ui.R
\name{dsFormQuantInput}
\alias{dsFormQuantInput}
\title{Dataset quantification inputs for dsFormInput}
\usage{
dsFormQuantInput(id)
}
\description{
Dataset quantification inputs for dsFormInput
}
\keyword{internal}
