% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsGenePlotly}
\alias{dsGenePlotly}
\title{Logic for Dataset Gene plotly}
\usage{
dsGenePlotly(input, output, session, eset, explore_genes, dataset_name)
}
\description{
Logic for Dataset Gene plotly
}
\keyword{internal}
