% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsQuantTable}
\alias{dsQuantTable}
\title{Logic for dataset quantification table}
\usage{
dsQuantTable(input, output, session, fastq_dir, labels, paired)
}
\description{
Logic for dataset quantification table
}
\keyword{internal}
