% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dtangleForm}
\alias{dtangleForm}
\title{Logic for deconvolution form}
\usage{
dtangleForm(
  input,
  output,
  session,
  show_dtangle,
  new_dataset,
  sc_dir,
  bulk_dir,
  explore_eset,
  dataset_dir,
  dataset_name
)
}
\description{
Logic for deconvolution form
}
\keyword{internal}
