% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{enableAll}
\alias{enableAll}
\title{Enable multiple ids}
\usage{
enableAll(ids, asis = rep(FALSE, length(ids)))
}
\arguments{
\item{ids}{Character vector of ids to enable}
}
\description{
Enable multiple ids
}
\keyword{internal}
