% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{evaluate_custom_metric}
\alias{evaluate_custom_metric}
\title{Evaluate custom single-cell metric}
\usage{
evaluate_custom_metric(metric, scseq)
}
\arguments{
\item{metric}{String with metric to evaluate}

\item{scseq}{\code{DietSCE} to evaluate \code{metric} for.}
}
\value{
data.frame with column name \code{metric} and result.
}
\description{
Evaluate custom single-cell metric
}
\keyword{internal}
