% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{exist_clusters}
\alias{exist_clusters}
\title{Test is there is at lest two clusters}
\usage{
exist_clusters(scseq)
}
\arguments{
\item{scseq}{}
}
\value{
TRUE if more than one cluster exists
}
\description{
Used by \code{\link{get_scseq_markers}} to prevent getting markers if there are no clusters to compare
}
\keyword{internal}
