% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_results.R
\name{explore_results}
\alias{explore_results}
\title{Explore query results}
\usage{
explore_results(cmap_res = NULL, l1000_res = NULL)
}
\arguments{
\item{cmap_res}{Named numeric vector returned from \code{\link{query_drugs}} using CMAP02 data.}

\item{l1000_res}{Named numeric vector returned from \code{\link{query_drugs}} using L1000 data.}
}
\description{
Explore query results
}
\examples{

# load CMAP02 data
cmap_path <- system.file('extdata', 'cmap_es_ind.rds', package = 'drugseqr')
cmap_es <- readRDS(cmap_path)

l1000_path <- system.file('extdata', 'l1000_es.rds', package = 'drugseqr')
l1000_es <- readRDS(l1000_path)

# run differential expression analysis
data_dir <- 'data-raw/example-data'
eset <- readRDS(system.file('extdata', 'IBD', 'eset.rds', package = 'drugseqr'))
# anal <- diff_expr(eset, data_dir)

# alternatively load previous analysis
anal <- readRDS(file.path(data_dir, 'diff_expr_symbol.rds'))

# get dprime effect size values for analysis
dprimes <- get_dprimes(anal)

# get correlations between query and drug signatures
cmap_res <- query_drugs(dprimes, cmap_es)
l1000_res <- query_drugs(dprimes, l1000_es)

explore_results(cmap_res, l1000_res)

}
