% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-docs-ui.R
\name{extract_section_info}
\alias{extract_section_info}
\title{Extracts info needed to construct docsSideNav}
\usage{
extract_section_info(docsSections)
}
\description{
Extracts info needed to construct docsSideNav
}
\keyword{internal}
