% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{file_exists}
\alias{file_exists}
\title{Safe file.exists}
\usage{
file_exists(x)
}
\arguments{
\item{x}{file path}
}
\value{
a logical vector of TRUE or FALSE values
}
\description{
Safe file.exists
}
