% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{filter_clinical}
\alias{filter_clinical}
\title{Subset by clinical phase}
\usage{
filter_clinical(query_table_nsig, show_clinical)
}
\arguments{
\item{query_table_nsig}{result of \link{filter_nsig}}

\item{show_clinical}{if \code{TRUE}, only drugs with a clinical phase annotation are kept.}
}
\value{
\code{query_table_nsig}.
  Drugs without a clinical phase annotation are removed if \code{show_clinical} is \code{FALSE}
}
\description{
Subset by clinical phase
}
\keyword{internal}
