% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{filter_doublets}
\alias{filter_doublets}
\title{Filter cells and clusters with high doublet score}
\usage{
filter_doublets(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} with \code{'doublet_score'} and \code{'cluster'} in \code{scseq@colData}}
}
\value{
\code{scseq} without cells/clusters that have outlying \code{'doublet_score'}
}
\description{
Filter cells and clusters with high doublet score
}
\keyword{internal}
