% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes in RNA-seq ExpressionSet}
\usage{
filter_genes(eset)
}
\arguments{
\item{eset}{ExpressionSet with 'counts' assayDataElement and group column in pData}
}
\value{
filtered \code{eset}
}
\description{
Filter genes in RNA-seq ExpressionSet
}
\keyword{internal}
