% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{filter_nsig}
\alias{filter_nsig}
\title{Subset by min signatures}
\usage{
filter_nsig(query_table_summarised, min_signatures)
}
\arguments{
\item{query_table_summarised}{result of \link{summarise_query_table}}

\item{min_signatures}{Number of independent perturbagen signatures below which a perturbation is excluded. Default is \code{3}.}
}
\value{
\code{query_table_summarised} with drugs filtered that contain fewer than \code{min_signatures}.
}
\description{
Subset by min signatures
}
\keyword{internal}
