% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{find_knee}
\alias{find_knee}
\title{Find knee point by looking for convergance of roryk and DropletUtils}
\usage{
find_knee(counts)
}
\arguments{
\item{counts}{dgCMatrix}
}
\value{
Knee point
}
\description{
Find knee point by looking for convergance of roryk and DropletUtils
}
\keyword{internal}
