% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{fit_lm_scseq}
\alias{fit_lm_scseq}
\title{Fit limma eBayes on single cell RNA-seq dataset}
\usage{
fit_lm_scseq(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}
}
\value{
List with model matrix and result of call to \code{\link[limma]{lmFit}}
}
\description{
Fit limma eBayes on single cell RNA-seq dataset
}
\keyword{internal}
