% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{format_query_res}
\alias{format_query_res}
\title{Format perturbation query result like web app.}
\usage{
format_query_res(
  query_res,
  drug_study,
  sort_by = c("avg_cor", "min_cor"),
  direction = c("opposing", "similar", "both"),
  sort_abs = FALSE,
  show_clinical = TRUE,
  min_signatures = 3,
  cells = NULL
)
}
\arguments{
\item{query_res}{Numeric vector of correlation values with names as HGNC symbols}

\item{drug_study}{Queried database giving rise to \code{query_res}.
One of \code{'CMAP02'},\code{'L1000 Genetic'}, or \code{'L1000 Drugs'}.}

\item{sort_by}{Metric to sort by. Either \code{'avg_cor'} (default) or \code{'min_cor'}.}

\item{direction}{Direction of correlation to sort results by. One of \code{'both'}, \code{'similar'}, or \code{'opposing'} (default).}

\item{sort_abs}{Should results be sorted based on absolute correlation? Default is \code{FALSE}.}

\item{show_clinical}{Should result only contain drugs with clinical phase annotation? Default is \code{TRUE}.}

\item{min_signatures}{Number of independent perturbagen signatures below which a perturbation is excluded. Default is \code{3}.}

\item{cells}{Character vector of cell types to include. Default (\code{NULL}) includes all cell types.}
}
\value{
\code{data.frame} with annotated, filtered, and sorted query result.
}
\description{
Used to get unfiltered query table from downloaded perturbation query result. A condensed top-hits is shown in the
web app for performance reasons.
}
\examples{

# generate fake result
annot <-  dseqr:::get_drugs_table('L1000_drugs')
query_res <- rnorm(nrow(annot), 0, 0.25)
names(query_res) <- annot$title
formatted <- dseqr:::format_query_res(query_res, 'L1000 Drugs', sort_by = 'min_cor')

}
\keyword{internal}
