% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{format_ridge_annot}
\alias{format_ridge_annot}
\title{Add cluster numbers to annotation}
\usage{
format_ridge_annot(annot)
}
\arguments{
\item{annot}{Character vector of cluster names}
}
\value{
\code{annot} with cluster numbers pre-pended to non-numeric values.
}
\description{
Adds cluster number to non-numeric cluster names.
E.g. if \code{annot = c('Monocytes', '2')} then
\code{annot = c('1: Monocytes', '2')} is returned
}
\keyword{internal}
