% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{format_sample_gene_plot}
\alias{format_sample_gene_plot}
\title{Format gene plots for sample comparison for drugseqr app}
\usage{
format_sample_gene_plot(plot, group, selected_gene, scseq)
}
\arguments{
\item{plot}{Returned by \code{\link{plot_umap_gene}}}

\item{group}{Level in \code{scseq$orig.ident} to show cells for. Either \code{'ctrl'} or \code{'test'}}

\item{scseq}{\code{Seurat} object.}
}
\value{
\code{plot} formatted for drugseqr app
}
\description{
Format gene plots for sample comparison for drugseqr app
}
\keyword{internal}
