% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{format_up_custom}
\alias{format_up_custom}
\title{Format uploaded custom query signature}
\usage{
format_up_custom(top_table)
}
\arguments{
\item{top_table}{data.frame with logFC, dprime, or effect sizes as column 1
and row.names as HGNC symbols}
}
\value{
\code{top_table} subsetted to rows with HGNC symbols in CMAP02/L1000 and
 with dprime column set to either \code{top_table$dprime}, \code{top_table$logFC},
 or \code{top_table[,1]}.
}
\description{
Format uploaded custom query signature
}
\keyword{internal}
