% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_whitelist.R
\name{geom_hist_whitelist}
\alias{geom_hist_whitelist}
\title{Generate single histogram for whitelist factors}
\usage{
geom_hist_whitelist(sce_df, x, xlab = "", ylab = "")
}
\arguments{
\item{sce_df}{\code{tibble} with columns \code{'whitelist'} and \code{get(x)}.}

\item{x}{Character vector indicating column in \code{sce_df} to use as x aesthetic.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}
}
\value{
\code{ggplot}
}
\description{
Generate single histogram for whitelist factors
}
\keyword{internal}
