% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{geom_tsne}
\alias{geom_tsne}
\title{Generates individual tsne plots for tsne_scseq_whitelist}
\usage{
geom_tsne(sce, colour_by, name = colour_by, xlab = "", ylab = "",
  scale = "distiller")
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{colour_by}{List within \code{sce} to color by.}

\item{name}{The name of the legend.}

\item{xlab}{Character, x-axis label. Default is \code{''}.}

\item{ylab}{Character, y-axis label. Default is \code{''}.}

\item{scale}{One if either \code{'distiller'} (default), \code{'diverge'} (for mito/ribo), or \code{'manual'} (for whitelist).}
}
\value{
ggplot2 plot
}
\description{
Generates individual tsne plots for tsne_scseq_whitelist
}
