% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{get_10x_whitepath}
\alias{get_10x_whitepath}
\title{Get path to 10x whitelist}
\usage{
get_10x_whitepath(indices_dir, bus_args)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices built with \code{build_kallisto_index}.}

\item{bus_args}{Character vector of arguments to bustools.}
}
\value{
Path to 10x whitelist.
}
\description{
Get path to 10x whitelist
}
\keyword{internal}
