% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{get_ambience}
\alias{get_ambience}
\title{Get genes that are ambient in at least one test and control sample}
\usage{
get_ambience(combined)
}
\arguments{
\item{scseqs}{List of \code{SingleCellExperiment} objects.}
}
\value{
List with test and control ambient genes
}
\description{
Get genes that are ambient in at least one test and control sample
}
\keyword{internal}
