% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_ambient}
\alias{get_ambient}
\title{Get ambient genes to exclude for diff_expr_scseq}
\usage{
get_ambient(scseq, markers, cluster_markers)
}
\arguments{
\item{scseq}{\code{Suerat} object.}

\item{top_table}{\code{data.frame} with results from \code{\link[limma]{topTable}}.}
}
\value{
Character vector of HGNC symbols representing ambient genes to exclude
}
\description{
Excludes upregulated genes if they are ambient in the test group.
Excludes downregulated genes if they are ambient in the control group.
}
\keyword{internal}
