% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{get_boxplotly_cell_args}
\alias{get_boxplotly_cell_args}
\title{Get arguments for cells boxplotly}
\usage{
get_boxplotly_cell_args(pdata, dtangle_est, dataset_name)
}
\arguments{
\item{pdata}{data.frame of phenotype data.}

\item{dtangle_est}{data.frame of proportion estimates from dtangle.}

\item{dataset_name}{Name of bulk dataset.}
}
\value{
List with items \code{'df'}, \code{'boxgap'}, \code{'boxgroupgap'}, and \code{'plot_fname'}.
}
\description{
Get arguments for cells boxplotly
}
\keyword{internal}
