% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{get_boxplotly_gene_args}
\alias{get_boxplotly_gene_args}
\title{Get arguments for gene boxplotly}
\usage{
get_boxplotly_gene_args(eset, explore_genes, dataset_name)
}
\arguments{
\item{eset}{ExpressionSet object with \code{'adjusted'} assayDataElement}

\item{explore_genes}{Character vector of genes to plot}

\item{dataset_name}{Name of bulk dataset.}
}
\value{
List with items \code{'df'}, \code{'boxgap'}, \code{'boxgroupgap'}, and \code{'plot_fname'}.
}
\description{
Get arguments for gene boxplotly
}
