% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_cell_choices}
\alias{get_cell_choices}
\title{Get cell choices data.frame for CMAP02 or L1000}
\usage{
get_cell_choices(drug_study)
}
\arguments{
\item{drug_study}{either 'CMAP02', 'L1000 Drugs', or 'L1000 Genetic'}
}
\value{
data.frame
}
\description{
Get cell choices data.frame for CMAP02 or L1000
}
\keyword{internal}
