% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cell_pcts}
\alias{get_cell_pcts}
\title{Get percentage of cells expressing each gene}
\usage{
get_cell_pcts(scseq, ident.1, ident.2)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{ident.1}{Test group level in \code{Idents(scseq)}.}

\item{ident.2}{Control group level in \code{Idents(scseq)}.}
}
\value{
data.frame with columns pct.1 and pct.2 indicating fraction of cells that express each gene.
}
\description{
Get percentage of cells expressing each gene
}
\keyword{internal}
