% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_cluster_markers}
\alias{get_cluster_markers}
\title{Loads markers for single cell cluster}
\usage{
get_cluster_markers(selected_clusters, dataset_dir)
}
\arguments{
\item{selected_clusters}{Character vector of integers}

\item{dataset_dir}{Directory to folder with single-cell dataset}
}
\value{
data.frame with markers
}
\description{
Loads markers for single cell cluster
}
\keyword{internal}
