% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_contrast_choices}
\alias{get_contrast_choices}
\title{Get contrast choices data.frame for selectize dropdown}
\usage{
get_contrast_choices(clusters, test)
}
\arguments{
\item{clusters}{Character vector of cluster names}

\item{test}{Name of test contrast}
}
\value{
data.frame with columns for rendering selectizeInput contrast choices
}
\description{
Get contrast choices data.frame for selectize dropdown
}
\keyword{internal}
