% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_contrast_markers}
\alias{get_contrast_markers}
\title{Get markers for one cluster against one other cluster}
\usage{
get_contrast_markers(con, markers)
}
\arguments{
\item{con}{String identifying clusters to compare e.g. \code{'1vs2'}}

\item{dataset_dir}{Path to folder for dataset}
}
\value{
Named list with data.frames for each comparison direction
}
\description{
Get markers for one cluster against one other cluster
}
\keyword{internal}
