% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_cors_html}
\alias{get_cors_html}
\title{Get HTML for correlation values.}
\usage{
get_cors_html(cors, titles, cor_titles, cors_range)
}
\arguments{
\item{cors}{List of numeric vectors of pearson correlations.}

\item{titles}{List of character vectors of treatment titles for pearson correlations (e.g. MCF7_1e-05M_6h_3).}

\item{cor_titles}{List of character vectors used for title element. Used
to allow right click on correlation point to load query for perturbation signature.}

\item{cors_range}{Numeric vector of length two specifying the range of correlation values.}
}
\value{
Character vector of HTML markup for the title/circle/text for a correlation plot.
}
\description{
Get HTML for correlation values.
}
\keyword{internal}
