% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_whitelist.R
\name{get_diverge}
\alias{get_diverge}
\title{Get a continuous color scale that emphasizes 0-10th and 90-100th percentiles.}
\usage{
get_diverge(x, name)
}
\arguments{
\item{x}{Numeric vector}

\item{name}{Character, name of scale}
}
\value{
\code{ScaleContinuous}
}
\description{
Get a continuous color scale that emphasizes 0-10th and 90-100th percentiles.
}
\keyword{internal}
