% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{get_dprimes}
\alias{get_dprimes}
\title{Get dprime effect size values.}
\usage{
get_dprimes(diff_exprs)
}
\arguments{
\item{diff_exprs}{List returned by \code{\link{dif_expr}}}
}
\description{
These are used to query against drug effect size matrices.
}
\examples{

# load result of previous differential expression analysis
data_dir <- file.path('data-raw/example-data')
anal <- readRDS(file.path(data_dir, 'diff_expr_symbol.rds'))

dprimes <- get_dprimes(anal)

}
