% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_drug_paths}
\alias{get_drug_paths}
\title{Get file paths to drug query results}
\usage{
get_drug_paths(data_dir, suffix, ftype = ".qs")
}
\arguments{
\item{data_dir}{Folder with drug query results}

\item{suffix}{string that appears after e.g. \code{'cmap_res_'} in the drug query result filenames.}
}
\value{
List with file paths to drug query results
}
\description{
Get file paths to drug query results
}
\keyword{internal}
