% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pug.R
\name{get_drugbank}
\alias{get_drugbank}
\title{Extract DrugBank URL from PUG-View}
\usage{
get_drugbank(pug_view)
}
\arguments{
\item{pug_view}{List from reading PUG-View JSON data with \code{\link[rjson]{fromJSON}}}
}
\value{
Character vector with DrugBank ID or NA if none exists
}
\description{
Extract DrugBank URL from PUG-View
}
\keyword{internal}
