% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugs_table.R
\name{get_drugs_table}
\alias{get_drugs_table}
\title{Get Drugs metadata table for CMAP02 or L1000}
\usage{
get_drugs_table(study)
}
\arguments{
\item{study}{either \code{'CMAP02'} or \code{'L1000_genes'} or \code{'L1000_drugs'}}
}
\value{
\code{data.frame} of meta data for CMAP02 or L1000
}
\description{
Get Drugs metadata table for CMAP02 or L1000
}
\examples{

cmap_meta <- get_drugs_table('CMAP02')
l1000_meta <- get_drugs_table('L1000_genes')

}
\keyword{internal}
