% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_scseq.R
\name{get_empty}
\alias{get_empty}
\title{Get empty droplets}
\usage{
get_empty(counts)
}
\arguments{
\item{counts}{sparse dgTMatrix returned by \code{\link{load_kallisto}}.}
}
\value{
Boolean with length \code{ncol(counts)} indicating if the droplet is empty (\code{TRUE}) or a cell(\code{FALSE}).
}
\description{
Used for calling empty droplets for testing kallisto and SoupX. Recommendation is currently to use alevin cell
filtering instead.
}
