% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{get_ensdb_package}
\alias{get_ensdb_package}
\title{Get ensembldb package name}
\usage{
get_ensdb_package(species, release)
}
\arguments{
\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link[drugseqr.data]{build_kallisto_index}}.}
}
\value{
Character vector with ensembldb package name. e.g. \code{'EnsDb.Hsapiens.v94'}.
}
\description{
Get ensembldb package name
}
\keyword{internal}
