% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_exclude_choices}
\alias{get_exclude_choices}
\title{Get choices for included cluster in integration}
\usage{
get_exclude_choices(anal_names, anal_colors, data_dir)
}
\arguments{
\item{anal_names}{Names of analyses selected for integration}

\item{anal_colors}{Character vector of colors to indicate analysis}

\item{data_dir}{Directory with single cell analyses.}
}
\value{
data.frame with columns for rendering selectizeInput include choices
}
\description{
Get choices for included cluster in integration
}
\keyword{internal}
