% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_expression_colors}
\alias{get_expression_colors}
\title{Get expression colors for scatterplot}
\usage{
get_expression_colors(ft.scaled)
}
\arguments{
\item{ft.scaled}{expression values scaled from 0 to 1}
}
\value{
character vector of colors
}
\description{
Get expression colors for scatterplot
}
\examples{

ft <- rnorm(100)
ft.scaled <- scales::rescale(ft)
colors <- get_expression_colors(ft.scaled)

}
