% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{get_fastq_id1s}
\alias{get_fastq_id1s}
\title{Get first sequence identifiers for fastq.gz files}
\usage{
get_fastq_id1s(fastq_paths)
}
\arguments{
\item{fastq_paths}{Character vector of paths to fastq.gz files}
}
\value{
Named character vector of first sequence id lines (start with @) in \code{fastq_paths}.
  Names are the \code{fastq_paths}.
}
\description{
Function is used to determine if an experiment is single or paired.
}
\keyword{internal}
