% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_fastqs}
\alias{get_fastqs}
\title{Download and fasterq-dump RNA-seq data from GEO}
\usage{
get_fastqs(gse_name, srp_meta, data_dir = getwd())
}
\arguments{
\item{gse_name}{GSE name. Will create folder with this name in \code{data_dir} and download data there.}

\item{srp_meta}{\code{data.frame} with SRP meta info. Returned from \code{\link{get_srp_meta}}.}

\item{data_dir}{Path to folder that \code{gse_name} folder will be created in.}
}
\description{
First tries to get RNA-Seq fastq files from EBI. If not successful, gets SRA from GEO and converts to fastq.gz.
}
