% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{get_feature_data}
\alias{get_feature_data}
\title{Get Data to Update Feature Plot}
\usage{
get_feature_data(plots_dir, scseq, feature)
}
\arguments{
\item{plots_dir}{Path to directory where feature data is saved}

\item{scseq}{SingleCellExperiment}

\item{feature}{Name of feature to get}
}
\value{
Named numeric vector of feature values. Names are cell names.
}
\description{
If feature data previously saved, then it is loaded. Otherwise, feature
data is saved for subsequent calls.
}
\keyword{internal}
