% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_founder}
\alias{get_founder}
\title{Determine the founder of a single-cell dataset}
\usage{
get_founder(sc_dir, dataset_name)
}
\arguments{
\item{sc_dir}{Directory with single-cell datasets.}

\item{dataset_name}{Name of dataset to determine founder for.}
}
\value{
Name of founder.
}
\description{
Founder is not necesarily the parent, but rather the original ancestor.
}
