% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_gene_choices}
\alias{get_gene_choices}
\title{Add cell percents to gene choices for single cell}
\usage{
get_gene_choices(
  markers,
  qc_metrics = NULL,
  qc_first = FALSE,
  species = "Homo sapiens",
  tx2gene = NULL
)
}
\arguments{
\item{markers}{data.frame of marker genes.}

\item{scseq}{\code{SingleCellExperiment} object}
}
\value{
data.frame of all genes, with markers on top and cell percent columns
}
\description{
Add cell percents to gene choices for single cell
}
\keyword{internal}
