% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_genego}
\alias{get_genego}
\title{Get gene to GO map for pathway analysis}
\usage{
get_genego(species = "Hs", gs_dir = NULL)
}
\arguments{
\item{species}{Species identifier}

\item{gs_dir}{Directory to save genego to}

\item{type}{either 'go' or 'kegg'}
}
\value{
genego
}
\description{
Get gene to GO map for pathway analysis
}
\keyword{internal}
