% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_geo_rnaseq}
\alias{get_geo_rnaseq}
\title{Download and fasterq-dump RNA-seq data from GEO}
\usage{
get_geo_rnaseq(gse_name, srp_meta, data_dir)
}
\arguments{
\item{gse_name}{GSE name. Will create folder with this name in \code{data_dir} and download data there.}

\item{srp_meta}{\code{data.frame} with SRP meta info. Returned from \code{\link{get_srp_meta}}.}

\item{data_dir}{Path to folder that \code{gse_name} folder will be created in.}
}
\description{
Download and fasterq-dump RNA-seq data from GEO
}
\examples{

 gse_name <- 'GSE117570'
 srp_meta <- get_srp_meta(gse_name)
 srp_meta <- srp_meta[grepl('_Normal$', srp_meta$title), ]

 data_dir <- 'data-raw/single-cell/reference-data'

 get_geo_rnaseq(gse_name, srp_meta, data_dir)

}
