% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_gonames}
\alias{get_gonames}
\title{Get names of gene set}
\usage{
get_gonames(genego, gs_dir = NULL)
}
\arguments{
\item{gs_dir}{Directory to save results to}

\item{gslist}{result of \code{\link{get_gslist}}}
}
\value{
Description of \code{gslist} gene sets
}
\description{
Get names of gene set
}
\keyword{internal}
