% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{get_group_colors}
\alias{get_group_colors}
\title{Get group colors for bulk data plots}
\usage{
get_group_colors(group_levels)
}
\arguments{
\item{group_levels}{result of \link{get_group_levels}}
}
\description{
Get group colors for bulk data plots
}
\keyword{internal}
