% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{get_group_levels}
\alias{get_group_levels}
\title{Get group levels for bulk data plots}
\usage{
get_group_levels(pdata)
}
\arguments{
\item{pdata}{Data.frame of phenotype data}
}
\description{
Get group levels for bulk data plots
}
\keyword{internal}
