% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{get_gs.names}
\alias{get_gs.names}
\title{Get names of gene set}
\usage{
get_gs.names(
  gslist,
  type = "go",
  species = "Hs",
  gs_dir = "/srv/drugseqr/gs_dir"
)
}
\arguments{
\item{gslist}{result of \code{\link{get_gslist}}}

\item{type}{either 'go' or 'kegg'}

\item{species}{species identifier}

\item{gs_dir}{Directory to save results to}
}
\value{
Description of \code{gslist} gene sets
}
\description{
Get names of gene set
}
