% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{get_hvgs}
\alias{get_hvgs}
\title{Get top highly variable genes for subsequent dimensionality reduction}
\usage{
get_hvgs(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object}
}
\description{
Runs after \code{preprocess_scseq}
}
