% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_integrated_ambient}
\alias{get_integrated_ambient}
\title{Get genes that are ambient in at least one test and control sample}
\usage{
get_integrated_ambient(scseqs)
}
\arguments{
\item{scseqs}{List of Seurat objects.}
}
\value{
List with test and control ambient genes
}
\description{
Get genes that are ambient in at least one test and control sample
}
\keyword{interal}
