% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{get_kegg_species}
\alias{get_kegg_species}
\title{Convert from species identifier to KEGG species identifier}
\usage{
get_kegg_species(species)
}
\arguments{
\item{species}{species identifier}
}
\value{
KEGG species identifier
}
\description{
Convert from species identifier to KEGG species identifier
}
