% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{get_knee}
\alias{get_knee}
\title{Get knee point and plot on barcode rank plot}
\usage{
get_knee(counts, knee_type = c("inflection", "roryk", "knee"))
}
\arguments{
\item{counts}{dgTMatrix of counts for all barcodes.}

\item{knee_type}{Knee type used for cell quality whitelist modeling by \code{get_scseq_whitelist}.}
}
\value{
count value corresponding to knee point
}
\description{
Get knee point and plot on barcode rank plot
}
\keyword{internal}
