% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_label_plot}
\alias{get_label_plot}
\title{Get plot to compare original labels for integrated single cell dataset.}
\usage{
get_label_plot(anal, scseq, annot, plot)
}
\arguments{
\item{anal}{Name of original analysis in \code{scseq$project} to plot.}

\item{scseq}{Integrated \code{Seurat} object.}

\item{annot}{Character vector of original labels for \code{anal}.}

\item{plot}{\code{ggplot} object from integrated \code{scseq}.}
}
\value{
\code{ggplot} object showing cells in \code{anal} with original labels but coordinates from \code{plot}
 of integrated \code{scseq}.
}
\description{
Get plot to compare original labels for integrated single cell dataset.
}
