% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_label_transfer_choices}
\alias{get_label_transfer_choices}
\title{Get label transfer choices data.frame}
\usage{
get_label_transfer_choices(anal_options, selected_anal, preds, species)
}
\arguments{
\item{anal_options}{Names list of analysis options with names \code{'Individual'} and \code{'Integrated'}.}

\item{preds}{Named list of predicted cluster labels. Names are values in \code{anal_options} lists.}
}
\value{
data.frame with columns \code{value}, \code{label}, \code{type}, and \code{preds}.
}
\description{
Used for Single Cell tab label transfer selectizeInput
}
\keyword{internal}
