% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_metric_choices}
\alias{get_metric_choices}
\title{Get choices data.frame for custom metrics (for subsetting)}
\usage{
get_metric_choices(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}
}
\value{
data.frame
}
\description{
Get choices data.frame for custom metrics (for subsetting)
}
\keyword{internal}
