% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_metric_features}
\alias{get_metric_features}
\title{Get features from custom metric}
\usage{
get_metric_features(metric)
}
\arguments{
\item{metric}{Custom metric to extract features from.}
}
\value{
Character vector of feature names extracted from \code{metric}
}
\description{
Get features from custom metric
}
\keyword{internal}
