% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_mods}
\alias{get_mods}
\title{Get model matrices for surrogate variable analysis}
\usage{
get_mods(pdata)
}
\arguments{
\item{eset}{Annotated eset with samples selected during \code{add_contrasts}.}
}
\value{
List with model matrix(mod) and null model matrix (mod0) used for \code{sva}.
}
\description{
Used by \code{add_adjusted} to create model matrix with surrogate variables.
}
