% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_nearest_row}
\alias{get_nearest_row}
\title{Search for closest row in a truth matrix for each row in a test matrix.}
\usage{
get_nearest_row(truth, test)
}
\arguments{
\item{truth}{matrix to pick nearest from for all rows in \code{test}.}

\item{test}{matrix to test each row of and find closest rows in \code{truth}.}
}
\value{
vector of rows in \code{truth} that are the nearest to each row in \code{test}.
}
\description{
from Stack Overflow: 40668623
}
\examples{
set.seed(123)  ## for reproducibility
D <- 2 #amount of dimensions
K <- 5
events <- 2*K #number of events
truth <- matrix(data=runif(events, min = 0, max = 1), nrow=K)
E <- 2
test <- matrix(data=runif(2*E, min = 0, max = 1), nrow=E)

get_nearest_row(truth, test)
#[1] 4 3
}
\keyword{internal}
