% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_npc_choices}
\alias{get_npc_choices}
\title{Get number of clusters different number of PCs}
\usage{
get_npc_choices(sce, type = "PCA")
}
\arguments{
\item{sce}{\code{SingleCellExperiement}}
}
\value{
result of \code{scran::getClusteredPCs}
}
\description{
Used to pick number of PCs to retain
}
\keyword{internal}
