% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_open_a}
\alias{get_open_a}
\title{Get Opening HTML a Tag}
\usage{
get_open_a(pre_url, ids, post_url, title)
}
\arguments{
\item{pre_url}{Character with url portion to paste before \code{id_col} column values.}

\item{ids}{non \code{NA} ids to be inserted between \code{pre_url} and \code{post_url} to form the link.}

\item{post_url}{Character with url portion to paste before \code{id_col} column values.}

\item{title}{Character that will be added to hyperlink title attribute. Default is \code{id_col}.}
}
\value{
Character vector of opening HTML a tags
}
\description{
If there are multiple-entry \code{ids} (e.g. \code{'2250, | 60823'}), the first entry
is added to the href attribute and subsequent entries are added to onclick javascript.
}
