% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_outliers}
\alias{get_outliers}
\title{Flag outliers}
\usage{
get_outliers(x)
}
\arguments{
\item{x}{Named numeric vector}
}
\value{
Boolean vector with \code{length(x)} indicating if values of \code{x} are outliers (TRUE) or not (FALSE).
}
\description{
Flag outliers
}
\keyword{internal}
