% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_path_directions}
\alias{get_path_directions}
\title{Get direction of pathways}
\usage{
get_path_directions(top_table)
}
\arguments{
\item{top_table}{}
}
\value{
a \code{data.frame} with columns: \itemize{
 \item is.up Boolean indicating if direction is mostly up
 \item label Character vector with values \code{'Mostly Up'} or \code{'Mostly Down'}.
 These are used for selectize input option group field.
}
}
\description{
Get direction of pathways
}
\keyword{internal}
