% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{get_path_res}
\alias{get_path_res}
\title{Get and save pathway results for ebfit object}
\usage{
get_path_res(ebfit, goana_path, kegga_path)
}
\arguments{
\item{kegga_path}{}
}
\value{

}
\description{
Used to avoid code reuse for single-cell and bulk
}
\keyword{internal}
