% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-server.R
\name{get_pathway_heatmap}
\alias{get_pathway_heatmap}
\title{Get heatmap of genes in a GO pathways}
\usage{
get_pathway_heatmap(eset, top_table, contrast_groups, path_id)
}
\arguments{
\item{eset}{ExpressionSet}

\item{top_table}{limma topTable}

\item{contrast_groups}{groups to include in heatmap}

\item{path_id}{GO pathway id string}
}
\value{
\code{pheatmap} object.
}
\description{
Get heatmap of genes in a GO pathways
}
