% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_pct_ambient}
\alias{get_pct_ambient}
\title{Determine ambient percent for each gene}
\usage{
get_pct_ambient(counts)
}
\arguments{
\item{counts}{\code{dgTMatrix} of counts. Rows are genes, columns are droplets.}
}
\value{
Named numeric vector of percentages for each gene.
}
\description{
Looks at droplets with counts less than or equal to 10 calculates the total percent for each gene
}
\keyword{internal}
