% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_salmon_bulk.R
\name{get_pkg_version}
\alias{get_pkg_version}
\title{Get version of salmon/kallisto from system command.}
\usage{
get_pkg_version(type)
}
\arguments{
\item{type}{Either \code{'salmon'} or \code{'kallisto'}.}
}
\value{
Version of salmon/kallisto.
}
\description{
Get version of salmon/kallisto from system command.
}
\keyword{internal}
