% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_pred_annot}
\alias{get_pred_annot}
\title{Get predicted annotation for label transfer}
\usage{
get_pred_annot(ref_preds, ref_name, anal_name, sc_dir, min.score = 0.9)
}
\arguments{
\item{ref_preds}{data.frame generated in \code{\link{labelTransferForm}} on event \code{submit_transfer}}

\item{ref_name}{Name of reference analysis that labels are transfered from.}

\item{anal_name}{Name of analysis that labels are transfered to.}

\item{sc_dir}{Directory containing folders with analyses for \code{ref_name} and \code{anal_name}.}

\item{min.score}{Minimum average prediction score used for label transfer.}
}
\value{
Character vector of predicted labels from \code{ref_name}.
}
\description{
Clusters with an average prediction scores below \code{min.score} retain their original labels.
}
\keyword{internal}
