% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{get_query_cols}
\alias{get_query_cols}
\title{Get query columns}
\usage{
get_query_cols(is_genetic)
}
\arguments{
\item{is_genetic}{Should gene columns be returned? If \code{FALSE} drug columns are returned.}
}
\value{
Character vector of column names.
}
\description{
Gets appropriate column names for query study.
}
\keyword{internal}
