% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_resoln_name}
\alias{get_resoln_name}
\title{Get the applied resolution dataset name}
\usage{
get_resoln_name(sc_dir, dataset_name)
}
\arguments{
\item{sc_dir}{directory with \code{dataset_name} folder}

\item{dataset_name}{name of single-cell dataset}
}
\value{
the resolution dataset name based on the last applied resolution
}
\description{
e.g. PBMCS_1/snn1
}
\keyword{internal}
