% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_sc_dataset_choices}
\alias{get_sc_dataset_choices}
\title{Get data.frame of single-cell dataset choices}
\usage{
get_sc_dataset_choices(sc_dir, new_dataset = NULL)
}
\arguments{
\item{sc_dir}{Directory containing single-cell datasets}
}
\value{
data.frame of single-cell dataset choices for selectizeInput
}
\description{
Get data.frame of single-cell dataset choices
}
\keyword{internal}
