% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_scseq_assay}
\alias{get_scseq_assay}
\title{Get assay to use from Seurat object.}
\usage{
get_scseq_assay(scseq)
}
\arguments{
\item{scseq}{Seurat object}
}
\value{
either \code{'SCT'} or \code{'RNA'}
}
\description{
Default is 'SCT' but if not present then 'RNA'
}
\keyword{internal}
