% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_scseq_markers}
\alias{get_scseq_markers}
\title{Combine pairwise wilcox tests between single-cell clusters}
\usage{
get_scseq_markers(tests, pval.type = "some", effect.field = "AUC", keep = NULL)
}
\arguments{
\item{tests}{Result of \code{pairwise_wilcox}}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{effect.field}{A string specifying the column name of each element of \code{de.lists} that contains the effect size.
If \code{NULL}, effect sizes are not reported in the output.}
}
\value{
List of data.frames
}
\description{
Combine pairwise wilcox tests between single-cell clusters
}
\keyword{internal}
