% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{get_scseq_whitelist}
\alias{get_scseq_whitelist}
\title{Get whitelist for high quality cells}
\usage{
get_scseq_whitelist(counts, data_dir)
}
\arguments{
\item{counts}{dgTMatrix of counts for all barcodes.}

\item{data_dir}{Directory that contains kallisto \code{'bus_output'} folder that counts were loaded from.
whitelist is loaded from here if previously saved.}
}
\value{
Character vector of barcodes called as high quality cells.
}
\description{
Get whitelist for high quality cells
}
\keyword{internal}
