% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_srp_meta}
\alias{get_srp_meta}
\title{Get metadata needed to download RNA-seq data for GSE}
\usage{
get_srp_meta(gse_name)
}
\arguments{
\item{gse_name}{GEO study name to get metadata for}
}
\value{
\code{data.frame} with sample annotations for each GSE
}
\description{
Goes to GSE page to get GSMs then goes to each GSM page to get SRX then to each SRX page to get some more metadata
Adapted from Alex's personal SRAdb package.
}
\examples{

srp_meta <- get_srp_meta('GSE117570')
}
