% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_results.R
\name{get_top}
\alias{get_top}
\title{Reduce entries in query results}
\usage{
get_top(query_res, nclinic = 100)
}
\arguments{
\item{query_res}{\code{data.frame} returned by \code{\link{append_pdata}}.}

\item{nclinic}{Integer indicating the number of entries with a clinical phase that should be included.}
}
\value{
\code{query_res} including only compounds that have a correlation less than the top entries with a clinical phase.
}
\description{
Gets all entries for compounds that have a correlation less than the top entries with a clinical phase.
}
