% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_top_cors}
\alias{get_top_cors}
\title{Get top correlated compounds for drug query}
\usage{
get_top_cors(query_table, ntop, is_genetic = FALSE)
}
\arguments{
\item{query_table}{data.frame with columns \code{'Correlation'} and \code{'Compound'}.}

\item{ntop}{Integer indicating the number of rows to keep after sorting by \code{arrange_by}.}

\item{is_genetic}{Boolean indicating if the query results are from L1000 genetic perturbations.}
}
\value{
data.table summarized by Compound with top results.
}
\description{
Get top correlated compounds for drug query
}
\keyword{internal}
