% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_top_tables}
\alias{get_top_tables}
\title{Get top table}
\usage{
get_top_tables(lm_fit, groups = c("test", "ctrl"), contrasts = NULL)
}
\arguments{
\item{lm_fit}{Result of \link{run_limma}}

\item{contrast}{String specifying contrast.}
}
\value{
result of \link[limma]{topTable}
}
\description{
Get top table
}
