% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_tsne_coords}
\alias{get_tsne_coords}
\title{Get median x-y coordinates for clusters in TSNE plot data}
\usage{
get_tsne_coords(plot_data)
}
\arguments{
\item{plot_data}{data.frame with columns \code{'ident'}, \code{'TSNE_1'}, and \code{'TSNE_2'}.}
}
\description{
Get median x-y coordinates for clusters in TSNE plot data
}
