% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{get_tx2gene}
\alias{get_tx2gene}
\title{Get transcript to gene map.}
\usage{
get_tx2gene(species, release)
}
\arguments{
\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link{build_index}}.}
}
\value{
\code{data.frame} with columns \code{tx_id}, \code{gene_name}, and \code{entrezid}
}
\description{
Get transcript to gene map.
}
\keyword{internal}
