% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_umap_coords}
\alias{get_umap_coords}
\title{Get median x-y coordinates for clusters in UMAP plot data}
\usage{
get_umap_coords(plot_data)
}
\arguments{
\item{plot_data}{data.frame with columns \code{'ident'}, \code{'UMAP_1'}, and \code{'UMAP_2'}.}
}
\description{
Get median x-y coordinates for clusters in UMAP plot data
}
