% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pug.R
\name{get_wikipedia}
\alias{get_wikipedia}
\title{Extract Wikipedia URL from PUG-View}
\usage{
get_wikipedia(pug_view)
}
\arguments{
\item{pug_view}{List from reading PUG-View JSON data with \code{\link[rjson]{fromJSON}}}
}
\value{
Character vector with Wikipedia page
}
\description{
Extract Wikipedia URL from PUG-View
}
\keyword{internal}
