% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot.R
\name{GetXYAesthetics}
\alias{GetXYAesthetics}
\title{Get X and Y aesthetics from a plot for a certain geom}
\usage{
GetXYAesthetics(plot, geom = "GeomPoint", plot.first = TRUE)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{geom}{Geom class to filter to}

\item{plot.first}{Use plot-wide X/Y aesthetics before geom-specific aesthetics}
}
\value{
A named list with values 'x' for the name of the x aesthetic and 'y' for the y aesthetic
}
\description{
Get X and Y aesthetics from a plot for a certain geom
}
