% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.R
\name{heatplot}
\alias{heatplot}
\title{Plot heatmap for genes from goana/kegga analysis}
\usage{
heatplot(path_res, foldChange = NULL)
}
\arguments{
\item{path_res}{data.frame, result of \code{limma::goana} or \code{limma::kegga} with
significant genes for each pathway added by \code{add_path_genes}.}

\item{foldChange}{Vector of logFC values with names equal to gene names as in \code{path_res$GeneName}}
}
\value{
\code{ggplot} object. If foldChange provided then tiles colored accordingly.
}
\description{
Plot heatmap for genes from goana/kegga analysis
}
