% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadablePlotly.R
\name{hiddenDownload}
\alias{hiddenDownload}
\title{Logic for hidden download link}
\usage{
hiddenDownload(input, output, session, check, fname_fun, data_fun)
}
\arguments{
\item{check}{reactive value to trigger download}

\item{fname_fun}{function that returns filename for download}

\item{data_fun}{function that saves file for download}
}
\description{
Logic for hidden download link
}
\keyword{internal}
