% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{hist_scseq_whitelist}
\alias{hist_scseq_whitelist}
\title{Plot histograms for whitelist factors}
\usage{
hist_scseq_whitelist(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} returned by \code{load_scseq}}
}
\value{
plot
}
\description{
Currently plots histograms for percent ribosomal reads, percent mitochondrial reads, log of total counts,
and log of total expressed features.
}
