% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{html_space}
\alias{html_space}
\title{Format character vectors that will be recognized by HTML}
\usage{
html_space(x, justify = "right")
}
\arguments{
\item{x}{any \R object (conceptually); typically numeric.}

\item{justify}{should a \emph{character} vector be left-justified (the
    default), right-justified, centred or left alone.  Can be abbreviated.}
}
\value{
Formatted \code{x} with &nbsp; substituted for each space.
}
\description{
Format character vectors that will be recognized by HTML
}
\keyword{internal}
