% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{id_from_tab}
\alias{id_from_tab}
\title{Convert tab name to formated id}
\usage{
id_from_tab(tab)
}
\arguments{
\item{tab}{The name of the tab (e.g. \code{'Single Cell'})}
}
\description{
used by navbarUI and *PageUI for drugseqr app
}
\keyword{internal}
