% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R, R/modules-sc-utils.R
\name{integrate_scseqs}
\alias{integrate_scseqs}
\title{Integrate multiple scRNA-seq samples}
\usage{
integrate_scseqs(scseqs, use_scalign = FALSE)

integrate_scseqs(scseqs, use_scalign = FALSE)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects}

\item{use_scalign}{whether or not to use scAlign. Default is \code{FALSE}}

\item{scseqs}{List of Seurat objects}
}
\value{
Integrated \code{Seurat} object with default assay of \code{"integrated"}

Integrated Seurat object.
}
\description{
Integrate multiple scRNA-seq samples

Integrate Seurat objects
}
\keyword{internal}
