% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{integrationForm}
\alias{integrationForm}
\title{Single Cell Integration form}
\usage{
integrationForm(input, output, session, data_dir, anal_options,
  show_integration)
}
\value{
\code{reactiveVal} that is either NULL or contains the name of a new integrated analysis
}
\description{
Single Cell Integration form
}
\keyword{internal}
