% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{interpret}
\alias{interpret}
\title{Check safety of evaluating string as R code.}
\usage{
interpret(
  expr_str,
  max_length = 200,
  whitelist = c("&", ">=", "<=", ">", "<", "|", "==", "!=", "*")
)
}
\arguments{
\item{expr_str}{String to evaluate,}

\item{max_length}{maximum length of string to evaluate,}

\item{whitelist}{R functions to allow use of.}
}
\value{
\code{TRUE} if expression is safe, otherwise \code{FALSE}.
}
\description{
Check safety of evaluating string as R code.
}
\keyword{internal}
