% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{iqr_replicates}
\alias{iqr_replicates}
\title{Removes features with replicated annotation.}
\usage{
iqr_replicates(eset, annot = "SYMBOL", rm.dup = FALSE)
}
\arguments{
\item{eset}{Annotated eset created by \code{\link{load_seq}}.}

\item{annot}{feature to use to remove replicates.}

\item{rm.dup}{remove duplicates (same measure, multiple ids)?}

\item{mod}{Model matrix without surrogate variables. generated by \code{diff_setup}.}

\item{svobj}{Result from \code{sva} function called during \code{diff_setup}.}
}
\value{
Expression set with unique features at probe or gene level.
}
\description{
For rows with duplicated annot, highested IQR retained.
}
