% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{is.ambient}
\alias{is.ambient}
\title{Determine ambient transcripts}
\usage{
is.ambient(counts)
}
\arguments{
\item{counts}{\code{dgTMatrix} of counts. Rows are genes, columns are droplets.}
}
\value{
Boolean vector indicating genes that are enriched in ambient RNA fraction.
}
\description{
Looks at droplets with counts less than or equal to 10 and flags genes with outlier total counts
as ambient.
}
\keyword{internal}
