% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{is_invertible}
\alias{is_invertible}
\title{Check uploaded bulk pdata to make sure the study design is invertible}
\usage{
is_invertible(pdata)
}
\description{
Check uploaded bulk pdata to make sure the study design is invertible
}
\keyword{internal}
