% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{isTruthy}
\alias{isTruthy}
\title{Truthy and falsy values}
\usage{
isTruthy(x)
}
\arguments{
\item{x}{An expression whose truthiness value we want to determine}
}
\value{
a logical vector of TRUE or FALSE values
}
\description{
Truthy and falsy values
}
