% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{isTruthyAll}
\alias{isTruthyAll}
\title{Check truthiness of multiple objects}
\usage{
isTruthyAll(...)
}
\arguments{
\item{...}{objects to check truthiness of}
}
\description{
Check truthiness of multiple objects
}
\keyword{internal}
