% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{jitter_umap}
\alias{jitter_umap}
\title{Add jitter to UMAP embeddings}
\usage{
jitter_umap(scseq, factor = 1, amount = 0)
}
\arguments{
\item{scseq}{\code{Seurat} object with umap reduction}

\item{factor}{numeric.}

\item{amount}{numeric; if positive, used as \emph{amount} (see below),
    otherwise, if \code{= 0} the default is \code{factor * z/50}.

    Default (\code{NULL}): \code{factor * d/5} where \code{d} is about
    the smallest difference between \code{x} values.}
}
\value{
\code{scseq} with jitter added to umap embedding
}
\description{
Add jitter to UMAP embeddings
}
