% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{justifiedButtonGroup}
\alias{justifiedButtonGroup}
\title{Full width button group with validation}
\usage{
justifiedButtonGroup(..., label, container_id = NULL, help_id = NULL)
}
\arguments{
\item{...}{actionButtons}

\item{label}{Character vector. Label above button group.}

\item{container_id}{id of container. Used to toggle 'has-error' class with shinyjs::toggleClass.}

\item{help_id}{id of help block. Used to show help message in change shinyjs::html}
}
\description{
Full width button group with validation
}
\keyword{internal}
