% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{keep_curr_selected}
\alias{keep_curr_selected}
\title{Take currently selected row and keep it in the same position}
\usage{
keep_curr_selected(datasets, prev, curr)
}
\arguments{
\item{datasets}{current data.frame of single-cell datasets}

\item{prev}{previous data.frame of single-cell datasets}

\item{curr}{name of currently selected row from \code{prev}}
}
\value{
\code{datasets} with \code{curr} at same row as it is in \code{prev}
}
\description{
Used to stop dataset change when adding new datasets
}
