% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{labelTransferForm}
\alias{labelTransferForm}
\title{Logic for label transfer between datasets}
\usage{
labelTransferForm(
  input,
  output,
  session,
  sc_dir,
  datasets,
  show_label_transfer,
  dataset_name,
  scseq,
  species
)
}
\description{
Logic for label transfer between datasets
}
\keyword{internal}
