% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{labelTransferFormInput}
\alias{labelTransferFormInput}
\title{Input form for transfering labels between single cell datasets}
\usage{
labelTransferFormInput(id)

labelTransferFormInput(id)
}
\description{
Input form for transfering labels between single cell datasets

Input form for transfering labels between single cell datasets
}
\keyword{internal}
