% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{load_bulk_anals}
\alias{load_bulk_anals}
\title{Load previous bulk anals dataframe}
\usage{
load_bulk_anals(data_dir)
}
\arguments{
\item{data_dir}{Path to folder container \code{'bulk'} and \code{'single-cell'} directories}
}
\value{
data.frame with columns "dataset_name" "dataset_dir" and "anal_name".
}
\description{
used for determining available anals for download
}
\keyword{internal}
