% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-utils.R
\name{load_bulk_datasets}
\alias{load_bulk_datasets}
\title{Load previous bulk datasets dataframe}
\usage{
load_bulk_datasets(data_dir)
}
\arguments{
\item{data_dir}{Path to folder container \code{'bulk'} and \code{'single-cell'} directories}
}
\value{
data.frame with columns "dataset_name" and "dataset_dir"
}
\description{
Load previous bulk datasets dataframe
}
\keyword{internal}
