% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_cell_ranger_counts}
\alias{load_cell_ranger_counts}
\title{Load cell ranger counts}
\usage{
load_cell_ranger_counts(data_dir, h5)
}
\arguments{
\item{data_dir}{Directory with raw and kallisto/bustools or CellRanger quantified single-cell RNA-Seq files.}

\item{h5}{Boolean indicating, for \code{type = 'cell_ranger'}, if \code{data_dir} container a hdf5 file.}
}
\value{
dgCMatrix
}
\description{
Mainly to avoid having to download massive datasets that have already been quantified.
))))
}
