% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_cellranger_counts}
\alias{load_cellranger_counts}
\title{Load cell ranger counts}
\usage{
load_cellranger_counts(data_dir)
}
\value{
dgCMatrix
}
\description{
Mainly to avoid having to download massive datasets that have already been quantified.
))))
}
