% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{load_custom_anals}
\alias{load_custom_anals}
\title{Load data.frame of custom anals for selectizeInput}
\usage{
load_custom_anals(data_dir)
}
\arguments{
\item{data_dir}{Path to directory containing \code{'custom_queries'} folder.}
}
\value{
\code{data.frame} used to show available custom queries in Drugs tab
}
\description{
Load data.frame of custom anals for selectizeInput
}
\keyword{internal}
