% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{load_custom_results}
\alias{load_custom_results}
\title{Load results of previous custom query}
\usage{
load_custom_results(res_paths)
}
\arguments{
\item{res_paths}{List of named strings with paths to results.}
}
\value{
List of names results loaded from \code{res_paths}.
}
\description{
Will only load \code{res_paths} that exist. L1000 results won't exist
if no genes selected for custom query were in the L1000 measured genes.
}
\keyword{internal}
