% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_drug_es}
\alias{load_drug_es}
\title{Load drug effect size matrices for drug queries}
\usage{
load_drug_es()
}
\value{
list of matrices
}
\description{
Load drug effect size matrices for drug queries
}
\keyword{internal}
