% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_kallisto_counts}
\alias{load_kallisto_counts}
\title{Read kallisto/bustools market matrix and annotations}
\usage{
load_kallisto_counts(data_dir)
}
\arguments{
\item{data_dir}{Directory with raw and kallisto/bustools or CellRanger quantified single-cell RNA-Seq files.}
}
\value{
sparse dgTMatrix with barcodes in columns and genes in rows.
}
\description{
Read kallisto/bustools market matrix and annotations
}
