% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{load_pert_anals}
\alias{load_pert_anals}
\title{Load data.frame of CMAP02/L1000 perturbations for right click load signature on correlation points}
\usage{
load_pert_anals()
}
\value{
\code{data.frame} used to show available pertubation queries in Drugs tab
}
\description{
Load data.frame of CMAP02/L1000 perturbations for right click load signature on correlation points
}
\keyword{internal}
