% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{load_pert_signature}
\alias{load_pert_signature}
\title{Get perturbation signature}
\usage{
load_pert_signature(pert, pert_type, pert_signature_dir, pvals = FALSE)
}
\arguments{
\item{pert}{Name of perturbation signature.}

\item{pert_type}{One of \code{'cmap'}, \code{'l1000_genes'}, or \code{'l1000_drugs'}.}
}
\description{
Get perturbation signature
}
\keyword{internal}
