% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_resoln}
\alias{load_resoln}
\title{Load the applied resolution name}
\usage{
load_resoln(dataset_dir)
}
\arguments{
\item{dataset_dir}{directory corresponding to single-cell dataset}
}
\value{
the last applied resolution name (e.g. \code{'snn1'})
}
\description{
e.g. snn1
}
\keyword{internal}
