% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq}
\alias{load_scseq}
\title{Load single-cell RNA-Seq data into a SingleCellExperiment.}
\usage{
load_scseq(data_dir)
}
\arguments{
\item{Directory}{with raw and quantified single-cell RNA-Seq files.}
}
\value{
\code{SingleCellExperiment} with \code{colData} attribute including alevin whitelist.
}
\description{
Load single-cell RNA-Seq data into a SingleCellExperiment.
}
\examples{

data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
load_scseq(data_dir)

}
