% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{load_scseq_anals}
\alias{load_scseq_anals}
\title{Generate data.frame of saved single cell RNA-Seq analyses}
\usage{
load_scseq_anals(data_dir, with_type = FALSE)
}
\arguments{
\item{data_dir}{Directory to folder with single-cell analysis folders.}

\item{with_type}{if \code{TRUE} include column \code{'type'} with values \code{'Single Cell'}. Used for optgroupField in
selectizeInput's. Default is \code{FALSE}}
}
\value{
data.frame with columns: \itemize{
 \item dataset_name \code{NA} included for consistency with \code{load_bulk_anals}.
 \item dataset_dir Directory name in \code{data_dir} with single cell analysis.
 \item anal_name Name of single cell analysis. Same as \code{dataset_dir}.

}
}
\description{
Used together with \code{\link{load_bulk_anals}} to create data.frame
for selectizeInput choices.
}
\keyword{internal}
