% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq_counts}
\alias{load_scseq_counts}
\title{Read kallisto/bustools market matrix and annotations}
\usage{
load_scseq_counts(data_dir)
}
\value{
sparse dgTMatrix with barcodes in columns and genes in rows.
}
\description{
Read kallisto/bustools market matrix and annotations
}
