% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{load_scseq_datasets}
\alias{load_scseq_datasets}
\title{Generate data.frame of integrated single cell RNA-Seq datasets}
\usage{
load_scseq_datasets(data_dir)
}
\arguments{
\item{data_dir}{Directory to folder with single-cell analysis folders.}
}
\value{
data.frame with columns: \itemize{
 \item dataset_name \code{NA} included for consistency with \code{load_bulk_anals}.
 \item dataset_dir Directory name in \code{data_dir} with single cell analysis.
 \item anal_name Name of single cell analysis. Same as \code{dataset_dir}.

}
}
\description{
Used together with \code{\link{load_bulk_datasets}} to create data.frame
for selectizeInput choices.
}
\keyword{internal}
