% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_scseq_subsets}
\alias{load_scseq_subsets}
\title{Load subsets of scseqs for integration/subsetting}
\usage{
load_scseq_subsets(
  dataset_names,
  sc_dir,
  subset_metrics = NULL,
  is_include = NULL,
  ...
)
}
\arguments{
\item{dataset_names}{Character vector of single cell analysis names to load.}

\item{sc_dir}{The directory with single-cell datasets}
}
\value{
List of \code{SingleCellExperiment} objects.
}
\description{
Sets orig.ident to \code{ident}.
}
\keyword{internal}
