% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_scseqs_for_integration}
\alias{load_scseqs_for_integration}
\title{Load scRNA-Seq datasets for integration}
\usage{
load_scseqs_for_integration(anal_names, exclude_clusters, sc_dir, ident)
}
\arguments{
\item{anal_names}{Character vector of single cell analysis names to load.}

\item{sc_dir}{The directory with single-cell datasets}

\item{ident}{Either \code{'test'} or \code{'ctrl'}}
}
\value{
List of \code{Seurat} objects.
}
\description{
Will restore SCT assay if was saved seperately. Downsamples very large datasets.
Also sets orig.ident to \code{ident}.
}
\keyword{internal}
