% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{load_seq}
\alias{load_seq}
\title{Load raw RNA-Seq data into an ExpressionSet.}
\usage{
load_seq(data_dir, species = "Homo sapiens", release = "94",
  overwrite = FALSE, dgel = FALSE)
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link{build_index}}.}

\item{overwrite}{If FALSE (default) and a saved \code{ExpressionSet} exists, will load from disk.}

\item{dgel}{If TRUE, will also save the \code{DGEList} object. Used for testing purposes.}
}
\value{
\code{\link[Biobase]{ExpressionSet}} with attributes/accessors:
\itemize{
  \item \code{sampleNames} from names of raw RNA-Seq files (excluding .fastq.gz suffix).
  \item \code{annotation} Character vector of annotation package used.
  \item \code{exprs} Length scaled counts generated from abundances for use in
    \code{\link[limma]{voom}} (see \code{vignette("tximport", package = "tximport")}).
  \item \code{phenoData} from \code{pdata_path} annotation file with added columns:
    \itemize{
      \item \code{quants_dir} directory within '\code{data_dir}/quants' containing salmon quantification results.
      \item \code{lib.size} library size from \code{\link[edgeR]{calcNormFactors}}.
      \item \code{norm.factors} library normalization factors from \code{\link[edgeR]{calcNormFactors}}.
    }
}
}
\description{
Load raw RNA-Seq data into an ExpressionSet.
}
\examples{

data_dir <- 'data-raw/example-data'
eset <- load_seq(data_dir, overwrite = TRUE)

}
