% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{make_unique}
\alias{make_unique}
\title{Make Character Strings Unique}
\usage{
make_unique(names, sep = ".")
}
\arguments{
\item{names}{a character vector.}

\item{sep}{a character string used to separate a duplicate name from its sequence number.}
}
\value{
A character vector of same length as names with duplicates changed, in the current locale's encoding.
}
\description{
Make Character Strings Unique
}
