% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{match_pdata}
\alias{match_pdata}
\title{Match and add file names for pdata rows.}
\usage{
match_pdata(pdata, fastqs)
}
\arguments{
\item{pdata}{\code{data.table} with sample annotations.}

\item{fastqs}{Character vector of fastq.gz files to find matches for.}
}
\value{
\code{pdata} with \code{file} column identifying file for each sample (row).
}
\description{
Attempts to find exact match between file names (excluding fastq.gz suffix) and first pdata column.
If an exact match is not possible, \code{grep} is used to search for a unique substring.
}
\keyword{internal}
