% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{merge_linkouts}
\alias{merge_linkouts}
\title{Merge columns with image links}
\usage{
merge_linkouts(query_res, cols)
}
\arguments{
\item{query_res}{\code{data.frame} after calling \code{\link{add_linkout}} to \code{cols}.}

\item{cols}{Character vector of columns in \code{query_res} that \code{\link{add_linkout}} has been called on.}
}
\value{
\code{query_res} with column \code{'External Links'} formed from pasting \code{cols} together. \code{cols} are removed.
}
\description{
Merge columns with image links
}
\keyword{internal}
