% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples that are from the sample subject}
\usage{
merge_samples(data_dir, merge_list, new_founder = NULL)
}
\arguments{
\item{data_dir}{of single-cell}

\item{merge_list}{Names are new name, values are current names to merge}

\item{new_founder}{Name of new founder for dataset for grouping in select menu.
Default (\code{NULL}) keeps previous founder.}
}
\value{
called for side effects
}
\description{
Merge samples that are from the sample subject
}
\examples{

if (interactive()) {

  # FILL THIS IN
  sc_dir <- ''

  dataset_names <- c('SJIALD_VS_HEALTHY_merged_QC1.0_fastMNN')

  merge_list <- list(
    'SJIALD_20200716' = c('SJIALD_1-20200716_QC3.0_NO_NEUTRO', 'SJIALD_2-20200716_QC3.0_NO_NEUTRO'),
    'SJIALD_FID12518' = c('SJIALD_FID12518_Diseased_QC2.0'))

  new_founder <- 'SJIALD_VS_HEALTHY_merged'

  for (dataset_name in dataset_names) {
    data_dir <- file.path(sc_dir, dataset_name)
    merge_samples(data_dir, merge_list, new_founder)
  }

}

}
