% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{move_new}
\alias{move_new}
\title{Move new dataset to end of datasets data.frame}
\usage{
move_new(new, datasets)
}
\arguments{
\item{new}{Name of new dataset}

\item{datasets}{data.frame of datasets from \link{get_sc_dataset_choices}}
}
\value{
\code{datasets} with \code{new} moved to end
}
\description{
Used to prevent de-selection after integration/subset/loading
}
\keyword{internal}
