% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{navbar2UI}
\alias{navbar2UI}
\title{UI for secondary navbar}
\usage{
navbar2UI(hide)
}
\arguments{
\item{hide}{should delete and add be disabled (for demo)}
}
\value{
shiny.tag with html for secondary navbar
}
\description{
UI for secondary navbar
}
\examples{

navbar2UI(hide = FALSE)

}
