% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{navbarUI}
\alias{navbarUI}
\title{UI for navbar}
\usage{
navbarUI(tabs, active, with_logout = FALSE)
}
\arguments{
\item{tabs}{Character vector of tab names to display}

\item{active}{the active tab name}

\item{with_logout}{include logout link? \code{TRUE}
if \code{data_dir} is '/srv/drugseqr' otherwise \code{FALSE}}
}
\value{
shiny.tag with html for navbar
}
\description{
UI for navbar
}
