% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot.R
\name{NoLegend}
\alias{NoLegend}
\title{Removes the legend}
\usage{
NoLegend(...)
}
\arguments{
\item{...}{Extra parameters to be passed to \code{theme}}
}
\description{
Removes the legend
}
\examples{
# Generate a plot with no legend
library(ggplot2)
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + drugseqr:::NoLegend()

}
\keyword{internal}
