% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{norm_scseq}
\alias{norm_scseq}
\title{Normalize single-cell libraries for cell-specific biases}
\usage{
norm_scseq(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} returned from \code{\link{load_scseq}}.}
}
\value{
Size-factor normalized \code{SingleCellExperiment}.
}
\description{
Normalize single-cell libraries for cell-specific biases
}
