% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{normalize_scseq}
\alias{normalize_scseq}
\title{Utility wrapper to run normalization and log transformation}
\usage{
normalize_scseq(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object}
}
\value{
Normalized and log transformed \code{scseq}.
}
\description{
Utility wrapper to run normalization and log transformation
}
