% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{pairwise_wilcox}
\alias{pairwise_wilcox}
\title{Run pairwise wilcox tests between single cell clusters}
\usage{
pairwise_wilcox(
  scseq,
  groups = scseq$cluster,
  direction = "up",
  block = NULL,
  restrict = NULL
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{groups}{A vector of length equal to \code{ncol(x)}, specifying the group assignment for each cell.
If \code{x} is a SingleCellExperiment, this is automatically derived from \code{\link{colLabels}}.}

\item{direction}{A string specifying the direction of differences to be considered in the alternative hypothesis.}

\item{block}{A factor specifying the blocking level for each cell.}

\item{restrict}{A vector specifying the levels of \code{groups} for which to perform pairwise comparisons.}
}
\value{
List of \code{data.frame}s, one for each cluster.
}
\description{
Run pairwise wilcox tests between single cell clusters
}
\keyword{internal}
