% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-server.R
\name{pathPert}
\alias{pathPert}
\title{Logic for perturbation selection in Pathways tab}
\usage{
pathPert(input, output, session, type, queries, pert_signature_dir)
}
\description{
Logic for perturbation selection in Pathways tab
}
\keyword{internal}
