% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-ui.R
\name{pathPertInput}
\alias{pathPertInput}
\title{Input for perturbation selection for Pathways page}
\usage{
pathPertInput(id, type)
}
\description{
Input for perturbation selection for Pathways page
}
\keyword{internal}
