% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{pick_roryk_cutoff}
\alias{pick_roryk_cutoff}
\title{Pick Roryk knee point}
\usage{
pick_roryk_cutoff(bcs)
}
\arguments{
\item{bcs}{Vector of total counts for all barcodes}
}
\value{
count value corresponding to Roryk knee point
}
\description{
from https://github.com/COMBINE-lab/salmon/issues/362
}
\keyword{internal}
