% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_biogps}
\alias{plot_biogps}
\title{Plot BioGPS data for a HGNC symbol}
\usage{
plot_biogps(gene)
}
\arguments{
\item{gene}{Character vector of gene name.}
}
\value{
ggplot
}
\description{
Plot BioGPS data for a HGNC symbol
}
\keyword{internal}
