% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot TSNE coloured by cluster}
\usage{
plot_cluster(
  scseq = NULL,
  reduction = "TSNE",
  plot_data = NULL,
  legend = FALSE,
  cols = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{scseq}{SingleCellExperiment}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{legend}{Should legend be kept? Default is \code{FALSE}.}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character
or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
By default, ggplot2 assigns colors}

\item{title}{Title of plot}

\item{...}{additional argument to \code{DimPlot}}
}
\value{
\code{ggplot}
}
\description{
Plot TSNE coloured by cluster
}
\keyword{internal}
