% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_cluster_labels}
\alias{plot_cluster_labels}
\title{Generate plotly for cluster origins of integrated dataset}
\usage{
plot_cluster_labels(scseq, clust, sc_dir)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}

\item{clust}{A cluster in \code{scseq} to generate plotly for}

\item{sc_dir}{Directory containing original datasets used when integrating \code{scseq}}
}
\value{
plotly
}
\description{
Generate plotly for cluster origins of integrated dataset
}
\keyword{internal}
