% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_combined_reports.R
\name{plot_combined_scseq_report}
\alias{plot_combined_scseq_report}
\title{Plot grid of combined scRNA-seq clusters and gene markers together}
\usage{
plot_combined_scseq_report(scseq, markers, pt.size = 3)
}
\arguments{
\item{scseq}{\code{Seurat} object.}

\item{markers}{Named list with a single character vector specifying the marker genes to plot.
The name is the cluster that will be highlighted.}
}
\value{
List of \code{ggplot} objects
}
\description{
For integrated data with multiple samples.
}
\keyword{internal}
