% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{plot_dprimes}
\alias{plot_dprimes}
\title{Generate plotly of dprimes values for Drugs tab.}
\usage{
plot_dprimes(path_df, drugs = TRUE)
}
\arguments{
\item{path_df}{result of \link{get_path_df}.}
}
\value{
plotly
}
\description{
Also used for pseudobulk single-cell datasets.
}
\keyword{internal}
