% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_ridge}
\alias{plot_ridge}
\title{Plot single-cell ridgeline plots}
\usage{
plot_ridge(
  feature,
  scseq,
  selected_cluster,
  by.sample = FALSE,
  with.height = FALSE,
  decreasing = feature \%in\% c("ribo_percent", "log10_sum", "log10_detected")
)
}
\arguments{
\item{feature}{Feature name to generate ridge plot for. Either a row or \code{colData} of \code{scseq}.}

\item{scseq}{\code{SingleCellExperiment}.}

\item{selected_cluster}{Name of the selected cluster.}

\item{by.sample}{if \code{TRUE} plot \code{feature} ridge for each \code{scseq$batch}. Default (\code{FALSE})
will plot \code{feature} for each \code{scseq$cluster}.}

\item{with.height}{Whether to return height of plot. See value for details.}

\item{decreasing}{if \code{TRUE}, ridgelines with smaller mean values of \code{feature} will show up on top.
Used to show features where smaller values indicate potential QC issues.}
}
\value{
ggplot object
}
\description{
Plot single-cell ridgeline plots
}
\keyword{internal}
