% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_combined_reports.R
\name{plot_scseq_barplot}
\alias{plot_scseq_barplot}
\title{Plot scRNA-seq barplot}
\usage{
plot_scseq_barplot(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}
}
\value{
\code{ggplot}
}
\description{
Plots percentage of cells within each \code{scseq$cluster} split by \code{scseq$orig.ident}.
}
\keyword{internal}
