% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_reports.R
\name{plot_scseq_report}
\alias{plot_scseq_report}
\title{Plot grid with scRNA-seq cluster and gene markers together}
\usage{
plot_scseq_report(scseq, markers, point_size = 3)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat}.}

\item{markers}{Named list of \code{data.frame}s where \code{row.names} are the marker genes. One list per cluster in \code{scseq}
with the same name as the cluster.}

\item{point_size}{Numeric scalar, specifying the size of the points. Defaults to 3.}
}
\value{
\code{ggplot}
}
\description{
Plot grid with scRNA-seq cluster and gene markers together
}
