% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_feature}
\alias{plot_tsne_feature}
\title{Plot TSNE coloured by gene or QC metric}
\usage{
plot_tsne_feature(
  scseq,
  feature,
  cols = c("lightgray", "blue"),
  reverse_scale = feature \%in\% c("ribo_percent", "log10_sum", "log10_detected"),
  title = paste0("Expression by Cell: ", feature)
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{feature}{Character vector specifying feature to colour cells by.}

\item{cols}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.
When blend is \code{TRUE}, takes anywhere from 1-3 colors:
\describe{
  \item{1 color:}{Treated as color for double-negatives, will use default colors 2 and 3 for per-feature expression}
  \item{2 colors:}{Treated as colors for per-feature expression, will use default color 1 for double-negatives}
  \item{3+ colors:}{First color used for double-negatives, colors 2 and 3 used for per-feature expression, all others ignored}
}}

\item{reverse_scale}{Should the scale be reversed (lower values of feature
are darker)?}

\item{title}{Title to add to plot.}
}
\value{
\code{ggplot}
}
\description{
Plot TSNE coloured by gene or QC metric
}
\keyword{internal}
