% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_feature_sample}
\alias{plot_tsne_feature_sample}
\title{Format gene plots for sample comparison for dseqr app}
\usage{
plot_tsne_feature_sample(
  feature,
  scseq,
  group,
  plot = NULL,
  cols = c("lightgray", "blue")
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{group}{Level in \code{scseq$orig.ident} to show cells for. Either \code{'ctrl'} or \code{'test'}}
}
\value{
\code{plot} formatted for dseqr app
}
\description{
Format gene plots for sample comparison for dseqr app
}
\keyword{internal}
