% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_gene}
\alias{plot_tsne_gene}
\title{Plot TSNE coloured by HGNC symbol}
\usage{
plot_tsne_gene(sce, gene, selected_groups = NULL,
  assay.type = "logcounts", point_size = 3)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{gene}{Character vector to colour cells by}

\item{selected_groups}{The groups in \code{sce$orig.ident} to show cell for. The default \code{NULL} shows all cells.}

\item{assay.type}{The assay data that gets used by \code{\link[scater]{plotTSNE}}. If \code{scseq} is a \code{Seurat}
object then the default ('logcounts') will be from the \code{data} slot of the \code{Seurat::DefaultAssay}.}
}
\value{
ggplot
}
\description{
Plot TSNE coloured by HGNC symbol
}
