% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{plot_tsne_gene_sample}
\alias{plot_tsne_gene_sample}
\title{Format gene plots for sample comparison for drugseqr app}
\usage{
plot_tsne_gene_sample(gene, scseq, group = "test")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{group}{Level in \code{scseq$orig.ident} to show cells for. Either \code{'ctrl'} or \code{'test'}}
}
\value{
\code{plot} formatted for drugseqr app
}
\description{
Format gene plots for sample comparison for drugseqr app
}
\keyword{internal}
