% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_umap_cluster}
\alias{plot_umap_cluster}
\title{Plot UMAP coloured by cluster}
\usage{
plot_umap_cluster(scseq, selected_groups = NULL, pt.size = 3)
}
\arguments{
\item{scseq}{\code{Seurat}.}

\item{selected_groups}{The groups in \code{scseq$orig.ident} to show cell for. The default \code{NULL} shows all cells.}

\item{pt.size}{Numeric scalar, specifying the size of the points. Defaults to 3.}

\item{...}{Additional arguments passed to \code{Seurat::DimPlot}.}
}
\value{
\code{ggplot}
}
\description{
Plot UMAP coloured by cluster
}
