% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin Plot}
\usage{
plot_violin(
  feature = NULL,
  scseq = NULL,
  selected_cluster = NULL,
  by.sample = FALSE,
  with_all = FALSE,
  with.height = FALSE,
  violin_data = NULL,
  is_mobile = FALSE,
  hide_pct = is_mobile,
  hide_ncells = is_mobile,
  decreasing = feature \%in\% c("ribo_percent", "log10_sum", "log10_detected")
)
}
\arguments{
\item{feature}{Character, gene name.}

\item{scseq}{\code{SingleCellExperiment} object.}

\item{selected_cluster}{Integer indicating which \code{level(scseq$cluster)} to highlight
(if \code{by.sample} is \code{FALSE}) or to subset data to (if \code{by.sample} is \code{TRUE}).
To plot by sample without subsetting to a cluster, set to \code{nlevels(scseq$cluster) + 1}.}

\item{by.sample}{Boolean indicating if violin plots should be by sample or by cluster (default).}

\item{with_all}{Set to \code{TRUE} if plotting by sample to get same colors as Shiny app.}

\item{with.height}{if \code{TRUE} returns a list object with plot and appropriate height.}

\item{violin_data}{Used for Shiny app. Result of \code{get_violin_data}.
Used when logcounts are seperate from \code{SingleCellExperiment} as in the Shiny app.}

\item{is_mobile}{If \code{TRUE} shortens up y-axis labels to fit mobile better.}

\item{hide_pct}{If \code{TRUE}, barplots indicating number of cells are hidden.}

\item{hide_ncells}{If \code{TRUE}, barplots indicating percent expression are hidden.}

\item{decreasing}{if \code{FALSE} (default) violins are ordered, bottom to top,
by increasing mean expression.}
}
\value{
\code{ggplot} or list
}
\description{
Violin Plot
}
