% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-utils.R
\name{plotlyCells}
\alias{plotlyCells}
\title{Generate boxplotly for vsd normalized gene data by group for Datasets tab}
\usage{
plotlyCells(pdata, dtangle_est)
}
\arguments{
\item{pdata}{data.frame with columns \code{'Group name'} (character) and \code{'Group'} (integer).}

\item{eset}{ExpressionSet object with \code{'vsd'} assayDataElement}

\item{explore_genes}{Character vector of genes to plot}
}
\value{
plotly
}
\description{
Generate boxplotly for vsd normalized gene data by group for Datasets tab
}
