% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-utils.R
\name{plotlyGene}
\alias{plotlyGene}
\title{Generate boxplotly for vsd normalized gene data by group for Datasets tab}
\usage{
plotlyGene(eset, explore_genes, pdata)
}
\arguments{
\item{eset}{ExpressionSet object with \code{'vsd'} assayDataElement}

\item{explore_genes}{Character vector of genes to plot}

\item{pdata}{data.frame with columns \code{'Group name'} (character) and \code{'Group'} (integer).}
}
\value{
plotly
}
\description{
Generate boxplotly for vsd normalized gene data by group for Datasets tab
}
