% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot sammon MDS with and without SVA}
\usage{
plotMDS(exprs, exprs_sva, group)
}
\arguments{
\item{exprs}{\code{matrix} of expression values.}

\item{exprs_sva}{\code{matrix} of expression values with surrogate variables regressed out.}

\item{group}{Character vector with values \code{'control'} and \code{'test'} indicating group membership.}
}
\description{
Plots interactive MDS plot of expression values with and without surrogate variable analysis.
}
