% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{plotStylesButton}
\alias{plotStylesButton}
\title{Button with sliders for adjusting plot jitter and point size}
\usage{
plotStylesButton(id)
}
\description{
Button with sliders for adjusting plot jitter and point size
}
