% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{prevent_integrated}
\alias{prevent_integrated}
\title{Switch away from integrated assay slot}
\usage{
prevent_integrated(scseq)
}
\arguments{
\item{scseq}{\code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
\code{scseq} with "SCT" as the default assay if it was previously "integrated".
}
\description{
The "integrated" assay slot is inappropriate for differential expression analysis. This will switch to \code{SCTransform}'ed
slot for \code{Seurat} objects. \code{SingleCellExperiment} objects check to make sure they weren't generated from
"integrated" data.
}
