% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{process_raw_scseq}
\alias{process_raw_scseq}
\title{Process Count Data for App}
\usage{
process_raw_scseq(
  scseq,
  dataset_name,
  sc_dir,
  founder = NULL,
  progress = NULL,
  value = 0
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}

\item{dataset_name}{Name of dataset to save}

\item{sc_dir}{Directory to save dataset to}

\item{progress}{Shiny progress object. Default (\code{NULL}) prints to stdout.}

\item{value}{Initial value of progress.}

\item{score_doublets}{Should doublet scores be computed? Intended for non-subsetted data within \code{\link{load_raw_scseq}}.
Default is \code{FALSE}.}
}
\description{
Process Count Data for App
}
