% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R, R/strain_scseq.R
\name{qc_scseq}
\alias{qc_scseq}
\title{Calculate QC metrics for SingleCellExperiment}
\usage{
qc_scseq(scseq)

qc_scseq(scseq)
}
\arguments{
\item{scseq}{Seurat object.}

\item{sce}{}
}
\value{
\code{scseq} with low quality cells removed.
}
\description{
This is used for filtering low quality cells from kallisto quantification after running \code{get_empty} and
either \code{strain_scseq} or just creating a Seurat object from the counts with empty droplets removed.
}
\details{
Removes cells that are three median absolute deviations above or below the median for:
percentage mitochondrial counts, percentage ribosomal counts, number of expressed features, and
number of total counts.
}
