% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{qread.safe}
\alias{qread.safe}
\title{Read qs file safely}
\usage{
qread.safe(file, .nofile = NULL, .nullfile = NULL)
}
\arguments{
\item{file}{the file name/path}

\item{.nofile}{return value when file doesn't exist}

\item{.nullfile}{return value when file value is \code{NULL}}
}
\value{
The de-serialized object
}
\description{
Read qs file safely
}
\keyword{internal}
