% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{query_budger}
\alias{query_budger}
\title{Find drugs that maximally affect a selection of genes}
\usage{
query_budger(query_genes, drug_es)
}
\arguments{
\item{query_genes}{Named list of character vectors with \code{'dn'} indicating genes that want to down-regulated and
\code{'up'} indicating genes that want to up-regulate.}

\item{drug_es}{A matrix of differential expression values for drugs.}
}
\description{
Results are based on the average effect on the query genes. Genes to upregulate are multiplied by -1 so that
strong positive results for these genes contribute towards a more negative average effect. All results are divided
by the absolute of the average minimum effect (after -1 multiplication of genes to upregulate) of the query genes.
This ensures that results will be between -1 and 1 for consistency with correlation values for \code{\link{query_drugs}}.
}
\keyword{internal}
