% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{query_drugs}
\alias{query_drugs}
\title{Get correlation between query and drug signatures.}
\usage{
query_drugs(query_genes, drug_es, ngenes = 200)
}
\arguments{
\item{query_genes}{Named numeric vector of differential expression values for
query genes. Names are HGNC symbols.}

\item{drug_es}{A matrix of differential expression values for drugs.}

\item{ngenes}{The number of top differentially-regulated (up and down) query genes to use.}
}
\value{
Named vector of pearson correlations between query and drug combination signatures.
}
\description{
Determines the pearson correlation between the query and each drug signature.
}
\details{
Drugs with the largest positive and negative pearson correlation are predicted to,
respectively, mimic and reverse the query signature. Values range from +1 to -1.
}
\examples{

# load CMAP02 data
cmap_es <- drugseqr.data::load_drug_es('cmap_es_ind.rds')

# use first cmap_es signature as query
dprimes <- cmap_es[1:100, 1]
names(dprimes) <- row.names(cmap_es)[1:100]

# get correlations between query and drug signatures
res <- drugseqr:::query_drugs(dprimes, cmap_es)

}
\keyword{internal}
