% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{querySignatureInput}
\alias{querySignatureInput}
\title{Query signature input for Drugs page}
\usage{
querySignatureInput(id)
}
\description{
Query signature input for Drugs page
}
\keyword{internal}
