% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{read.delim1}
\alias{read.delim1}
\title{Helper function to read single column text files}
\usage{
read.delim1(file)
}
\arguments{
\item{file}{File to read from}
}
\value{
Character vector from \code{file}.
}
\description{
Helper function to read single column text files
}
