% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{read_alevin_mm}
\alias{read_alevin_mm}
\title{Read alevin 0.14.0 quants_mat.mtx.gz output}
\usage{
read_alevin_mm(alevin_dir)
}
\arguments{
\item{files}{path to alevin directory}
}
\value{
matrix of counts
}
\description{
Read alevin 0.14.0 quants_mat.mtx.gz output
}
\keyword{internal}
