% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto.R
\name{read_kallisto}
\alias{read_kallisto}
\title{Read kallisto/bustools market matrix and annotations}
\usage{
read_kallisto(data_dir)
}
\arguments{
\item{data_dir}{Path to folder with fastq.gz scRNA-seq files}
}
\value{
sparse dgTMatrix with barcodes in columns and genes in rows.
}
\description{
Primarily for testing purposes. Recommendation is to use \code{run_alevin} and \code{load_scseq}.
}
\examples{

}
