% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{readAlevin}
\alias{readAlevin}
\title{Read alevin 0.14.0 output}
\usage{
readAlevin(files)
}
\arguments{
\item{files}{path to alevin quants_mat.gz output}
}
\value{
matrix of counts
}
\description{
Read alevin 0.14.0 output
}
\keyword{internal}
