% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{readAlevin_sparse}
\alias{readAlevin_sparse}
\title{Read alevin 0.14.0 quants_mat.gz output}
\usage{
readAlevin_sparse(files)
}
\arguments{
\item{files}{path to quants_mat.gz files}
}
\value{
sparse dgTMatrix
}
\description{
Read alevin 0.14.0 quants_mat.gz output
}
\keyword{internal}
