% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{readRDS.safe}
\alias{readRDS.safe}
\title{Read RDS file safely}
\usage{
readRDS.safe(file, .nofile = NULL, .nullfile = NULL)
}
\arguments{
\item{file}{a \link[base]{connection} or the name of the file where the \R object
    is saved to or read from.}

\item{.nofile}{return value when file doesn't exist}

\item{.nullfile}{return value when file value is \code{NULL}}
}
\value{
For \code{readRDS}, an \R object.

  For \code{saveRDS}, \code{NULL} invisibly.
}
\description{
Read RDS file safely
}
\keyword{internal}
