% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{reduce_dims}
\alias{reduce_dims}
\title{Perform PCA and TSNE dimensionality reduction}
\usage{
reduce_dims(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object with \code{'hvg'} in \code{rowData}}
}
\value{
\code{sce} with \code{'PCA'} and \code{'TSNE'} reducedDim slots.
}
\description{
Runs after \code{add_hvgs} so that runs faster and on interesting genes.
}
