% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{remove_axes}
\alias{remove_axes}
\title{Remove ggplot axis title, text, and ticks}
\usage{
remove_axes(plot, types)
}
\arguments{
\item{plot}{ggplot}

\item{types}{Character vector of axis types to remove. Either or both of \code{'xaxis'} and \code{'yaxis'}.}
}
\value{
ggplot
}
\description{
Remove ggplot axis title, text, and ticks
}
