% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{remove_bulk_anals}
\alias{remove_bulk_anals}
\title{Remove analysis info from anals dataframe}
\usage{
remove_bulk_anals(dataset_name, data_dir)
}
\arguments{
\item{dataset_name}{Name of dataset}

\item{data_dir}{Path to folder container \code{'bulk'} and \code{'single-cell'} directories}
}
\description{
Used to clear previous analysis when groupings change
}
\keyword{internal}
