% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{remove_dataset_files}
\alias{remove_dataset_files}
\title{Delete stale dataset files after changing sample groups or running sva}
\usage{
remove_dataset_files(
  data_dir,
  patterns = c("^adjusted_\\\\d+svs.rds$", "^iqr_keep_\\\\d+svs.rds$", "^svobj.rds$",
    "^numsv.rds$", "^lm_fit_\\\\d+svs.rds$", "cmap_res_.+_\\\\d+svs.rds$",
    "goana_.+_\\\\d+svs.rds$", "kegga_.+_\\\\d+svs.rds$",
    "l1000_drugs_res_.+_\\\\d+svs.rds$", "l1000_genes_res_.+_\\\\d+svs.rds$",
    "diff_expr_symbol_.+_\\\\d+svs.rds$"),
  exclude = NULL
)
}
\arguments{
\item{data_dir}{Path to folder with dataset files}

\item{patterns}{patterns to remove. Default is all.}
}
\description{
Delete stale dataset files after changing sample groups or running sva
}
\keyword{internal}
