% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{req}
\alias{req}
\title{Check for required values}
\usage{
req(..., cancelOutput = FALSE)
}
\arguments{
\item{...}{Values to check for truthiness.}

\item{cancelOutput}{If TRUE and an output is being evaluated, stop processing
as usual but instead of clearing the output, leave it in whatever state it happens to be in.}
}
\value{
The first value that was passed in.
}
\description{
Check for required values
}
