% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{rightClickMenu}
\alias{rightClickMenu}
\title{Custom right click menu for selecting correlation point as query}
\usage{
rightClickMenu()
}
\description{
Custom right click menu for selecting correlation point as query
}
\keyword{internal}
