% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_comparison}
\alias{run_comparison}
\title{Run genes and pathways test vs ctrl scseq comparisons}
\usage{
run_comparison(
  scseq,
  selected_clusters,
  sc_dir,
  anal_name,
  session,
  with_path = FALSE,
  with_drugs = FALSE
)
}
\arguments{
\item{scseq}{\code{Suerat} object}

\item{selected_clusters}{the selected clusters in \code{Seurat::Idents(scseq)}}

\item{sc_dir}{Path to directory with single cell analysis folders.}

\item{anal_name}{Name of analysis. A directory in \code{sc_dir}.}

\item{session}{Shiny session object used for progress bar.}

\item{with_path}{Boolean to include pathway analysis or not.}

\item{with_drugs}{Boolean to include drug query results or not.}
}
\value{
Named list with names. \code{anal} \code{cluster_markers} and optionally \code{path}.
}
\description{
Used in Single Cell tab (no pathway analysis) and Pathways tab (with pathway analysis).
}
\keyword{internal}
