% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{run_custom_query}
\alias{run_custom_query}
\title{Run custom query}
\usage{
run_custom_query(query_genes, res_paths, session)
}
\arguments{
\item{query_genes}{Named list of character vectors with \code{'up'} indicating genes to upregulate and \code{'dn'}
containing genes to downregulate.}

\item{res_paths}{List of paths to save results to.}

\item{session}{session object from Shiny. Used to indicate progress.}
}
\value{
List with query results.
}
\description{
Run custom query
}
\keyword{internal}
