% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_dream}
\alias{run_dream}
\title{Run dream mixed effect model on RNA-seq ExpressionSet}
\usage{
run_dream(eset)
}
\arguments{
\item{eset}{\code{ExpressionSet} with 'lib.size', 'norm.factors', 'group' and 'pair' columns in \code{pData}.}
}
\value{
\code{\link[limma]{topTable}} result from \code{\link[varianceParition]{dream}} with 'group'
 treated as fixed effect and 'pair' as a random effect.
}
\description{
Run dream mixed effect model on RNA-seq ExpressionSet
}
