% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_drug_queries}
\alias{run_drug_queries}
\title{Run drug queries}
\usage{
run_drug_queries(top_table, drug_paths, es, ngenes = 200)
}
\value{
\code{res} with drug query results added to \code{'cmap'} \code{'l1000'} slots.
}
\description{
Used by run_comparison
}
\keyword{internal}
