% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_drugs_comparison}
\alias{run_drugs_comparison}
\title{Run drugs comparison}
\usage{
run_drugs_comparison(res_paths, session, res = list(), ambient = NULL)
}
\arguments{
\item{res_paths}{List of paths with names \code{'cmap'} \code{'l1000'} and \code{'anal'} to
saved cmap, l1000, and differential expression analysis results.}

\item{session}{Shiny session object used for progress bar.}

\item{res}{List to store drug query results in.}

\item{ambient}{Character vector of ambient genes to exclude from drug queries.}
}
\value{
\code{res} with drug query results added to \code{'cmap'} \code{'l1000'} slots.
}
\description{
Used by run_comparison
}
\keyword{internal}
