% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{run_drugseqr}
\alias{run_drugseqr}
\title{Run drugseqr app}
\usage{
run_drugseqr(data_dir, pert_query_dir = NULL, test = FALSE,
  test_data = FALSE)
}
\arguments{
\item{data_dir}{Directory containing folders \code{'bulk'} \code{'single-cell'}.
Ignored if \code{test_data} is \code{TRUE}.}

\item{pert_query_dir}{Path to directory where pert query results (using CMAP02/L1000 as query signature) will be downloaded as requested.}

\item{test}{Boolean indicating if \code{shinytest} should be run (default is \code{FALSE}).
If \code{TRUE} test data will be used.}

\item{test_data}{Boolean indicating if test data should be used. Default is \code{TRUE}}
}
\description{
Used for local development
}
\examples{

data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/IBD_20190712'
data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/IBD'
data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/amnon'
data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/sjia'
data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/example'
pert_query_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/drug_gene_queries/data'

run_drugseqr(data_dir, pert_query_dir, test_data = TRUE)
run_drugseqr(data_dir, pert_query_dir, test_data = FALSE)

}
