% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr_table.R
\name{run_drugseqr_table}
\alias{run_drugseqr_table}
\title{Run app with drug query table only}
\usage{
run_drugseqr_table(
  query_res,
  drug_study = "L1000 Drugs",
  cells = NULL,
  sort_by = "avg_cor",
  show_clinical = TRUE,
  min_signatures = 3,
  is_pert = FALSE,
  direction = "opposing",
  ntop = "all",
  app_dir = system.file("dt", package = "drugseqr", mustWork = TRUE),
  host = "0.0.0.0",
  port = 3838
)
}
\arguments{
\item{query_res}{Numeric vector of correlation values with names as HGNC symbols}

\item{drug_study}{Queried database giving rise to \code{query_res}.
One of \code{'CMAP02'},\code{'L1000 Genetic'}, or \code{'L1000 Drugs'}.}

\item{cells}{Character vector of cell types to include. Default (\code{NULL}) includes all cell types.}

\item{sort_by}{Metric to sort by. Either \code{'avg_cor'} (default) or \code{'min_cor'}.}

\item{show_clinical}{Should result only contain drugs with clinical phase annotation? Default is \code{TRUE}.}

\item{min_signatures}{Number of independent perturbagen signatures below which a perturbation is excluded. Default is \code{3}.}

\item{direction}{Direction of correlation to sort results by. One of \code{'both'}, \code{'similar'}, or \code{'opposing'} (default).}
}
\value{

}
\description{
Used to show full query table and use programmatic filter options
}
