% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dseqr.R
\name{run_dseqr}
\alias{run_dseqr}
\title{Run dseqr app}
\usage{
run_dseqr(
  app_name,
  data_dir = "/srv/dseqr",
  app_dir = system.file("app", package = "dseqr", mustWork = TRUE),
  pert_query_dir = file.path(data_dir, "pert_query_dir"),
  pert_signature_dir = file.path(data_dir, "pert_signature_dir"),
  gs_dir = file.path(data_dir, "gs_dir"),
  indices_dir = file.path(data_dir, "indices"),
  tabs = c("Single Cell", "Bulk Data", "Drugs"),
  test = FALSE,
  test_data = FALSE,
  host = "0.0.0.0",
  port = 3838,
  logout_url = NULL,
  is_example = FALSE
)
}
\arguments{
\item{app_name}{Name for new dseqr app.}

\item{data_dir}{Directory containing folder \code{app_name} with saved analyses.}

\item{app_dir}{Directory containing shiny app files. Can be 'inst/app' if working from source code.}

\item{pert_query_dir}{Path to directory where pert query results (using CMAP02/L1000 as query signature) will be downloaded as requested.}

\item{pert_signature_dir}{Path to directory where pert signatures for CMAP02/L1000 will be downloaded as requested.}

\item{indices_dir}{Path to directory containing \code{kallisto} indices and whitelists.}

\item{tabs}{Character vector of tabs to include in order desired. Must be subset of 'Single Cell', 'Bulk Data', and 'Drugs'.}

\item{test}{Boolean indicating if \code{shinytest} should be run (default is \code{FALSE}).
If \code{TRUE} test data will be used.}

\item{test_data}{Boolean indicating if test data should be used. Default is \code{TRUE}}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}
}
\value{
Runs dseqr app
}
\description{
Run dseqr application to explore single-cell and bulk RNA-seq datasets.
}
\examples{

# create directory structure for new datasets
data_dir <- tempdir()
app_name <- 'example'
init_dseqr(app_name, data_dir)


# run app
# run_dseqr(app_name, data_dir)

}
