% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{run_fasterq_dump}
\alias{run_fasterq_dump}
\title{Run fasterq-dump from ncbi/sra-tools}
\usage{
run_fasterq_dump(srr_name, out_dir, args = NULL)
}
\arguments{
\item{srr_name}{Run name starting with \code{'SRR'}}

\item{out_dir}{Directory to dump fastq files into.}

\item{args}{Character vector of additional arguments to fasterq-dump}
}
\description{
Downloads SRR.sra file and converts to fastq files.
}
\keyword{internal}
