% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_harmony}
\alias{run_harmony}
\title{Run harmony integration}
\usage{
run_harmony(logcounts, subset_row, batch, pairs = NULL)
}
\arguments{
\item{logcounts}{dgCMatrix of logcounts.}

\item{subset_row}{Vector specifying the subset of features to use for dimensionality reduction.
This can be a character vector of row names, an integer vector of row indices or a logical vector.}

\item{pairs}{}
}
\value{
Integrated \code{DietSCE} with logcounts assay, corrected reducedDim, and batch annotation.
}
\description{
Run harmony integration
}
\keyword{internal}
