% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto.R
\name{run_kallisto}
\alias{run_kallisto}
\title{Run kallisto/bustools for quantifying scRNA-seq}
\usage{
run_kallisto(index_path, whitelist_path, data_dir)
}
\arguments{
\item{index_path}{Path to kallisto index file}

\item{whitelist_path}{path to 10X Chromium version 2 whitelist}

\item{data_dir}{Path to folder with fastq.gz scRNA-seq files}
}
\description{
Primarily for testing purposes. Recommendation is to use \code{run_alevin}.
}
