% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{run_kallisto_scseq_commands}
\alias{run_kallisto_scseq_commands}
\title{Runs kallisto scseq commands}
\usage{
run_kallisto_scseq_commands(bus_args, whitepath, out_dir, inspection = FALSE)
}
\arguments{
\item{bus_args}{arguments to \code{system2} for kallisto bus}

\item{whitepath}{Path to whitelist}

\item{out_dir}{Directory to write to}

\item{inspection}{Boolean indicating if kallisto inspect should be run. Used for \code{detect_10x_chemistry}.}
}
\value{
stdout from kallisto inspect if \code{inspection = TRUE}, otherwise \code{NULL}.
}
\description{
Runs kallisto scseq commands
}
\keyword{internal}
