% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kb_scseq.R
\name{run_kb_scseq}
\alias{run_kb_scseq}
\title{Run kallisto/bustools for quantifying 10X scRNA-seq data}
\usage{
run_kb_scseq(
  indices_dir,
  data_dir,
  out_dir = file.path(data_dir, "bus_output"),
  species = "human",
  threads = 4
)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices built with \code{build_kallisto_index}.}

\item{data_dir}{Path to folder with 10X fastq.gz scRNA-seq files}

\item{out_dir}{Path to save output to.}

\item{species}{Species to use index from. Currently supports 'human' or 'mouse'.}

\item{threads}{Number of threads to use. Default is 1.}
}
\value{
called for side effects.
}
\description{
Run kallisto/bustools for quantifying 10X scRNA-seq data
}
\seealso{
\link{download_kb_index}
}
