% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_liger}
\alias{run_liger}
\title{Run liger integration}
\usage{
run_liger(logcounts, scseqs, batch)
}
\arguments{
\item{logcounts}{dgCMatrix of logcounts.}

\item{scseqs}{list of \code{SingleCellExperiment} objects to integrate.}

\item{batch}{Character vector indicating the original datasets for each cell in scseqs.
Must be in same order as \code{scseqs}.}
}
\value{
Integrated \code{SingleCellExperiment} with logcounts assay, corrected reducedDim, and batch annotation.
}
\description{
Run liger integration
}
\keyword{internal}
