% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{run_limma_pbulk}
\alias{run_limma_pbulk}
\title{Run pseudo bulk limma fit}
\usage{
run_limma_pbulk(dataset_dir, species = "Homo sapiens", release = "94")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object}
}
\description{
Run pseudo bulk limma fit
}
