% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_limma_scseq}
\alias{run_limma_scseq}
\title{Run limma on pseudobulk experiment}
\usage{
run_limma_scseq(
  summed,
  meta,
  species,
  trend = FALSE,
  method = "TMMwsp",
  with_fdata = FALSE,
  progress = NULL,
  value = 0,
  ...
)
}
\arguments{
\item{summed}{pseudobulk \code{SingleCellExperiment}. If \code{NULL}, \code{summed_path} must be supplied
and the call is assumed to originate from \link[callr]{r_bg}.}

\item{meta}{data.frame of sample metadata with column \code{group} and \code{row.names}
set with \code{summed$batch}}

\item{species}{species name used to retrieve feature annotation.}

\item{trend}{Should limma-trend analysis be run (default is \code{FALSE})? Slightly faster than
limma-voom (default). Used for pseudobulk grid differential expression analyses.}

\item{method}{normalization method to be used.}

\item{with_fdata}{if \code{TRUE}, adds Entrez IDs to fit object which are needed to
downstream GO pathway analyses.}

\item{progress}{progress object or \code{NULL}.}

\item{value}{Initial value of progress.}

\item{...}{additional arguments to \link[edgeR]{filterByExpr}.}
}
\value{
List with a fit object, model matrix, and normalized expression matrix.
}
\description{
Run limma on pseudobulk experiment
}
\keyword{internal}
