% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{run_lmfit}
\alias{run_lmfit}
\title{Perform lmFit analysis from limma.}
\usage{
run_lmfit(eset, mod, rna_seq = TRUE)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}. Non-selected samples
and duplicate features removed by \code{add_contrasts} and
\code{iqr_replicates}.}

\item{mod}{Model matrix generated by \code{diff_setup}. With
or without surrogate variables.}

\item{rna_seq}{Is this an RNA-seq \code{eset}? Default is \code{TRUE}.}
}
\value{
result from call to limma \code{lmFit}.
}
\description{
If paired samples, runs \code{\link{duplicateCorrelation}} to estimate intra-patient variance.
}
