% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_path_comparison}
\alias{run_path_comparison}
\title{Run single cell RNA-Seq test vs ctrl pathway comparison}
\usage{
run_path_comparison(scseq, selected_clusters, sc_dir, anal_name, res,
  ambient)
}
\arguments{
\item{scseq}{\code{Suerat} object}

\item{selected_clusters}{Character vector of selected clusters to run comparison for.}

\item{sc_dir}{Path to folder with single cell analyses.}

\item{anal_name}{Folder name in \code{sc_dir} that contains single cell analysis.}

\item{res}{List with previous differential expression analysis results in slot \code{'anal'}.}

\item{ambient}{Character vector of ambient genes to exclude from pathway analysis.}
}
\value{
\code{res} with slot \code{'path'} containing pathway analysis results. The \code{'anal'}
 slot is also subsetted such that ambient genes are excluded.
}
\description{
Used by \code{run_comparison} for pathway analysis.
}
