% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA on SingleCellExperiment}
\usage{
run_pca(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} with \code{rowData(sce)$hvg} a boolean
vector indicating rows (genes) to use as highly variable genes.}
}
\value{
\code{sce} with PCA reduction
}
\description{
Run PCA on SingleCellExperiment
}
\keyword{internal}
