% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_post_cluster}
\alias{run_post_cluster}
\title{Post-Cluster Processing of SingleCellExperiment}
\usage{
run_post_cluster(
  scseq,
  dataset_name,
  sc_dir,
  resoln = 1,
  progress = NULL,
  value = 0,
  reset_annot = TRUE
)
}
\arguments{
\item{scseq}{SingleCellExperiment}

\item{dataset_name}{Name of dataset}

\item{sc_dir}{Directory with single-cell datasets}

\item{resoln}{resolution parameter}

\item{progress}{progress}

\item{value}{value}

\item{reset_annot}{if \code{TRUE} then overwrites saved annotation with cluster indices.
Set to \code{FALSE} when Azimuth is used as cluster annotation is defined.}
}
\description{
Performs the following:
- downsampling (used for label transfer)
- pseudobulk
- saving data
}
