% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{run_reduction}
\alias{run_reduction}
\title{Run TSNE or UMAP}
\usage{
run_reduction(sce, type = c("auto", "TSNE", "UMAP"), dimred = "PCA")
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{type}{One of \code{c('auto', 'TSNE', 'UMAP')}. If \code{'auto'} (default),
runs UMAP is more than 5000 cells, otherwise TSNE.}

\item{dimred}{reducedDim to run TSNE or UMAP on}
}
\value{
\code{sce} with \code{'TSNE'} or \code{'UMAP'} \code{reducedDim}
}
\description{
Run TSNE or UMAP
}
\keyword{internal}
