% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_salmon_bulk.R
\name{run_salmon_bulk}
\alias{run_salmon_bulk}
\title{Runs salmon quantification.}
\usage{
run_salmon_bulk(
  indices_dir,
  data_dir,
  pdata = NULL,
  species = "homo_sapiens",
  flags = c("--validateMappings", "--posBias", "--seqBias", "--gcBias")
)
}
\arguments{
\item{indices_dir}{Directory with salmon indices. See \code{\link[drugseqr.data]{build_salmon_index}}.}

\item{data_dir}{Directory with raw fastq.gz RNA-Seq files.}

\item{pdata}{Previous result of call to \code{run_salmon} or \code{\link{select_pairs}}. Used to bypass another call to \code{select_pairs}.}

\item{species}{Species name. Default is \code{homo_sapiens}.
Used to determine transcriptome index to use.}

\item{flags}{Character vector of flags to pass to salmon.}
}
\description{
For pair-ended experiments, reads for each pair should be in a seperate file.
}
\examples{

# first place IBD data in data-raw/example-data
indices_dir <- '~/Documents/Batcave/zaklab/drugseqr.data/inst/indices'
data_dir <- file.path('data-raw', 'example-data')
pdata_path <- file.path(data_dir, 'Phenotypes.csv')
run_salmon_bulk(indices_dir, data_dir, pdata_path)

}
