% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_qc.R
\name{run_scseq_qc}
\alias{run_scseq_qc}
\title{Run quality control for single-cell dataset}
\usage{
run_scseq_qc(
  sce,
  metrics = c("low_lib_size", "low_n_features", "high_subsets_mito_percent",
    "low_subsets_ribo_percent", "high_doublet_score")
)
}
\arguments{
\item{sce}{\code{DietSCE}.}

\item{metrics}{Character vector of metrics to remove outliers for.}
}
\value{
\code{sce} with outliers removed.
}
\description{
Run quality control for single-cell dataset
}
\keyword{internal}
