% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{run_sva}
\alias{run_sva}
\title{Run surrogate variable analysis}
\usage{
run_sva(mods, eset, rna_seq = TRUE)
}
\arguments{
\item{mods}{result of \code{get_mods}}

\item{eset}{Expression set.}

\item{rna_seq}{Boolean is it RNA-seq?}
}
\description{
Run surrogate variable analysis
}
