% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_tsne}
\alias{run_tsne}
\title{Run TSNE}
\usage{
run_tsne(sce, dimred = "PCA")
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{dimred}{reducedDim to run TSNE on}
}
\value{
\code{sce} with \code{'TSNE'} \code{reducedDim}
}
\description{
Run TSNE
}
