% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP for visualizing single cell data.}
\usage{
run_umap(scseq, dims = 1:30, reduction = "pca")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object.}
}
\value{
\code{scseq} with TSNE results.
}
\description{
If \code{scseq} is a \code{SingleCellExperiment} object then uses \code{scater::runUMAP}.
If \code{scseq} is a \code{Seurat} object then uses \code{Seurat::RunUMAP}.
}
