% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_varPart}
\alias{run_varPart}
\title{Run variance partition on RNA-seq ExpressionSet}
\usage{
run_varPart(eset)
}
\arguments{
\item{eset}{\code{ExpressionSet} with 'lib.size', 'norm.factors', 'group' and 'pair' columns in \code{pData}.}
}
\value{
results of call to \code{\link[varianceParition]{fitExtractVarPartModel}} with
 'group' and 'pair' covariates treated as random effects
}
\description{
Run variance partition on RNA-seq ExpressionSet
}
