% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{salmon_lt_0.14.0}
\alias{salmon_lt_0.14.0}
\title{Check if salmon version is less than 0.14.0}
\usage{
salmon_lt_0.14.0(salmon_version)
}
\arguments{
\item{salmon_version}{String giving salmon version}
}
\value{
\code{TRUE} if salmon version is less than 0.14.0, otherwise \code{FALSE}.
}
\description{
Check if salmon version is less than 0.14.0
}
\keyword{internal}
