% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{sampleComparison}
\alias{sampleComparison}
\title{Logic to for sample comparison input}
\usage{
sampleComparison(input, output, session, selected_anal, scseq, annot)
}
\description{
Logic to for sample comparison input
}
\keyword{internal}
