% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{sampleComparisonInput}
\alias{sampleComparisonInput}
\title{Selection form/button for sample comparisons (test vs control)}
\usage{
sampleComparisonInput(id)
}
\description{
Selection form/button for sample comparisons (test vs control)
}
\keyword{internal}
