% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{save_bulk_anals}
\alias{save_bulk_anals}
\title{Save new analysis info to anals dataframe}
\usage{
save_bulk_anals(dataset_name, dataset_dir, anal_name, data_dir)
}
\arguments{
\item{dataset_name}{Name of dataset}

\item{dataset_dir}{Folder name for dataset}

\item{anal_name}{Name of new analysis}

\item{data_dir}{Path to folder container \code{'bulk'} and \code{'single-cell'} directories}
}
\description{
Save new analysis info to anals dataframe
}
\keyword{internal}
