% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_combined_reports.R
\name{save_combined_scseq_reports}
\alias{save_combined_scseq_reports}
\title{Save PDF of combined scRNA-seq reports}
\usage{
save_combined_scseq_reports(scseq, markers, orig_clusters, fname,
  point_size = 3)
}
\arguments{
\item{scseq}{\code{Seurat} object with \code{scseq$orig_clusters} providing the originally identified clusters for
each sample. This permits visualizing how integration shift the TSNE coordinates and cluster assignments.}

\item{markers}{Named list of character vectors specifying the marker genes to plot for each cluster. The names are the
clusters that reports will be generated for in order.}

\item{fname}{String giving the name of the PDF file to save.}
}
\value{
Saves report to \code{fname}.
}
\description{
For integrated data with multiple samples.
}
