% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{save_scle}
\alias{save_scle}
\title{Save SingleCellExperiment as loom file}
\usage{
save_scle(scseq, dataset_dir, overwrite = TRUE)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}

\item{dataset_dir}{Directory to save in}

\item{overwrite}{Overwrite existing? Default is \code{TRUE}}
}
\description{
Save SingleCellExperiment as loom file
}
\keyword{internal}
