% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{save_scseq_args}
\alias{save_scseq_args}
\title{Save arguments for integration/subsetting}
\usage{
save_scseq_args(args, dataset_name, sc_dir)
}
\arguments{
\item{args}{Arguments to save}

\item{dataset_name}{Name of analysis}

\item{sc_dir}{Directory with single-cell analyses}
}
\description{
Save arguments for integration/subsetting
}
