% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{save_scseq_reports}
\alias{save_scseq_reports}
\title{Save PDF of scRNA-seq reports}
\usage{
save_scseq_reports(sce, markers, fname, point_size = 3)
}
\arguments{
\item{point_size}{}
}
\description{
Save PDF of scRNA-seq reports
}
