% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{sc_dl_filename}
\alias{sc_dl_filename}
\title{Utility to generate filename for single cell download csv}
\usage{
sc_dl_filename(cluster, anal, comparison_type)
}
\arguments{
\item{cluster}{Character vector of cluster names}

\item{anal}{result of diff_expr_scseq}

\item{comparison_type}{either 'samples' or 'clusters'}
}
\value{
File name string
}
\description{
Utility to generate filename for single cell download csv
}
\keyword{internal}
