% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{scalign_integrate}
\alias{scalign_integrate}
\title{Integrate Single Cell datasets using scAlign}
\usage{
scalign_integrate(scseqs, genes)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects}

\item{genes}{Highly variable genes to integrate with}
}
\description{
Integrate Single Cell datasets using scAlign
}
\keyword{internal}
