% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{scalign_to_scseq}
\alias{scalign_to_scseq}
\title{Convert scAlign integrated dataset into Seurat object}
\usage{
scalign_to_scseq(scalign)
}
\arguments{
\item{scalign}{SingleCellExperiment returned from \code{scalign_integrate}}
}
\value{
Seurat object
}
\description{
Convert scAlign integrated dataset into Seurat object
}
\keyword{internal}
