% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scAnal}
\alias{scAnal}
\title{Logic for single cell cluster analyses for Single Cell, Drugs, and Pathways tabs}
\usage{
scAnal(input, output, session, dataset_dir, is_sc = function() TRUE)
}
\description{
Logic for single cell cluster analyses for Single Cell, Drugs, and Pathways tabs
}
\keyword{internal}
