% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scClusterPlot}
\alias{scClusterPlot}
\title{Logic for cluster plots}
\usage{
scClusterPlot(input, output, session, scseq, plot_styles, selected_group,
  cached_plot)
}
\description{
Logic for cluster plots
}
\keyword{internal}
