% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scClusterPlotOutput}
\alias{scClusterPlotOutput}
\title{Output plot of single cell clusters}
\usage{
scClusterPlotOutput(id)

scClusterPlotOutput(id)
}
\description{
Output plot of single cell clusters

Output plot of single cell clusters
}
\keyword{internal}
