% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scForm}
\alias{scForm}
\title{Logic for form on Single Cell Exploration page}
\usage{
scForm(input, output, session, sc_dir, indices_dir)
}
\description{
Logic for form on Single Cell Exploration page
}
\keyword{internal}
