% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scLabelsComparison}
\alias{scLabelsComparison}
\title{Logic for selecting cluster to plot label origin for integrated dataset}
\usage{
scLabelsComparison(input, output, session, cluster_choices)
}
\description{
Logic for selecting cluster to plot label origin for integrated dataset
}
\keyword{internal}
