% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scLabelsPlot}
\alias{scLabelsPlot}
\title{Logic for labels plot for integrated dataset}
\usage{
scLabelsPlot(input, output, session, sc_dir, selected_cluster, scseq)
}
\description{
Logic for labels plot for integrated dataset
}
\keyword{internal}
