% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scLabelsPlotOutput}
\alias{scLabelsPlotOutput}
\title{Output plotly for labels comparison with integrated datasets}
\usage{
scLabelsPlotOutput(id)
}
\description{
Output plotly for labels comparison with integrated datasets
}
\keyword{internal}
