% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scPage}
\alias{scPage}
\title{Logic for Single Cell Tab}
\usage{
scPage(
  input,
  output,
  session,
  sc_dir,
  indices_dir,
  tx2gene_dir,
  gs_dir,
  is_mobile,
  add_sc,
  remove_sc,
  integrate_sc,
  export_sc
)
}
\arguments{
\item{input}{standard shiny module boilerplate}

\item{output}{standard shiny module boilerplate}

\item{session}{standard shiny module boilerplate}

\item{sc_dir}{sub folder of \code{data_dir} where single-cell data is stored}

\item{indices_dir}{Path to directory containing \code{kallisto} indices and whitelists.}

\item{tx2gene_dir}{Path to directory containing transcript to gene maps
produced by \link[dseqr.data]{load_tx2gene}.}

\item{gs_dir}{Path to directory where gene to Gene Ontology maps produced by \link{get_genego} are saved.}

\item{is_mobile}{is the page being shown on a mobile device? If \code{TRUE},
then various styles are updated accordingly.}

\item{add_sc}{reactive that is used to trigger the modal to import single-cell datasets.}

\item{remove_sc}{reactive that is used to trigger the modal to delete single-cell datasets.}

\item{integrate_sc}{reactive that is used to trigger the modal to integrate single-cell datasets.}

\item{export_sc}{reactive that is used to trigger the modal to export single-cell datasets.}
}
\value{
Called with \link[shiny]{callModule} to generate logic for
  single-cell tab.
}
\description{
Logic for Single Cell Tab
}
