% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scPageUI}
\alias{scPageUI}
\title{UI for Single Cell Exploration page}
\usage{
scPageUI(id, tab, active)
}
\arguments{
\item{id}{Identification string that is names-paced using \link[shiny]{NS}.}

\item{tab}{Name to appear on tab}

\item{active}{Name of current active \code{tab}}
}
\value{
shiny.tag with html for single-cell tab
}
\description{
UI for Single Cell Exploration page
}
