% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-server.R
\name{scPathClusters}
\alias{scPathClusters}
\title{Logic for single cell clusters selector for pathForm}
\usage{
scPathClusters(input, output, session, data_dir, anal, is_sc)
}
\description{
Logic for single cell clusters selector for pathForm
}
\keyword{internal}
