% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-ui.R
\name{scPathClustersInput}
\alias{scPathClustersInput}
\title{Input for single cell clusters in pathFormInput}
\usage{
scPathClustersInput(ns)
}
\description{
Input for single cell clusters in pathFormInput
}
\keyword{internal}
