% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scRidgePlot}
\alias{scRidgePlot}
\title{Logic for Ridge plot for clusters}
\usage{
scRidgePlot(input, output, session, selected_gene, selected_cluster, scseq)
}
\description{
Logic for Ridge plot for clusters
}
\keyword{internal}
