% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSampleClusters}
\alias{scSampleClusters}
\title{Logic for single cell sample comparison cluster for Single Cell and Drugs tabs}
\usage{
scSampleClusters(
  input,
  output,
  session,
  scseq,
  lm_fit,
  groups,
  dataset_dir,
  resoln_dir,
  plots_dir,
  feature_plot,
  dataset_name,
  sc_dir,
  input_annot = function() NULL,
  show_dprimes = function() TRUE,
  is_integrated = function() TRUE,
  is_sc = function() TRUE,
  exclude_ambient = function() FALSE,
  comparison_type = function() "samples",
  applied = function() TRUE,
  is_mobile = function() FALSE
)
}
\description{
Logic for single cell sample comparison cluster for Single Cell and Drugs tabs
}
\keyword{internal}
