% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R, R/modules-path-ui.R
\name{scSampleComparisonInput}
\alias{scSampleComparisonInput}
\title{Input for Single Cell sample comparison}
\usage{
scSampleComparisonInput(ns)

scSampleComparisonInput(ns)
}
\description{
Used in both Drugs and Pathways tab
}
\keyword{internal}
