% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSampleGroups}
\alias{scSampleGroups}
\title{Logic for single cell sample comparison groups for Single Cell and Drugs tabs}
\usage{
scSampleGroups(
  input,
  output,
  session,
  dataset_dir,
  resoln_dir,
  dataset_name,
  input_scseq = function() NULL
)
}
\description{
Logic for single cell sample comparison groups for Single Cell and Drugs tabs
}
\keyword{internal}
