% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSampleMarkerPlot}
\alias{scSampleMarkerPlot}
\title{Logic for single-cell sample comparison plots}
\usage{
scSampleMarkerPlot(input, output, session, selected_gene, plot_fun)
}
\description{
setup to allow for ggplot/plotly
}
\keyword{internal}
