% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scSampleMarkerPlotOutput}
\alias{scSampleMarkerPlotOutput}
\title{Output plot/plotly for samples comparison with integrated datasets}
\usage{
scSampleMarkerPlotOutput(id)
}
\description{
Output plot/plotly for samples comparison with integrated datasets
}
\keyword{internal}
