% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSelectedAnal}
\alias{scSelectedAnal}
\title{Logic for selected analysis part of scForm}
\usage{
scSelectedAnal(input, output, session, sc_dir, new_anal, new_dataset)
}
\description{
Logic for selected analysis part of scForm
}
\keyword{internal}
