% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scSelectedAnalInput}
\alias{scSelectedAnalInput}
\title{Input form/associated buttons for selecting single cell analysis}
\usage{
scSelectedAnalInput(id)
}
\description{
Input form/associated buttons for selecting single cell analysis
}
\keyword{internal}
