% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSelectedDataset}
\alias{scSelectedDataset}
\title{Logic for selected dataset part of scForm}
\usage{
scSelectedDataset(input, output, session, sc_dir, new_dataset)
}
\description{
Logic for selected dataset part of scForm
}
\keyword{internal}
