% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scSelectedDatasetInput}
\alias{scSelectedDatasetInput}
\title{Input form/associated buttons for selecting single cell dataset}
\usage{
scSelectedDatasetInput(id)
}
\description{
Input form/associated buttons for selecting single cell dataset
}
\keyword{internal}
