% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{scseq_part_path}
\alias{scseq_part_path}
\title{Get path to saved scseq part}
\usage{
scseq_part_path(data_dir, anal_name, part)
}
\arguments{
\item{data_dir}{Path to directory with analyses.}

\item{anal_name}{Name of analysis.}

\item{part}{either \code{'annot'}, \code{'scseq'}, or \code{'markers'}.}
}
\value{
Path to analysis \code{part}.
}
\description{
Get path to saved scseq part
}
\keyword{internal}
