% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_contrast.R
\name{select_contrast}
\alias{select_contrast}
\title{Select contrast for an experiment.}
\usage{
select_contrast(eset)
}
\arguments{
\item{eset}{\code{ExpressionSet}}
}
\value{
Named list of length two. In order, list items contain rows corresponding to control and test group. Names are specified group names.
}
\description{
Used to select a contrast (control and test group samples) for an experiment.
}
\examples{

eset_path <- system.file('extdata', 'IBD', 'eset.rds', package='drugseqr')
eset <- readRDS(eset_path)

select_contrast(eset)


}
