% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pairs.R
\name{select_pairs}
\alias{select_pairs}
\title{Select pairs and replicates for an experiment.}
\usage{
select_pairs(data_dir)
}
\arguments{
\item{data_dir}{Directory with raw fastq.gz RNA-Seq files.}
}
\value{
data.frame with character column \code{'File Name'} and integer columns \code{'Pair'} (if pair-ended) and \code{'Replicate'} (if added).
}
\description{
Used to select a pairs for paired-end experiments and replicate samples. Please follow prompt, ensuring correct file name matching and
end-type experiment identification.
}
\examples{

data_dir <- system.file('extdata', 'IBD', package='drugseqr', mustWork = TRUE)
pdata <- select_pairs(data_dir)

}
