% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R, R/modules-sc-server.R
\name{selectedAnal}
\alias{selectedAnal}
\title{Logic for selected dataset/analysis in Drugs and Pathways tabs}
\usage{
selectedAnal(input, output, session, sc_dir, new_anal, new_dataset)

selectedAnal(input, output, session, sc_dir, new_anal, new_dataset)
}
\description{
Logic for selected dataset/analysis in Drugs and Pathways tabs

Logic for selected analysis part of scForm
}
\keyword{internal}
