% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{selectedAnalInput}
\alias{selectedAnalInput}
\title{Analysis input for Drugs and Pathways page}
\usage{
selectedAnalInput(id)
}
\description{
Analysis input for Drugs and Pathways page
}
\keyword{internal}
