% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{selectedAnnot}
\alias{selectedAnnot}
\title{Logic for selected anals in integrated dataset to show original annotation plots for.}
\usage{
selectedAnnot(input, output, session, scseq, is.integrated, sc_dir)
}
\description{
Logic for selected anals in integrated dataset to show original annotation plots for.
}
\keyword{internal}
