% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectedAnnotDatasetInput}
\alias{selectedAnnotDatasetInput}
\title{Input for selecting datasets to show original labels for}
\usage{
selectedAnnotDatasetInput(id)
}
\description{
Input for selecting datasets to show original labels for
}
\keyword{internal}
