% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectedClusterInput}
\alias{selectedClusterInput}
\title{Input form and buttons to select a cluster or contrast and rename a cluster}
\usage{
selectedClusterInput(id)
}
\description{
Input form and buttons to select a cluster or contrast and rename a cluster
}
