% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{selectedDrugStudy}
\alias{selectedDrugStudy}
\title{Logic for selected drug study}
\usage{
selectedDrugStudy(input, output, session)
}
\description{
Logic for selected drug study
}
\keyword{internal}
