% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{selectedDrugStudyInput}
\alias{selectedDrugStudyInput}
\title{Select drugs study (CMAP or L1000) for drugs page}
\usage{
selectedDrugStudyInput(id)
}
\description{
Select drugs study (CMAP or L1000) for drugs page
}
\keyword{internal}
