% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{selectedGene}
\alias{selectedGene}
\title{Logic for selected gene to show plots for}
\usage{
selectedGene(
  input,
  output,
  session,
  dataset_name,
  dataset_dir,
  scseq,
  is_integrated,
  selected_markers,
  selected_cluster,
  type,
  qc_metrics = function() NULL,
  ambient = function() NULL
)
}
\description{
Logic for selected gene to show plots for
}
\keyword{internal}
