% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectedGeneInput}
\alias{selectedGeneInput}
\title{Input form to select gene for scBioGpsPlotOutput and scMarkerPlotOutput}
\usage{
selectedGeneInput(id)
}
\description{
Input form to select gene for scBioGpsPlotOutput and scMarkerPlotOutput
}
