% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectedGroupsInput}
\alias{selectedGroupsInput}
\title{Input form to control/test/all groups for integrated datasets}
\usage{
selectedGroupsInput(id)
}
\description{
Input form to control/test/all groups for integrated datasets
}
\keyword{internal}
