% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{selectedPertSignature}
\alias{selectedPertSignature}
\title{Logic for showing pert selection for gene plot}
\usage{
selectedPertSignature(
  input,
  output,
  session,
  data_dir,
  query_res,
  query_type,
  pert_signature_dir,
  show_genes
)
}
\description{
Logic for showing pert selection for gene plot
}
\keyword{internal}
