% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{selectedPertSignatureInput}
\alias{selectedPertSignatureInput}
\title{UI for seperate drugsPertInput for CMAP/L1000}
\usage{
selectedPertSignatureInput(id)
}
\description{
UI for seperate drugsPertInput for CMAP/L1000
}
\keyword{internal}
