% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectizeInputMultWithButton}
\alias{selectizeInputMultWithButton}
\title{selectizeInput with button}
\usage{
selectizeInputMultWithButton(id, label, button)
}
\description{
Adjusts height is multiple.
}
\keyword{internal}
