% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{selectizeInputWithButtons}
\alias{selectizeInputWithButtons}
\title{selectizeInput with buttons and validation}
\usage{
selectizeInputWithButtons(id, label, ..., options = NULL,
  container_id = NULL, help_id = NULL)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{...}{selectizeInput}
}
\description{
selectizeInput with buttons and validation
}
\keyword{internal}
