% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{selectizeInputWithValidation}
\alias{selectizeInputWithValidation}
\title{selectizeInput with validation}
\usage{
selectizeInputWithValidation(id, label, options = NULL,
  container_id = NULL, help_id = NULL)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{options}{A list of options. See the documentation of \pkg{selectize.js}
for possible options (character option values inside \code{\link[base]{I}()} will
be treated as literal JavaScript code; see \code{\link{renderDataTable}()}
for details).}

\item{container_id}{id of container. Used to toggle 'has-error' class with shinyjs::toggleClass.}

\item{help_id}{id of help block. Used to show help message in change shinyjs::html}
}
\description{
selectizeInput with validation
}
\keyword{internal}
