% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_error.R
\name{send_slack_error}
\alias{send_slack_error}
\title{Send shiny errors to slack}
\usage{
send_slack_error(project, user)
}
\arguments{
\item{project}{The project that the error originated from}

\item{user}{The user that the error originated from}
}
\value{
Generates an alert in browser
}
\description{
Send shiny errors to slack
}
