% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{setup_fdata}
\alias{setup_fdata}
\title{Setup feature annotation data}
\usage{
setup_fdata(tx2gene)
}
\arguments{
\item{tx2gene}{\code{data.frame} mapping transcripts to gene names. Returned from from \code{\link{get_tx2gene}}.}
}
\value{
\code{data.table} with columns \code{SYMBOL_9606} and \code{ENTREZID_HS} corresponding to
  HGNC symbols and human entrez ids respectively.
}
\description{
Setup feature annotation data
}
\keyword{internal}
