% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{shorten_cluster_labels}
\alias{shorten_cluster_labels}
\title{Shorted cluster labels when there are lots of clusters}
\usage{
shorten_cluster_labels(labels)
}
\arguments{
\item{labels}{Character vector of labels}
}
\value{
\code{labels} where each word is shortened. trailing underscore numbers are preserved.
}
\description{
Shorted cluster labels when there are lots of clusters
}
\keyword{internal}
