% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{showIntegration}
\alias{showIntegration}
\title{Logic for show integration button in selectedAnal}
\usage{
showIntegration(input, output, session)
}
\description{
Logic for show integration button in selectedAnal
}
\keyword{internal}
