% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{showIntegrationButton}
\alias{showIntegrationButton}
\title{Button with to toggle display of integrationFormInput}
\usage{
showIntegrationButton(id)
}
\description{
Button with to toggle display of integrationFormInput
}
\keyword{internal}
