% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{showLabelTransfer}
\alias{showLabelTransfer}
\title{Logic for show label transfer button}
\usage{
showLabelTransfer(input, output, session)
}
\description{
Logic for show label transfer button
}
\keyword{internal}
