% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{showLabelTransferButton}
\alias{showLabelTransferButton}
\title{Button with to toggle display of label transfer inputs}
\usage{
showLabelTransferButton(id)
}
\description{
Button with to toggle display of label transfer inputs
}
\keyword{internal}
