% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleExIPlot.R
\name{SingleExIPlot}
\alias{SingleExIPlot}
\title{Plot a single expression by identity on a plot}
\usage{
SingleExIPlot(
  data,
  idents,
  hl = NULL,
  title = NULL,
  split = NULL,
  type = "violin",
  sort = FALSE,
  y.max = NULL,
  adjust = 1,
  pt.size = 0.05,
  cols = NULL,
  seed.use = 42,
  log = FALSE,
  color = NULL,
  color_dark = NULL,
  nsel = 1,
  ncells = NULL
)
}
\arguments{
\item{data}{Data to plot}

\item{idents}{Idents to use}

\item{type}{Make either a 'ridge' or 'violin' plot}

\item{sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{y.max}{Maximum Y value to plot}

\item{adjust}{Adjust parameter for geom_violin}

\item{cols}{Colors to use for plotting}

\item{seed.use}{Random seed to use. If NULL, don't set a seed}

\item{log}{plot Y axis on log scale}
}
\value{
A ggplot-based Expression-by-Identity plot
}
\description{
Plot a single expression by identity on a plot
}
