% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleViolinIPlot.R
\name{SingleViolinIPlot}
\alias{SingleViolinIPlot}
\title{Violin plot by identity}
\usage{
SingleViolinIPlot(
  data,
  idents,
  hl = NULL,
  title = NULL,
  sort = FALSE,
  y.max = NULL,
  adjust = 1,
  pt.size = 1,
  pt.shape = 20,
  seed.use = 42,
  log = FALSE,
  color = NULL,
  color_dark = NULL,
  nsel = 1,
  ncells = NULL,
  pct.cells = NULL
)
}
\arguments{
\item{data}{Data to plot}

\item{idents}{Idents to use}

\item{hl}{Factor used to highlight specific \code{idents} via
fill, alpha, and color geoms.}

\item{title}{Plot title}

\item{sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{y.max}{Maximum Y value to plot}

\item{adjust}{Adjust parameter for geom_violin}

\item{pt.size}{size parameter for geom_jitter}

\item{pt.shape}{shape parameter for geom_jitter}

\item{seed.use}{Random seed to use. If NULL, don't set a seed}

\item{log}{plot Y axis on log scale}

\item{color}{Colors to use for fill with length equal to one less than
\code{length(levels(hl))}. The last level of \code{hl} is filled with gray.}

\item{color_dark}{Darker versions of \code{color} used for points and lines.}

\item{nsel}{Number of selected \code{idents}. Default is 1.}

\item{ncells}{Vector specifying number of cells in each {levels(idents)}.
If specified, an annotated bar plot is drawn to indicate number of cells.}

\item{pct.cells}{Vector specifying percent of cells that express the marker,
once for each {levels(idents)}. If specified, an annotated bar plot is drawn.}
}
\value{
A ggplot-based Violin-by-Identity plot
}
\description{
Violin plot by identity
}
