% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{sort_clusters}
\alias{sort_clusters}
\title{shortest distance between cluster centroids}
\usage{
sort_clusters(label_coords)
}
\arguments{
\item{label_coords}{data.frame with columns label, x, and y giving coordinates
of cluster labels}
}
\value{
\code{label_coords} sorted by shortest path
}
\description{
TODO: use to order clusters by similarity
}
\examples{

label_coords <- data.frame(
  label = c('A', 'B', 'C'),
  x = c(1, 5, 2),
  y = c(1, 5, 2))

sort_clusters(label_coords)

}
\keyword{internal}
