% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{sort_query_table_clin}
\alias{sort_query_table_clin}
\title{final sorting/filtering of query table}
\usage{
sort_query_table_clin(
  query_table_clin,
  sort_by,
  sort_abs,
  direction,
  drug_study,
  remove_html = FALSE
)
}
\arguments{
\item{query_table_clin}{result of \code{filter_clinical}}

\item{sort_by}{Metric to sort by. Either \code{'avg_cor'} (default) or \code{'min_cor'}.}

\item{sort_abs}{Should results be sorted based on absolute correlation? Default is \code{FALSE}.}

\item{direction}{Direction of correlation to sort results by. One of \code{'both'}, \code{'similar'}, or \code{'opposing'} (default).}

\item{drug_study}{Queried database giving rise to \code{query_res}.
One of \code{'CMAP02'},\code{'L1000 Genetic'}, or \code{'L1000 Drugs'}.}

\item{remove_html}{Should html columns be removed? For non-webapp usage.}
}
\value{
Final query table
}
\description{
final sorting/filtering of query table
}
\keyword{internal}
