% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{srt_to_sce}
\alias{srt_to_sce}
\title{Convert Seurat object to SingleCellExperiment}
\usage{
srt_to_sce(srt, assay = NULL)
}
\arguments{
\item{srt}{\code{Seurat} object}
}
\value{
\code{SingleCellExperiment} object
}
\description{
Also adds mrna and rrna
}
