% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{srt_to_sce_shim}
\alias{srt_to_sce_shim}
\title{Process Legacy Seurat Datasets with SCE}
\usage{
srt_to_sce_shim(srt, sc_dir, dataset_name)
}
\arguments{
\item{srt}{\code{Seurat} object}

\item{sc_dir}{Path to folder with single-cell datasets}

\item{dataset_name}{Name of dataset}
}
\value{
Throws error and asks to reset app
}
\description{
Remove this once all legacy datasets updated.
}
