% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{stabilize_scseq}
\alias{stabilize_scseq}
\title{Model and account for mean-variance relationship}
\usage{
stabilize_scseq(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}
}
\value{
Variance stabilized \code{sce}
}
\description{
Model and account for mean-variance relationship
}
