% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{standardize_cellranger}
\alias{standardize_cellranger}
\title{Rename CellRanger files for loading by Read10X}
\usage{
standardize_cellranger(data_dir)
}
\arguments{
\item{data_dir}{Path to folder with CellRanger files.}
}
\description{
Rename CellRanger files for loading by Read10X
}
\keyword{internal}
