% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_scseq.R
\name{strain_scseq}
\alias{strain_scseq}
\title{Remove ambient gene expression from sc-RNAseq counts}
\usage{
strain_scseq(counts, empty, project)
}
\arguments{
\item{counts}{sparse dgTMatrix returned by \code{\link{load_kallisto}}.}

\item{empty}{Boolean with indicating columns in \code{counts} that are empty droplets. Return by \code{\link{get_empty}}}

\item{project}{String. Passed to \code{CreateSeuratObject}.}
}
\value{
\code{Seurat} object with counts corrected for ambient contamination.
}
\description{
Uses \code{SoupX} to estimate and adjust counts for ambient contamination.
Currently only works with kallisto quantification results.
}
