% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_scseq.R
\name{subset_scseq}
\alias{subset_scseq}
\title{Remove cells above/below three mads from the median. Used by \code{qc_scseq}.}
\usage{
subset_scseq(scseq, by)
}
\arguments{
\item{scseq}{Seurat object}

\item{by}{String, a numeric column in \code{scseq@meta.data} to subset by.}
}
\value{
\code{scseq} with cells that are within three mads of the median of \code{by}.
}
\description{
Remove cells above/below three mads from the median. Used by \code{qc_scseq}.
}
\keyword{internal}
