% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{subsetForm}
\alias{subsetForm}
\title{Logic for subsetting a datatset}
\usage{
subsetForm(
  input,
  output,
  session,
  sc_dir,
  scseq,
  datasets,
  show_subset,
  selected_dataset,
  dataset_dir,
  cluster_choices
)
}
\description{
Logic for subsetting a datatset
}
\keyword{internal}
