% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{subsetFormInput}
\alias{subsetFormInput}
\title{Input form for subsetting single cell datasets}
\usage{
subsetFormInput(id)
}
\description{
Input form for subsetting single cell datasets
}
\keyword{internal}
