% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_table.R
\name{summarise_query_table}
\alias{summarise_query_table}
\title{Summarise annotated query table}
\usage{
summarise_query_table(
  query_table_annot,
  is_genetic,
  cells,
  sort_abs,
  remove_html = FALSE,
  ntop = 1500
)
}
\arguments{
\item{query_table_annot}{result of \link{query_table_annot}}

\item{is_genetic}{Boolean indicating if the query results are from L1000 genetic perturbations.}

\item{cells}{Character vector of cell types to include. Default (\code{NULL}) includes all cell types.}

\item{sort_abs}{Should results be sorted based on absolute correlation? Default is \code{FALSE}.}

\item{remove_html}{Should html columns be removed? For non-webapp usage.}

\item{ntop}{Integer indicating the number of rows to keep after sorting by \code{arrange_by}.}
}
\value{
summarised query table
}
\description{
Summarise annotated query table
}
\keyword{internal}
