% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{supress.genes}
\alias{supress.genes}
\title{Move genes to bottom of markers data.frame}
\usage{
supress.genes(markers, supress)
}
\arguments{
\item{markers}{data.frame with genes as row names.}

\item{supress}{Genes to move to bottom of \code{markers}.}
}
\value{
\code{markers} with genes in \code{supress} at bottom.
}
\description{
Used to supress ambient outliers and positive genes for single-cell sample comparisons
}
\keyword{internal}
