% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-ui.R
\name{svaButton}
\alias{svaButton}
\title{Button with sliders for adjusting number of surrogate variables}
\usage{
svaButton(inputId, sliderId, max_svs = 0, prev_svs = 0)
}
\description{
Button with sliders for adjusting number of surrogate variables
}
\keyword{internal}
