% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{tabPane}
\alias{tabPane}
\title{UI for a tab pane}
\usage{
tabPane(tab, active, ...)
}
\arguments{
\item{tab}{The name of the tab}

\item{active}{The name of the active tab}

\item{...}{The UI elements to place in the tab}
}
\value{
shiny div tag with UI for tab
}
\description{
UI for a tab pane
}
\keyword{internal}
