% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{textAreaInputWithButtons}
\alias{textAreaInputWithButtons}
\title{textAreaInput with buttons}
\usage{
textAreaInputWithButtons(
  id,
  label,
  ...,
  container_id = NULL,
  help_id = NULL,
  tooltip = TRUE,
  placeholder = ""
)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{...}{actionButtons}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\description{
textAreaInput with buttons
}
\keyword{internal}
