% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{textInputWithButtons}
\alias{textInputWithButtons}
\title{textInput with buttons}
\usage{
textInputWithButtons(id, label, ..., container_id = NULL,
  help_id = NULL)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{...}{actionButtons}
}
\description{
textInput with buttons
}
\keyword{internal}
