% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{textInputWithValidation}
\alias{textInputWithValidation}
\title{textInput with validation}
\usage{
textInputWithValidation(id, label, container_id = NULL, help_id = NULL)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{container_id}{id of container. Used to toggle 'has-error' class with shinyjs::toggleClass.}

\item{help_id}{id of help block. Used to show help message in change shinyjs::html}
}
\description{
textInput with validation
}
\keyword{internal}
