% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_reports.R
\name{theme_no_xaxis}
\alias{theme_no_xaxis}
\title{Remove ggplot xaxis title, text, and ticks}
\usage{
theme_no_xaxis()
}
\value{
ggplot2 \code{theme} object
}
\description{
Remove ggplot xaxis title, text, and ticks
}
\examples{

library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_no_xaxis()

}
