% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{theme_no_yaxis}
\alias{theme_no_yaxis}
\title{Remove ggplot yaxis title, text, and ticks}
\usage{
theme_no_yaxis()
}
\value{
\code{theme}
}
\description{
Remove ggplot yaxis title, text, and ticks
}
\keyword{internal}
