% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{toggleAll}
\alias{toggleAll}
\title{Toggle disabled state for multiple ids}
\usage{
toggleAll(ids)
}
\arguments{
\item{ids}{Character vector of ids to disable}
}
\description{
Toggle disabled state for multiple ids
}
\keyword{internal}
