% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{toggleSelectizeButtons}
\alias{toggleSelectizeButtons}
\title{Show/hide buttons in selectizeInputWithButtons}
\usage{
toggleSelectizeButtons(selectize_id, button_ids, condition)
}
\arguments{
\item{selectize_id}{id of selectizeInputWithButtons element}

\item{button_ids}{character vector of ids for buttons in selectizeInputWithButtons}

\item{condition}{will show if TRUE and hide if FALSE}
}
\description{
Show/hide buttons in selectizeInputWithButtons
}
