% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{transfer_labels}
\alias{transfer_labels}
\title{Get predicted cells types for a query dataset based on a reference dataset}
\usage{
transfer_labels(reference, query, updateProgress = NULL, n = 3,
  n_init = 2)
}
\arguments{
\item{reference}{\code{Seurat} object to use as a reference}

\item{query}{\code{Seurat} object to obtain predictions for}
}
\value{
\code{data.frame} with predictions
}
\description{
Get predicted cells types for a query dataset based on a reference dataset
}
