% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{transfer_prev_annot}
\alias{transfer_prev_annot}
\title{Transfer annotation when change resolution}
\usage{
transfer_prev_annot(resoln, prev_resoln, dataset_name, sc_dir)
}
\arguments{
\item{resoln}{new resolution}

\item{prev_resoln}{previous resolution}

\item{dataset_name}{name of single-cell dataset}

\item{sc_dir}{directory with folder named \code{dataset_name}}
}
\value{
saves new annotation in resolution subdirectory
}
\description{
Transfer annotation when change resolution
}
\keyword{internal}
