% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{transition_efs}
\alias{transition_efs}
\title{Utility to move files on EFS from IA to SA}
\usage{
transition_efs(fpath)
}
\arguments{
\item{fpath}{Path of file}
}
\description{
If time since last access time is greater than \code{Sys.getenv('EFS_LIFECYCLE')},
\code{fpath} to copied in order to move it out of EFS infrequent access.
Ignored if EFS_LIFECYCLE not set.
}
