% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_whitelist.R
\name{tsne_scseq_whitelist}
\alias{tsne_scseq_whitelist}
\title{tSNE plots of whitelisted cells}
\usage{
tsne_scseq_whitelist(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} returned by \code{load_scseq}}
}
\value{
\code{ggplot}
}
\description{
tSNE plots of whitelisted cells
}
