% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{update_feature_plot}
\alias{update_feature_plot}
\title{Update Feature Plot}
\usage{
update_feature_plot(
  plot,
  feature_data,
  feature,
  reverse_scale = feature \%in\% c("ribo_percent", "log10_sum", "log10_detected"),
  is_qc = feature \%in\% c("ribo_percent", "mito_percent", "log10_sum",
    "log10_detected", "doublet_score")
)
}
\arguments{
\item{plot}{ggplot2 object}

\item{feature_data}{Named numeric vector of feature values. Names are cell
names and must be a subset of \code{row.names(plot$data)}}

\item{feature}{Name of feature}

\item{reverse_scale}{Should low values be emphasized? Used for features where
low values indicate possible quality issues.}
}
\value{
ggplot2 object
}
\description{
Used to change feature with fast re-plot
}
\keyword{internal}
