% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{validate_custom_query}
\alias{validate_custom_query}
\title{Validate custom query}
\usage{
validate_custom_query(dn_genes, up_genes, custom_name)
}
\arguments{
\item{dn_genes}{Character vector of genes to upregulate}

\item{up_genes}{Character vector of genes to downregulate}

\item{custom_name}{Name for new custom analysis}
}
\value{
\code{NULL} if valid, otherwise an error message
}
\description{
Validate custom query
}
\keyword{internal}
