% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{validate_integration}
\alias{validate_integration}
\title{Validate dataset selection for integration}
\usage{
validate_integration(types, name, azimuth_ref, dataset_names, sc_dir)
}
\value{
\code{NULL} if valid, otherwise an error message
}
\description{
Validate dataset selection for integration
}
\keyword{internal}
