% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{validate_metric}
\alias{validate_metric}
\title{Validate a custom metric}
\usage{
validate_metric(metric, scseq)
}
\arguments{
\item{metric}{String with custom metric to evaluate.}

\item{scseq}{\code{DietSCE} to evaluate \code{metric} for.}
}
\value{
Either a data.frame is successful or a object with class 'try-error'.
}
\description{
Validate a custom metric
}
\keyword{internal}
