% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pairs.R
\name{validate_pairs}
\alias{validate_pairs}
\title{Validate sample pairing for pair-ended RNA seq}
\usage{
validate_pairs(pairs, rows, reps)
}
\arguments{
\item{pairs}{Numeric vector of integers and/or \code{NA}. Positions with the same integer
value indicate samples that are paired in a pair-ended experiment.}

\item{rows}{Numeric vector of integers indicating selected rows.}

\item{reps}{Numeric vector of integers and/or \code{NA}. Positions with the same integer
value indicate samples that replicates.}
}
\value{
\code{TRUE} if the pairing is valid, otherwise \code{FALSE}.
}
\description{
Validate sample pairing for pair-ended RNA seq
}
\keyword{internal}
