% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-utils.R
\name{validate_pdata}
\alias{validate_pdata}
\title{Validate pdata before differential expression analysis}
\usage{
validate_pdata(pdata)
}
\arguments{
\item{pdata}{data.frame with column \code{'Group'}}
}
\value{
NULL if valid, otherwise a character vector indicating what's wrong
}
\description{
Validate pdata before differential expression analysis
}
\keyword{internal}
