% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{validate_preds}
\alias{validate_preds}
\title{Check that label predictions aren't from overwritten datasets}
\usage{
validate_preds(preds, sc_dir)
}
\arguments{
\item{preds}{List of predictions}

\item{sc_dir}{Directory with single-cell datasets}
}
\value{
\code{preds} that were generated from current datasets
}
\description{
Check that label predictions aren't from overwritten datasets
}
\keyword{internal}
