% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pairs.R
\name{validate_reps}
\alias{validate_reps}
\title{Validate sample replicates for RNA seq}
\usage{
validate_reps(pairs, rows, reps)
}
\arguments{
\item{pairs}{Numeric vector of integers and/or \code{NA}. Positions with the same integer
value indicate samples that are paired in a pair-ended experiment.}

\item{rows}{Numeric vector of integers indicating selected rows.}

\item{reps}{Numeric vector of integers and/or \code{NA}. Positions with the same integer
value indicate samples that replicates.}
}
\value{
\code{TRUE} if the replicate is valid, otherwise \code{FALSE}.
}
\description{
Validate sample replicates for RNA seq
}
\keyword{internal}
