% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{validate_up_custom}
\alias{validate_up_custom}
\title{Validate uploaded custom query}
\usage{
validate_up_custom(top_table, custom_name)
}
\arguments{
\item{top_table}{data.frame with logFC, dprime, or effect sizes as column 1
and row.names as HGNC symbols}

\item{custom_name}{Name for new custom analysis}
}
\value{
\code{NULL} if valid, otherwise an error message
}
\description{
Validate uploaded custom query
}
\keyword{internal}
