% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{add_combined_metrics}
\alias{add_combined_metrics}
\title{Add QC metrics to integrated SingleCellExperiment}
\usage{
add_combined_metrics(combined, scseqs)
}
\arguments{
\item{combined}{Integrated \code{SingleCellExperiment}}

\item{scseqs}{list of \code{SingleCellExperiment}s used to create \code{combined}}
}
\value{
\code{combined} with QC metrics from \code{scseqs} added
}
\description{
Add QC metrics to integrated SingleCellExperiment
}
\keyword{internal}
