% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_hvgs}
\alias{add_hvgs}
\title{Get top highly variable genes for subsequent dimensionality reduction}
\usage{
add_hvgs(sce, hvgs = NULL)
}
\arguments{
\item{sce}{\code{DietSCE} object}
}
\description{
Runs after \code{preprocess_scseq}
}
