% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{add_hvgs}
\alias{add_hvgs}
\title{Add top highly variable genes for subsequent dimensionality reduction}
\usage{
add_hvgs(sce, hvgs = NULL)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object}

\item{hvgs}{Character vector of genes to use. Used to specify external
set of highly variable genes.}
}
\value{
\code{sce} a boolean vector added to \code{rowData(sce)$hvg} that
indicates which rows correspond to highly variable genes.
}
\description{
Runs after \code{preprocess_scseq}
}
\keyword{internal}
