% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_scseq_clusters}
\alias{add_scseq_clusters}
\title{Add SNN Graph based clusters to SingleCellExperiment}
\usage{
add_scseq_clusters(scseq, use.dimred = "PCA")
}
\arguments{
\item{sce}{\code{SingleCellExperiment} returned by \code{\link{stabilize_scseq}} (contains \code{sce@reducedDims$PCA})}
}
\value{
\code{sce} with column \code{cluster} in \code{colData(sce)}
}
\description{
Add SNN Graph based clusters to SingleCellExperiment
}
