% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_scseq_clusters}
\alias{add_scseq_clusters}
\title{Cluster SingleCellExperiment}
\usage{
add_scseq_clusters(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}
}
\value{
\code{sce} with column \code{cluster} in colData and \code{'npcs'} in metadata
}
\description{
Cluster SingleCellExperiment
}
