% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{add_scseq_clusters}
\alias{add_scseq_clusters}
\title{Add clusters to single cell RNA-seq object}
\usage{
add_scseq_clusters(sce)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}
}
\value{
column \code{cluster} in \code{colData(sce)} is added.
}
\description{
Runs after \code{reduce_dims} so that \code{'PCA'} reducedDim is available.
}
