% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_results.R
\name{add_table_html}
\alias{add_table_html}
\title{Add HTML to query results table}
\usage{
add_table_html(query_res)
}
\arguments{
\item{query_res}{\code{data.frame} returned by \code{\link{summarize_compound}}}
}
\value{
\code{query_res} with pubchem cid links and correlation plot HTML.
}
\description{
Add HTML to query results table
}
