% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{build_ensdb}
\alias{build_ensdb}
\title{Build ensembldb annotation package.}
\usage{
build_ensdb(species = "Homo sapiens", release = "94")
}
\arguments{
\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link[drugseqr.data]{build_kallisto_index}}.}
}
\description{
Build ensembldb annotation package.
}
\examples{

# build ensembldb annotation package for human
build_ensdb()

}
