% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{build_ensdb_index}
\alias{build_ensdb_index}
\title{Download ensembl transcriptome and build index for salmon quantification}
\usage{
build_ensdb_index(species = "homo_sapiens", release = "94")
}
\arguments{
\item{species}{The species. Default is \code{homo_sapiens.}}

\item{release}{ensembl release. Default is \code{94} (latest in release for AnnotationHub -
needs to match with \code{\link{build_ensdb}}) and corresponds to Gencode release 29 for \code{\link{build_gencode_index}}.}
}
\description{
This index is used for bulk RNA seq quantification. See \code{\link{build_gencode_index}} for single cell RNA-seq equivalent.
}
\examples{
# build salmon index for humans
build_ensdb_index()

}
