% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{build_index}
\alias{build_index}
\title{Download ensembl transcriptome and build index for salmon quantification}
\usage{
build_index(species = "homo_sapiens", version = "94")
}
\arguments{
\item{species}{The species. Default is \code{homo_sapiens.}}

\item{version}{ensembl version. Default is \code{94} (latest in release for AnnotationHub - needs to match with \code{\link{build_ensdb}}).}
}
\description{
Download ensembl transcriptome and build index for salmon quantification
}
\examples{
# build salmon index for humans
build_index()

}
