% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{bulkAnal}
\alias{bulkAnal}
\title{Logic for bulk group analyses for Bulk, Drugs, and Pathways tabs}
\usage{
bulkAnal(
  input,
  output,
  session,
  pdata,
  dataset_name,
  eset,
  numsv,
  svobj,
  fastq_dir,
  is_bulk = function() TRUE
)
}
\description{
Logic for bulk group analyses for Bulk, Drugs, and Pathways tabs
}
\keyword{internal}
