% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{bulkAnal}
\alias{bulkAnal}
\title{Logic for bulk data analyses}
\usage{
bulkAnal(
  input,
  output,
  session,
  pdata,
  eset,
  numsv,
  svobj,
  fastq_dir,
  dl_content,
  dl_fname,
  enable_dl
)
}
\description{
Logic for bulk data analyses
}
\keyword{internal}
