% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{bulkAnal}
\alias{bulkAnal}
\title{Logic for bulk data analyses}
\usage{
bulkAnal(
  input,
  output,
  session,
  pdata,
  dataset_name,
  eset,
  numsv,
  svobj,
  fastq_dir,
  is_bulk = function() TRUE
)
}
\value{
list with name, lm_fit, and is_lmfit
}
\description{
Logic for bulk data analyses
}
\keyword{internal}
