% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-ui.R
\name{bulkAnalInput}
\alias{bulkAnalInput}
\title{Bulk Differential expression analysis input}
\usage{
bulkAnalInput(id, with_dl = TRUE, label = "Select groups to compare:")
}
\description{
Bulk Differential expression analysis input
}
\keyword{internal}
