% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkFormQuant}
\alias{bulkFormQuant}
\title{Logic for dataset quantification part of bulkForm}
\usage{
bulkFormQuant(
  input,
  output,
  session,
  error_msg,
  dataset_name,
  pdata,
  fastq_dir,
  data_dir,
  indices_dir
)
}
\description{
Logic for dataset quantification part of bulkForm
}
\keyword{internal}
