% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{bulkMDSplotly}
\alias{bulkMDSplotly}
\title{Logic for Bulk Data MDS plotly}
\usage{
bulkMDSplotly(input, output, session, dataset_name, explore_eset, numsv)
}
\description{
Logic for Bulk Data MDS plotly
}
\keyword{internal}
