% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{comparisonType}
\alias{comparisonType}
\title{Logic for comparison type toggle for integrated datasets}
\usage{
comparisonType(input, output, session, scseq, is_integrated)
}
\description{
Logic for comparison type toggle for integrated datasets
}
\keyword{internal}
