% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{construct_path_df}
\alias{construct_path_df}
\title{Used by get_all_df and get_path_df to construct the return result}
\usage{
construct_path_df(top_table, nmax = min(nrow(top_table), 200))
}
\arguments{
\item{top_table}{Filtered result of \code{\link[limma]{topTable}}}

\item{nmax}{Maximum number of rows to keep. Default is all. Used by \code{construct_all_df}
to limit number of plotted genes.}
}
\description{
Used by get_all_df and get_path_df to construct the return result
}
\keyword{internal}
