% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{construct_pbulk_esets}
\alias{construct_pbulk_esets}
\title{Run pseudo bulk limma fit}
\usage{
construct_pbulk_esets(summed, species = "Homo sapiens")
}
\arguments{
\item{summed}{Pseudobulk \code{SingleCellExperiment}}
}
\value{
Normalized \code{ExpressionSet}
}
\description{
Run pseudo bulk limma fit
}
\keyword{internal}
