% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbulk_qcsets.R
\name{construct_pbulk_qcsets}
\alias{construct_pbulk_qcsets}
\title{Construct Pseudobulk QC expression sets}
\usage{
construct_pbulk_qcsets(pbulk_esets)
}
\arguments{
\item{pbulk_esets}{List of pseudobulk expression sets}
}
\value{
List of pseodobulk qc expression sets
}
\description{
Used for testing if significant changes in e.g. sum of mtRNA or ribosomal counts
}
\examples{
pbulk_esets <- readRDS("~/Documents/Batcave/zaklab/drugseqr/data-raw/patient_data/sjia/single-cell/sjia_pbmcs_lungdisease_vs_healthy_harmony/pbulk_esets.rds")
qcsets <- construct_pbulk_qcsets(pbulk_esets)
lm_fit <- run_limma_scseq(qcsets)
get_top_table(lm_fit[[1]], robust = FALSE)

}
