% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{detect_10x_chemistry}
\alias{detect_10x_chemistry}
\title{Auto Detect 10x Chemistry}
\usage{
detect_10x_chemistry(
  indices_dir,
  index_path,
  data_dir,
  bus_args,
  fqs,
  techs = c("10xv2", "10xv3")
)
}
\arguments{
\item{indices_dir}{Folder that containes whitelist for each \code{tech}.}

\item{index_path}{Path to kallisto index.}

\item{data_dir}{Path to folder with 10x fastq.gz files.}

\item{bus_args}{Additional arguments to kallisto bus.}

\item{fqs}{Vector of fastq file names.}

\item{techs}{10x chemistries to check. Passed to kallisto -x argument.}
}
\value{
one of \code{techs} corresponding to chemistry with most reads that agree with whitelist.
}
\description{
Creates samples of 10,000 reads per lane and then determins the number of reads
with barcodes in agreement with the chemistry whitelist.
}
