% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_qc.R
\name{detect_cells}
\alias{detect_cells}
\title{Determine non-empty droplets}
\usage{
detect_cells(counts, species = "Homo sapiens")
}
\arguments{
\item{counts}{dgCMatrix of counts}

\item{species}{Character vector with names of species. Supports Homo sapiens or Mus musculus.}
}
\value{
Indices of columns of counts corresponding to non-empty droplets
}
\description{
Determine non-empty droplets
}
\keyword{internal}
