% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{diff_abundance}
\alias{diff_abundance}
\title{Run differential abundance analysis}
\usage{
diff_abundance(scseq, annot)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}
}
\description{
Run differential abundance analysis
}
\keyword{internal}
