% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{diff_abundance}
\alias{diff_abundance}
\title{Run differential abundance analysis}
\usage{
diff_abundance(
  obj,
  annot = NULL,
  pairs = NULL,
  orig.ident = NULL,
  filter = TRUE
)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} or count matrix}

\item{orig.ident}{factor of group identities with length \code{ncol(obj)}}
}
\description{
Run differential abundance analysis
}
\keyword{internal}
