% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{diff_path_scseq}
\alias{diff_path_scseq}
\title{Run PADOG pathway analysis on a single cell RNA-Seq dataset}
\usage{
diff_path_scseq(scseq, prev_anal, ambient, data_dir, anal_name,
  clusters_name, NI = 1000)
}
\arguments{
\item{scseq}{\code{Seurat} object subsetted to selected clusters}

\item{prev_anal}{Result of call to \code{diff_expr_scseq}}

\item{data_dir}{Directory to folders with single cell analyses}

\item{anal_name}{Name of folder in \code{data_dir} to save results to}

\item{clusters_name}{String with sorted comma seperated integers of selected clusters returned from \code{collapse_sorted}.}
}
\value{
result of \code{\link[PADOG]{padog}}
}
\description{
Analysis is performed comparing test to control cells for all selected clusters.
}
\keyword{internal}
