% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{dl_pert_result}
\alias{dl_pert_result}
\title{Download CMAP02/L1000 pert query result from S3}
\usage{
dl_pert_result(res_path)
}
\arguments{
\item{res_path}{Path to download file to.}
}
\description{
Download CMAP02/L1000 pert query result from S3
}
\examples{
data_dir <- tempdir()
res_path <- file.path(data_dir, 'cmap_res_BRD-K45319408_PC3_5um_24h.rds')
dseqr:::dl_pert_result(res_path)

}
\keyword{internal}
