% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{dl_pert_signature}
\alias{dl_pert_signature}
\title{Download CMAP02/L1000 pert signature from S3}
\usage{
dl_pert_signature(sig_path, pert_type)
}
\arguments{
\item{sig_path}{Path to download file to.}

\item{pert_type}{One of \code{'cmap'}, \code{'l1000_drugs'}, or \code{'l1000_genes'}.}
}
\value{
\code{NULL}. Called for side effects.
}
\description{
Download CMAP02/L1000 pert signature from S3
}
\examples{
data_dir <- tempdir()
sig_path <- file.path(data_dir, '0317956-0000_PC3_1e-06M_6h.rds')
dseqr:::dl_pert_signature(sig_path, pert_type = 'cmap')

}
\keyword{internal}
