% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-docs-ui.R
\name{docsPageUI}
\alias{docsPageUI}
\title{UI for docs page (add new sections here)}
\usage{
docsPageUI(id, tab, active)
}
\description{
UI for docs page (add new sections here)
}
\keyword{(add}
\keyword{UI}
\keyword{docs}
\keyword{for}
\keyword{here)}
\keyword{internal}
\keyword{new}
\keyword{page}
\keyword{sections}
