% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{drugsGenesPlotly}
\alias{drugsGenesPlotly}
\title{Logic query/drug genes plotly}
\usage{
drugsGenesPlotly(
  input,
  output,
  session,
  data_dir,
  top_table,
  drug_study,
  show_genes,
  pert_signature
)
}
\description{
Logic query/drug genes plotly
}
\keyword{internal}
