% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{drugsPage}
\alias{drugsPage}
\title{Logic for Drugs page}
\usage{
drugsPage(
  input,
  output,
  session,
  new_anal,
  data_dir,
  pert_query_dir,
  pert_signature_dir
)
}
\description{
Logic for Drugs page
}
\keyword{internal}
