% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{drugsTable}
\alias{drugsTable}
\title{Logic for drug table}
\usage{
drugsTable(
  input,
  output,
  session,
  query_res,
  drug_study,
  anal,
  cells,
  show_clinical,
  sort_by,
  min_signatures,
  is_pert,
  direction
)
}
\description{
Logic for drug table
}
\keyword{internal}
