% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsDataset}
\alias{dsDataset}
\title{Logic for selected dataset part of dsFrom}
\usage{
dsDataset(
  input,
  output,
  session,
  sc_dir,
  bulk_dir,
  data_dir,
  new_dataset,
  explore_eset
)
}
\description{
Logic for selected dataset part of dsFrom
}
\keyword{internal}
