% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsExploreTable}
\alias{dsExploreTable}
\title{Logic for differential expression analysis table}
\usage{
dsExploreTable(
  input,
  output,
  session,
  eset,
  labels,
  data_dir,
  dataset_dir,
  dataset_name,
  svobj_r,
  numsv_r
)
}
\description{
Logic for differential expression analysis table
}
\keyword{internal}
