% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsFormAnal}
\alias{dsFormAnal}
\title{Logic for differential expression analysis part of dsForm}
\usage{
dsFormAnal(
  input,
  output,
  session,
  data_dir,
  sc_dir,
  bulk_dir,
  dataset_name,
  dataset_dir,
  new_anal,
  new_dataset,
  explore_eset
)
}
\description{
Logic for differential expression analysis part of dsForm
}
\keyword{internal}
