% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsFormAnal}
\alias{dsFormAnal}
\title{Logic for differential expression analysis part of dsForm}
\usage{
dsFormAnal(input, output, session, error_msg, dataset_name, data_dir,
  new_anal, dataset_dir)
}
\description{
Logic for differential expression analysis part of dsForm
}
\keyword{internal}
