% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-server.R
\name{dsMDSplotly}
\alias{dsMDSplotly}
\title{Logic for Dataset MDS plotly}
\usage{
dsMDSplotly(input, output, session, data_dir, dataset_dir, anal_name, new_anal)
}
\description{
Logic for Dataset MDS plotly
}
\keyword{internal}
