% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_results.R
\name{explore_results}
\alias{explore_results}
\title{Explore query results}
\usage{
explore_results(query_res)
}
\arguments{
\item{query_res}{\code{data.frame} returned from \code{\link{append_pdata}}}
}
\description{
Explore query results
}
\examples{

# load CMAP02 data
cmap_path <- system.file('extdata', 'cmap_es_ind.rds', package = 'drugseqr')
cmap_es <- readRDS(cmap_path)

# load previous differential expression analysis
data_dir <- 'data-raw/example-data'
anal <- readRDS(file.path(data_dir, 'diff_expr_symbol.rds'))

# get dprime effect size values for analysis
dprimes <- get_dprimes(anal)

# get correlations between query and drug signatures
res <- query_drugs(dprimes, cmap_es)

# append drug pdata
res <- append_pdata(res, 'CMAP02')

explore_results(res)

}
