% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{explore_scseq_clusters}
\alias{explore_scseq_clusters}
\title{Explore Single Cell Clusters}
\usage{
explore_scseq_clusters(base_dir, test = FALSE, test_data = TRUE)
}
\description{
Explore Single Cell Clusters
}
\examples{

base_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/IBD_20190712'

explore_scseq_clusters(base_dir, test_data = FALSE)

}
