% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{explore_scseq_clusters}
\alias{explore_scseq_clusters}
\title{Explore Single Cell Clusters}
\usage{
explore_scseq_clusters(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}
}
\description{
Explore Single Cell Clusters
}
\examples{

# import alevin quants
data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
sce <- load_scseq(data_dir)

# subset by alevin whitelist norm/stabilize using good cell only
sce <- sce[, sce$whitelist]
sce <- norm_scseq(sce)
sce <- stabilize_scseq(sce)

# get clusters and run tSNE
sce <- add_scseq_clusters(sce)

set.seed(1000)
sce <- scater::runTSNE(sce, use_dimred="PCA")

explore_scseq_clusters(sce)

}
