% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{explore_scseq_clusters}
\alias{explore_scseq_clusters}
\title{Explore Single Cell Clusters}
\usage{
explore_scseq_clusters(data_dir, test = FALSE, test_data = TRUE)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object}

\item{markers}{Named list of \code{data.frame}s where \code{row.names} are the marker genes. One list per cluster in \code{scseq}
with the same name as the cluster.}
}
\description{
Explore Single Cell Clusters
}
\examples{

data_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/single-cell/example-anals/sjia'

explore_scseq_clusters(data_dir, test_data = FALSE)

}
