% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{explore_scseq_clusters}
\alias{explore_scseq_clusters}
\title{Explore Single Cell Clusters}
\usage{
explore_scseq_clusters(scseq, markers = NULL, pt.size = 3)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object}

\item{markers}{Named list of \code{data.frame}s where \code{row.names} are the marker genes. One list per cluster in \code{scseq}
with the same name as the cluster.}
}
\description{
Explore Single Cell Clusters
}
\examples{

# import kallisto quants
data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
scseq <- load_scseq(data_dir)

# subset by whitelist norm/stabilize using good cell only
scseq <- scseq[, scseq$whitelist]
scseq <- preprocess_scseq(scseq)

# get clusters and run tSNE
scseq <- add_scseq_clusters(scseq, resolution = 1.6)
scseq <- run_tsne(scseq)

markers <- get_scseq_markers(scseq)

explore_scseq_clusters(scseq, markers)

}
