% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{exploreEset}
\alias{exploreEset}
\title{Logic to setup explore_eset for Bulk Data plots}
\usage{
exploreEset(eset, fastq_dir, explore_pdata, numsv, svobj)
}
\description{
Logic to setup explore_eset for Bulk Data plots
}
\keyword{internal}
