% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{fit_ebayes}
\alias{fit_ebayes}
\title{Perform eBayes analysis from limma.}
\usage{
fit_ebayes(eset, contrasts, mod, rna_seq = TRUE)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}. Non-selected samples
and duplicate features removed by \code{add_contrasts} and
\code{iqr_replicates}.}

\item{contrasts}{Character vector of contrast names generated by
\code{add_contrasts}.}

\item{mod}{Model matrix generated by \code{diff_setup}. With
or without surrogate variables.}
}
\value{
result from call to limma \code{eBayes}.
}
\description{
Generates contrast matrix then runs eBayes analysis from limma.
}
