% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{fit_lm_scseq}
\alias{fit_lm_scseq}
\title{Run limma differential expression between test and control groups in single cell RNA-Seq dataset}
\usage{
fit_lm_scseq(scseq, dataset_dir, clusters_name)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{clusters}{Character vector of clusters to include in analysis}
}
\value{
Named list with slots: \itemize{
 \item top_table result of limma::topTable.
 \item ebayes_sv results of \code{\link{fit_ebayes_scseq}}.
 \item pdata \code{data.frame} with column \code{'group'} indicating test and control samples.
}
}
\description{
Run limma differential expression between test and control groups in single cell RNA-Seq dataset
}
\seealso{
\code{\link{get_ambient}}
}
\keyword{internal}
