% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{fit_lm_scseq}
\alias{fit_lm_scseq}
\title{Fit limma eBayes on single cell RNA-seq dataset}
\usage{
fit_lm_scseq(scseq)
}
\arguments{
\item{scseq}{\code{Seurat} object.}

\item{ident.1}{String with control group name in \code{Idents(scseq)}.}

\item{ident.2}{String with test group name in \code{Idents(scseq)}.}
}
\value{
result of call to \code{\link[limma]{eBayes}}
}
\description{
Fit limma eBayes on single cell RNA-seq dataset
}
\keyword{internal}
