% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{from_crossmeta}
\alias{from_crossmeta}
\title{Covert from crossmeta to drugseqr formats}
\usage{
from_crossmeta(gse_name, data_dir)
}
\arguments{
\item{gse_name}{Name of GSE accession dataset, must be folder in data_dir}

\item{data_dir}{directory that contains \code{gse_name} folder}
}
\value{

}
\description{
Covert from crossmeta to drugseqr formats
}
\examples{
data_dir <- 'data-raw/patient_data/covid19/bulk'
gse_name <- 'GSE17400'
}
