% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{get_ambience}
\alias{get_ambience}
\title{Get sample ambience profiles}
\usage{
get_ambience(combined)
}
\arguments{
\item{combined}{\code{SingleCellExperiment} object.}
}
\value{
matrix of ambience profiles for each sample
}
\description{
Get sample ambience profiles
}
\keyword{internal}
