% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cluster_choices}
\alias{get_cluster_choices}
\title{Get cluster choices data.frame for selectize dropdown}
\usage{
get_cluster_choices(
  clusters,
  dataset_dir,
  scseq = NULL,
  sample_comparison = FALSE,
  top_tables = NULL
)
}
\arguments{
\item{clusters}{Character vector of cluster names}

\item{dataset_dir}{Directory with single cell dataset.}

\item{sample_comparison}{is this for test vs control comparion? Default is \code{FALSE}.}
}
\value{
data.frame with columns for rendering selectizeInput cluster choices
}
\description{
Get cluster choices data.frame for selectize dropdown
}
\keyword{internal}
