% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cluster_choices}
\alias{get_cluster_choices}
\title{Get cluster choices data.frame for selectize dropdown}
\usage{
get_cluster_choices(clusters, anal_name, sc_dir,
  sample_comparison = FALSE)
}
\arguments{
\item{clusters}{Character vector of cluster names}

\item{sample_comparison}{is this for test vs control comparion? Default is \code{FALSE}.}

\item{scseq}{\code{Seurat} object}

\item{value}{Character vector for value column which is returned from \code{selectizeInput}. Default is \code{clusters}.}
}
\value{
data.frame with columns for rendering selectizeInput cluster choices
}
\description{
Get cluster choices data.frame for selectize dropdown
}
\keyword{internal}
