% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cluster_choices}
\alias{get_cluster_choices}
\title{Get cluster choices data.frame for selectize dropdown}
\usage{
get_cluster_choices(clusters, sample_comparison = FALSE, with_all = FALSE, ...)
}
\arguments{
\item{clusters}{Character vector of cluster names}

\item{sample_comparison}{is this for test vs control comparion? Default is \code{FALSE}.}

\item{...}{Named arguments to \code{\link{get_cluster_stats}}.}

\item{dataset_dir}{Directory with single cell dataset.}
}
\value{
data.frame with columns for rendering selectizeInput cluster choices
}
\description{
Get cluster choices data.frame for selectize dropdown
}
\keyword{internal}
