% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cluster_stats}
\alias{get_cluster_stats}
\title{Get/Save cluster stats for single-cell related selectizeInputs}
\usage{
get_cluster_stats(
  resoln_dir = NULL,
  scseq = NULL,
  top_tables = NULL,
  use_disk = FALSE,
  sample_comparison = FALSE,
  contrast_dir = NULL
)
}
\arguments{
\item{resoln_dir}{Sub directory with single cell dataset info specific to resolution.}

\item{scseq}{\code{DietSCE} object to get/save stats for.
if \code{NULL} (Default), will be loaded.}

\item{top_tables}{List of \code{limma::topTable} results used by
scSampleComparison for integrated datasets.}

\item{use_disk}{Should results by saved to disk? used by scSampleComparison
to persist results to disk.}

\item{sample_comparison}{is this for test vs control comparion? Default is \code{FALSE}.}

\item{has_replicates}{Boolean indicating if integrated dataset has multiple
samples in at least one group. Used by scSampleComparison.}
}
\value{
List with cluster stats
}
\description{
Get/Save cluster stats for single-cell related selectizeInputs
}
