% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{get_cluster_stats}
\alias{get_cluster_stats}
\title{Get/Save cluster stats for single-cell related selectizeInputs}
\usage{
get_cluster_stats(
  dataset_dir = NULL,
  scseq = NULL,
  top_tables = NULL,
  has_replicates = FALSE,
  use_disk = FALSE,
  sample_comparison = FALSE
)
}
\arguments{
\item{dataset_dir}{Directory with single cell dataset.}

\item{scseq}{\code{SingleCellExperiment} object to get/save stats for. if \code{NULL} (Default), will be loaded.}
}
\value{
List with cluster stats
}
\description{
Get/Save cluster stats for single-cell related selectizeInputs
}
