% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{get_dprimes}
\alias{get_dprimes}
\title{Get dprime effect size values.}
\usage{
get_dprimes(top_table)
}
\value{
Named numeric vector. Names are gene names, values are effect size values.
}
\description{
These are used to query against drug effect size matrices.
}
\examples{

# load result of previous differential expression analysis
data_dir <- file.path('data-raw/example-data')
anal <- readRDS(file.path(data_dir, 'diff_expr_symbol.rds'))

dprimes <- get_dprimes(anal)

}
