% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{get_dprimes}
\alias{get_dprimes}
\title{Get dprime effect size values.}
\usage{
get_dprimes(top_table)
}
\arguments{
\item{top_table}{data.frame of differential expression results.}
}
\value{
Named numeric vector. Names are gene names, values are effect size values.
}
\description{
These are used to query against drug effect size matrices.
}
\examples{

# generate fake previous result
top_table <- data.frame(dprimes=rnorm(5), row.names = paste0('gene', 1:5))

dprimes <- drugseqr:::get_dprimes(top_table)

}
\keyword{internal}
