% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pairs.R
\name{get_fastq_id1s}
\alias{get_fastq_id1s}
\title{Get first sequence identifiers for fastq.gz files}
\usage{
get_fastq_id1s(fastq_paths)
}
\arguments{
\item{fastq_paths}{Character vector of paths to fastq.gz files}
}
\value{
Named character vector of first sequence id lines (start with @) in \code{fastq_paths}.
  Names are the \code{fastq_paths}.
}
\description{
Function is used to determine if an experiment is single or paired.
}
\examples{
# required fastq.gz files in data-raw/example-data (e.g. IBD example)
fastq_paths <- list.files(file.path('data-raw', 'example-data'), '.fastq.gz$', full.names = TRUE)
fastq_id1s <- get_fastq_id1s(fastq_paths)

}
\keyword{internal}
