% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_gslist}
\alias{get_gslist}
\title{Get gslist for pathway analysis}
\usage{
get_gslist(species = "Hs", type = "go", gs_dir = "/srv/drugseqr/gs_dir")
}
\arguments{
\item{species}{Species identifier}

\item{type}{either 'go' or 'kegg'}

\item{gs_dir}{Directory to save gslist to}
}
\value{
gslist
}
\description{
Get gslist for pathway analysis
}
