% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_bulk.R
\name{get_kallisto_index}
\alias{get_kallisto_index}
\title{Get path to kallisto index}
\usage{
get_kallisto_index(indices_dir, species = "homo_sapiens", release = "94")
}
\arguments{
\item{indices_dir}{Path to folder with indices dir}

\item{species}{Character vector giving species}

\item{release}{Character vector of EnsDB release number. Release \code{'94'} is the default because HGNC symbols
most closely overlap with those used to annotate CMAP02 and L1000.}
}
\value{
Path to kallisto index
}
\description{
Get path to kallisto index
}
\keyword{internal}
