% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{get_knee}
\alias{get_knee}
\title{Get Roryk knee point and plot on barcode rank plot}
\usage{
get_knee(counts)
}
\arguments{
\item{counts}{dgTMatrix of counts for all barcodes.}
}
\value{
count value corresponding to Roryk knee point
}
\description{
Get Roryk knee point and plot on barcode rank plot
}
\keyword{internal}
