% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_mods}
\alias{get_mods}
\title{Get model matrices for surrogate variable analysis}
\usage{
get_mods(eset)
}
\arguments{
\item{eset}{Annotated eset with samples selected during \code{add_contrasts}.}

\item{svanal}{Perform surrogate variable analysis? Default is \code{TRUE}.}

\item{rna_seq}{Is this an RNA-seq experiment? Inferred from \code{eset} in \code{\link{diff_expr}}.}
}
\value{
List with model matrix(mod), model matrix with surrogate
        variables(modsv), and result of \code{sva} function.
}
\description{
Used by \code{diff_expr} to create model matrix with surrogate variables
in order to run \code{diff_anal}.
}
\seealso{
\code{\link{add_contrast}}, \code{\link{diff_expr}}.
}
