% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{get_palette}
\alias{get_palette}
\title{Get a pallete for cluster plots}
\usage{
get_palette(levs, dark = FALSE, with_all = FALSE)
}
\arguments{
\item{levs}{Character vector of levels to get colour pallete for.}
}
\value{
Character vector with colour codes of \code{length(levs)}.
}
\description{
Get a pallete for cluster plots
}
\keyword{internal}
