% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{get_path_df}
\alias{get_path_df}
\title{Get data.frame for plotting gene expression values of a pathway}
\usage{
get_path_df(anal, path_id = NULL, path_genes = NULL)
}
\arguments{
\item{anal}{Result of call to \code{\link{diff_expr_scseq}}}

\item{path_id}{String with KEGG pathway id.}
}
\value{
\code{data.frame} with columns: \itemize{
 \item Gene gene names.
 \item Dprime standardized unbiased effect size values.
 \item sd standard deviations of \code{Dprime}.
 \item Link url to GeneCards page for gene.
}
}
\description{
Get data.frame for plotting gene expression values of a pathway
}
\keyword{internal}
