% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{get_path_res}
\alias{get_path_res}
\title{Get and save pathway results for ebfit object}
\usage{
get_path_res(ebfit, go_path, kegg_path, species = "Hs", min.genes = 4)
}
\arguments{
\item{ebfit}{Result of \code{fit_ebayes}}

\item{go_path}{Path to save Gene Ontology result}

\item{kegg_path}{Path to save KEGG result}
}
\value{
List with GO and KEGG results
}
\description{
Used to avoid code reuse for single-cell and bulk
}
\keyword{internal}
