% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-server.R
\name{get_path_res}
\alias{get_path_res}
\title{Get and save pathway results for ebfit object}
\usage{
get_path_res(ebfit, goana_path, kegga_path, species = "Hs")
}
\arguments{
\item{ebfit}{Result of \code{fit_ebayes}}

\item{goana_path}{Path to save \code{goana} result}

\item{kegga_path}{Path to save \code{kegga} result}
}
\value{
List with \code{goana} and \code{kegga} results
}
\description{
Used to avoid code reuse for single-cell and bulk
}
\keyword{internal}
