% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-bulk-utils.R
\name{get_path_res}
\alias{get_path_res}
\title{Get and save pathway results for ebfit object}
\usage{
get_path_res(
  de,
  goana_path,
  gs_dir,
  species = "Hs",
  genego = NULL,
  gonames = NULL,
  coef = ncol(de),
  nmin = 50,
  max_fdr = 0.05,
  min_abs_logfc = 0
)
}
\arguments{
\item{de}{Result of \code{\link[crossmeta]{fit_ebayes}} or
\code{link[crossmeta]{get_top_table}}}

\item{goana_path}{Path to save goana Gene Ontology result}
}
\value{
List with GO and KEGG results
}
\description{
Used to avoid code reuse for single-cell and bulk
}
\keyword{internal}
