% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_presto_markers}
\alias{get_presto_markers}
\title{Get cluster markers using presto}
\usage{
get_presto_markers(scseq)
}
\arguments{
\item{scseq}{SingleCellExperiment object}
}
\value{
list of data.frames, one for each cluster
}
\description{
Get cluster markers using presto
}
\keyword{internal}
