% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_salmon_version}
\alias{get_salmon_version}
\title{Get version of salmon from system command.}
\usage{
get_salmon_version(command)
}
\arguments{
\item{command}{System command for salmon. Used to determine salmon version.}
}
\value{
Version of salmon.
}
\description{
Get version of salmon from system command.
}
\keyword{internal}
