% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_scseq_markers}
\alias{get_scseq_markers}
\title{Get markers genes for single cell clusters}
\usage{
get_scseq_markers(scseq, assay.type = "logcounts")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object.}

\item{assay.type}{Used by \code{\link[scran]{findMarkers}}}
}
\value{
List of \code{data.frame}s, one for each cluster.
}
\description{
If \code{scseq} is a \code{SingleCellExperiment} object then uses \code{scran::findMarkers}.
If \code{scseq} is a \code{Seurat} object then uses \code{Seurat::FindAllMarkers}.
}
