% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_scseq_markers}
\alias{get_scseq_markers}
\title{Get markers genes for single cell clusters}
\usage{
get_scseq_markers(
  scseq,
  groups = scseq$cluster,
  direction = "up",
  test = "wilcox",
  pval.type = "some",
  block = NULL,
  restrict = NULL
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{...}{arguments to \code{findMarkers}}
}
\value{
List of \code{data.frame}s, one for each cluster.
}
\description{
Get markers genes for single cell clusters
}
