% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scseq_whitelist.R
\name{get_scseq_whitelist}
\alias{get_scseq_whitelist}
\title{Get whitelist for high quality cells}
\usage{
get_scseq_whitelist(
  counts,
  data_dir,
  overwrite = FALSE,
  knee_type = c("inflection", "roryk", "knee"),
  species = "Homo sapiens"
)
}
\arguments{
\item{counts}{dgTMatrix of counts for all barcodes.}

\item{data_dir}{Directory with raw and kallisto/bustools or CellRanger quantified single-cell RNA-Seq files.}

\item{knee_type}{Knee type used for cell quality whitelist modeling by \code{get_scseq_whitelist}.}
}
\value{
Character vector of barcodes called as high quality cells.
}
\description{
Get whitelist for high quality cells
}
\keyword{internal}
