% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{get_species}
\alias{get_species}
\title{Get tx2gene for human or mouse depending on intersections}
\usage{
get_species(counts)
}
\arguments{
\item{counts}{dgCMatrix where rownames are gene ids}
}
\value{
tx2gene for human or mouse
}
\description{
Get tx2gene for human or mouse depending on intersections
}
\keyword{internal}
