% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{get_species}
\alias{get_species}
\title{Get species for human or mouse depending on gene ids}
\usage{
get_species(x)
}
\arguments{
\item{x}{dgCMatrix or data.frame where rownames are gene ids}
}
\value{
either 'Homo sapiens' or 'Mus musculus'
}
\description{
Get species for human or mouse depending on gene ids
}
\keyword{internal}
