% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{get_top}
\alias{get_top}
\title{Reduce entries in query results}
\usage{
get_top(query_cors, arrange_by, ntop, decreasing = FALSE)
}
\arguments{
\item{query_cors}{\code{data.frame} with columns \code{'Compound'} and \code{arrange_by}.}

\item{arrange_by}{String indicating column name in \code{query_cors} do sort by.}

\item{ntop}{Integer indicating the number of rows to keep after sorting by \code{arrange_by}.}
}
\value{
Character vector of \code{ntop} compounds.
}
\description{
Gets all entries for compounds that have a correlation less than the top entries with a clinical phase.
}
