% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{identify_sc_files}
\alias{identify_sc_files}
\title{Get 10X FastQ paths for a read type and sort by increasing lane number}
\usage{
identify_sc_files(data_dir, read_type = "R1")
}
\arguments{
\item{data_dir}{Path to folder with 10X fastq.gz scRNA-seq files}

\item{read_type}{One of \code{'R1'} or \code{'R2'} for CB+UMI and read sequences respectively.}
}
\value{
Character vector of lane ordered paths to FastQ read type files.
}
\description{
Get 10X FastQ paths for a read type and sort by increasing lane number
}
\keyword{internal}
