% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{import_quants}
\alias{import_quants}
\title{Import salmon quant.sf files}
\usage{
import_quants(data_dir, tx2gene)
}
\arguments{
\item{data_dir}{Directory with a folder named 'quants' that contains salmon quantification folders for each sample.}

\item{tx2gene}{\code{data.frame} mapping transcripts to gene names. Returned from from \code{\link{get_tx2gene}}.}
}
\value{
\code{DGEList} with length scaled counts. Lowly expressed genes are filtered.
}
\description{
Import salmon quant.sf files
}
\keyword{internal}
