% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{import_quants}
\alias{import_quants}
\title{Import kallisto quants}
\usage{
import_quants(
  data_dir,
  type,
  species = "Homo sapiens",
  release = "94",
  return_deseq = FALSE
)
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{type}{Either \code{'salmon'} or \code{'kallisto'}. The package used for quantification. Must be on the PATH.}

\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link[drugseqr.data]{build_kallisto_index}}.}

\item{return_deseq}{if \code{TRUE}, returns tximport result for DESeq2 workflow.}
}
\value{
\code{DGEList} with length scaled counts.
}
\description{
Import kallisto quants
}
\keyword{internal}
