% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{init_drugseqr}
\alias{init_drugseqr}
\title{Initialize drugseqr folders/files for a new app}
\usage{
init_drugseqr(app_name)
}
\arguments{
\item{app_name}{Name for new drugseqr app.}
}
\description{
Creates necessary folders/files for a new drugseqr app inside of /srv/shiny-server/drugseqr.
}
\examples{

init_drugseqr('example')

}
