% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{init_drugseqr}
\alias{init_drugseqr}
\title{Initialize drugseqr folders/files for a new app}
\usage{
init_drugseqr(app_name, app_dir = "/srv/drugseqr")
}
\arguments{
\item{app_name}{Name for new drugseqr app.}

\item{app_dir}{Path to put \code{app_name} directory where app will be initialized. Default is \code{'/srv/drugseqr'} for shiny.}
}
\description{
Creates necessary folders/files for a new drugseqr app inside of /srv/shiny-server/drugseqr.
}
\examples{

# app_dir for local development
init_drugseqr('example', app_dir = 'data-raw/patient_data')

}
