% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dseqr.R
\name{init_dseqr}
\alias{init_dseqr}
\title{Initialize dseqr folders/files for a new app}
\usage{
init_dseqr(project_name, data_dir = "/srv/dseqr")
}
\arguments{
\item{data_dir}{Path to put \code{anal_name} directory where app will be
initialized. Default is \code{'/srv/dseqr'} (for hosting app on server).}

\item{anal_name}{Name for new dseqr app.}
}
\description{
Creates necessary folders/files for a new dseqr app inside of /srv/shiny-server/dseqr.
}
\examples{

data_dir <- tempdir()
init_dseqr('example', data_dir)

}
