% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{integrate_saved_scseqs}
\alias{integrate_saved_scseqs}
\title{Integrate previously saved scseqs}
\usage{
integrate_saved_scseqs(
  sc_dir,
  test,
  ctrl,
  exclude_clusters,
  integration_name,
  integration_type = c("harmony", "liger", "fastMNN"),
  pairs = NULL,
  progress = NULL,
  value = 0
)
}
\arguments{
\item{sc_dir}{Directory with saved single-cell datasets.}

\item{test}{Character vector of test analysis names.}

\item{ctrl}{Character vector of control analysis names.}

\item{progress}{optional Shiny \code{Progress} object.}

\item{dataset_name}{Name for new integrated analysis.}
}
\description{
Performs integration and saves as a new analysis.
Used by \code{explore_scseq_clusters} shiny app.
}
\keyword{internal}
