% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{integrate_saved_scseqs}
\alias{integrate_saved_scseqs}
\title{Integrate previously saved scseqs}
\usage{
integrate_saved_scseqs(
  sc_dir,
  test,
  ctrl,
  exclude_clusters,
  anal_name,
  updateProgress = NULL,
  use_scalign = FALSE
)
}
\arguments{
\item{sc_dir}{Directory with saved single-cell datasets.}

\item{test}{Character vector of test analysis names.}

\item{ctrl}{Character vector of control analysis names.}

\item{anal_name}{Name for new integrated analysis.}

\item{progress}{optional Shiny \code{Progress} object.}
}
\description{
Performs integration and saves as a new analysis.
Used by \code{explore_scseq_clusters} shiny app.
}
\keyword{internal}
