% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{integrate_scseqs}
\alias{integrate_scseqs}
\title{Integrate multiple scRNA-seq samples}
\usage{
integrate_scseqs(scseqs)
}
\arguments{
\item{scseqs}{List of \code{Seurat} objects}
}
\value{
Integrated \code{Seurat} object with default assay of \code{"integrated"}
}
\description{
Integrate multiple scRNA-seq samples
}
