% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{integrate_scseqs}
\alias{integrate_scseqs}
\title{Integrate multiple scRNA-seq samples}
\usage{
integrate_scseqs(
  scseqs,
  type = c("harmony", "fastMNN", "Azimuth", "symphony"),
  pairs = NULL,
  hvgs = NULL,
  ref_name = NULL
)
}
\arguments{
\item{scseqs}{List of \code{SingleCellExperiment} objects}

\item{type}{One of \code{'harmony'} (default) or \code{'fastMNN'} specifying integration to use.}

\item{pairs}{data.frame with columns \code{'sample'} with sample
names of \code{scseqs} and \code{'pair'} with integers indicating paired
samples. If not \code{NULL} (default), then harmony includes pairings
as a covariate.}
}
\value{
Integrated \code{SingleCellExperiment} object.
}
\description{
Integrate multiple scRNA-seq samples
}
\keyword{internal}
