% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{integrate_scseqs}
\alias{integrate_scseqs}
\title{Integrate multiple scRNA-seq samples}
\usage{
integrate_scseqs(scseqs, type = c("harmony", "liger", "fastMNN"), pairs = NULL)
}
\arguments{
\item{scseqs}{List of \code{SingleCellExperiment} objects}

\item{type}{One of \code{'harmony'} (default), \code{'liger'} or \code{'fastMNN'} specifying integration to use.}
}
\value{
Integrated \code{SingleCellExperiment} object.
}
\description{
Integrate multiple scRNA-seq samples
}
