% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{integrationForm}
\alias{integrationForm}
\title{Logic for integration form toggled by showIntegration}
\usage{
integrationForm(
  input,
  output,
  session,
  sc_dir,
  datasets,
  show_integration,
  dataset_name
)
}
\description{
Logic for integration form toggled by showIntegration
}
\keyword{internal}
