% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{integrationForm}
\alias{integrationForm}
\title{Logic for integration form toggled by showIntegration}
\usage{
integrationForm(input, output, session, sc_dir, anal_options,
  show_integration, selected_anal, scseq)
}
\description{
Logic for integration form toggled by showIntegration
}
\keyword{internal}
