% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{integrationFormInput}
\alias{integrationFormInput}
\title{Input form for integrating single cell datasets}
\usage{
integrationFormInput(id)

integrationFormInput(id)
}
\description{
Input form for integrating single cell datasets

Input form for integrating single cell datasets
}
\keyword{internal}
