% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{iqr_replicates}
\alias{iqr_replicates}
\title{Removes features with replicated annotation.}
\usage{
iqr_replicates(
  eset,
  mod = NULL,
  svobj = NULL,
  annot = "SYMBOL",
  rm.dup = FALSE,
  rna_seq = TRUE
)
}
\arguments{
\item{eset}{Annotated eset created by \code{\link{load_seq}}.}

\item{mod}{Model matrix without surrogate variables. generated by \code{diff_setup}.}

\item{svobj}{Result from \code{sva} function called during \code{diff_setup}.}

\item{annot}{feature to use to remove replicates.}

\item{rm.dup}{remove duplicates (same measure, multiple ids)?}

\item{rna_seq}{Is this an RNA-seq experiment? Inferred from \code{eset} in \code{\link{diff_expr}}.}
}
\value{
List with:
   \item{eset}{Expression set with unique features at probe or gene level.}
   \item{exprs_sva}{Expression data from eset with effect of surrogate
      variable removed.}
}
\description{
For rows with duplicated annot, highested IQR retained.
}
