% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{jitter_umap}
\alias{jitter_umap}
\title{Add jitter to UMAP embeddings}
\usage{
jitter_umap(scseq, factor = 1, amount = 0)
}
\arguments{
\item{scseq}{\code{Seurat} object with umap reduction}
}
\value{
\code{scseq} with jitter added to umap embedding
}
\description{
Add jitter to UMAP embeddings
}
