% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_results.R
\name{limit_cells}
\alias{limit_cells}
\title{Limit Query Results to Specific Cell Lines}
\usage{
limit_cells(query_res, cells)
}
\arguments{
\item{query_res}{\code{data.frame} returned by \code{\link{append_pdata}}.}

\item{cells}{Character vector of cell lines to limit \code{query_res} by.}
}
\value{
\code{query_res} for specified \code{cells}
}
\description{
Limit Query Results to Specific Cell Lines
}
