% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-utils.R
\name{limit_cells}
\alias{limit_cells}
\title{Limit Query Results to Specific Cell Lines}
\usage{
limit_cells(query_table, cells = NULL)
}
\arguments{
\item{query_table}{\code{data.frame} with \code{cell_line} column.}

\item{cells}{Character vector of cell lines to limit \code{query_table} by. If \code{NULL} (default),
no filtering occurs.}
}
\value{
\code{query_table} for specified \code{cells}
}
\description{
Limit Query Results to Specific Cell Lines
}
\keyword{internal}
