% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_raw_scseq}
\alias{load_raw_scseq}
\title{Process raw single cell fastq or cellranger files for app}
\usage{
load_raw_scseq(
  dataset_name,
  fastq_dir,
  sc_dir,
  indices_dir,
  progress = NULL,
  recount = FALSE,
  value = 0,
  metrics = c("low_lib_size", "low_n_features", "high_subsets_mito_percent",
    "low_subsets_ribo_percent", "high_doublet_score", "high_outlyingness")
)
}
\arguments{
\item{dataset_name}{Name of dataset}

\item{fastq_dir}{Directory with fastq or cellranger files}

\item{sc_dir}{Single cell directory for app. Will store results in \code{dataset_name} subdirectory}

\item{progress}{Optional shiny \code{Progress} object. Default will print progress.}
}
\description{
Process raw single cell fastq or cellranger files for app
}
