% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_saved_scseq}
\alias{load_saved_scseq}
\title{Load saved single cell RNA-seq dataset}
\usage{
load_saved_scseq(anal, data_dir)
}
\arguments{
\item{anal}{Name of single cell analysis and containing folder.}

\item{data_dir}{Path to directory containing \code{anal} folder.}
}
\value{
\code{SingleCellExperiment} object
}
\description{
Used for both dataset integration and label transfer.
}
\keyword{internal}
