% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq}
\alias{load_scseq}
\title{Load kallisto/bustools quantification into a Seurat or SingleCellExperiment object.}
\usage{
load_scseq(data_dir, type = "Seurat", project = "SeuratProject",
  soupx = FALSE)
}
\arguments{
\item{type}{Object type to return. Either \code{'Seurat'} or \code{'SingleCellExperiment'}.}

\item{Directory}{with raw and kallisto/bustools-quantified single-cell RNA-Seq files.}
}
\value{
\code{Seurat} (default) or \code{SingleCellExperiment} with whitelist meta data.
}
\description{
Load kallisto/bustools quantification into a Seurat or SingleCellExperiment object.
}
\examples{

data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
load_scseq(data_dir)

}
