% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq}
\alias{load_scseq}
\title{Load SingleCellExperiment}
\usage{
load_scseq(dataset_dir, default_clusters = TRUE)
}
\arguments{
\item{dataset_dir}{Directory with scle.loom or scseq.qs}
}
\value{
\code{SingleCellExperiment} or \code{SingleCellLoomExperiment}
}
\description{
Loads scle.loom if exists otherwise scseq.qs
}
\keyword{internal}
