% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq}
\alias{load_scseq}
\title{Load kallisto/bustools quantification into a Seurat or SingleCellExperiment object.}
\usage{
load_scseq(
  data_dir,
  project = "SeuratProject",
  type = c("kallisto", "cellranger"),
  soupx = FALSE
)
}
\arguments{
\item{data_dir}{Directory with raw and kallisto/bustools or CellRanger quantified single-cell RNA-Seq files.}

\item{project}{String identifying sample.}

\item{type}{Quantification file type. One of either \code{'kallisto'} or \code{'cell_ranger'}.}

\item{soupx}{Boolean indicating if \code{SoupX} should be used to remove backgroud counts (defaul is \code{FALSE}).
Experimental and not currently recommended.}

\item{h5}{Boolean indicating, for \code{type = 'cell_ranger'}, if \code{data_dir} container a hdf5 file.}
}
\value{
\code{Seurat} object with whitelist meta data.
}
\description{
Load kallisto/bustools quantification into a Seurat or SingleCellExperiment object.
}
\examples{

data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
load_scseq(data_dir)

}
