% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{load_scseq_qcgenes}
\alias{load_scseq_qcgenes}
\title{Load mitochondrial and ribsomal gene names}
\usage{
load_scseq_qcgenes(species = "Homo sapiens")
}
\value{
Named list with \code{rrna} and \code{mrna} character vectors.
}
\description{
This are the genes used by alevin for whitelisting.
}
