% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{load_scseq_qs}
\alias{load_scseq_qs}
\title{Load SingleCellExperiment from qs file}
\usage{
load_scseq_qs(dataset_dir)
}
\arguments{
\item{dataset_dir}{Path to folder with scseq.qs file}
}
\value{
SingleCellExperiment
}
\description{
Also attaches clusters from last applied leiden resolution and stores resolution.
}
\keyword{internal}
