% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{load_seq}
\alias{load_seq}
\title{Load raw RNA-Seq data into an ExpressionSet.}
\usage{
load_seq(data_dir, pdata_path, species = "Homo sapiens",
  release = "94")
}
\arguments{
\item{data_dir}{Directory with raw RNA-Seq files.}

\item{pdata_path}{Path to text file with sample annotations.}

\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link{build_index}}.}
}
\description{
Load raw RNA-Seq data into an ExpressionSet.
}
