% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{load_seq}
\alias{load_seq}
\title{Load bulk RNA-Seq data into an ExpressionSet.}
\usage{
load_seq(
  data_dir,
  type = "kallisto",
  species = "Homo sapiens",
  release = "94",
  load_saved = TRUE,
  save_eset = TRUE,
  save_dgel = FALSE
)
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{type}{Either \code{'salmon'} or \code{'kallisto'}. The package used for quantification. Must be on the PATH.}

\item{species}{Character vector indicating species. Genus and species should be space seperated, not underscore. Default is \code{Homo sapiens}.}

\item{release}{EnsemblDB release. Should be same as used in \code{\link[drugseqr.data]{build_kallisto_index}}.}

\item{load_saved}{If TRUE (default) and a saved \code{ExpressionSet} exists, will load from disk.}

\item{save_eset}{If TRUE (default) and either \code{load_saved} is \code{FALSE} or a saved \code{ExpressionSet} does not exist,
then an ExpressionSet will be saved to disk. Will overwrite if already exists.}

\item{save_dgel}{If TRUE, will save the \code{DGEList} object. Used for testing purposes. Default is \code{FALSE}.}
}
\value{
\code{\link[Biobase]{ExpressionSet}} with attributes/accessors:
\itemize{
  \item \code{sampleNames} from names of raw RNA-Seq files (excluding .fastq.gz suffix).
  \item \code{annotation} Character vector of annotation package used.
  \item \code{exprs} Length scaled counts generated from abundances for use in
    \code{\link[limma]{voom}} (see \code{vignette("tximport", package = "tximport")}).
  \item \code{phenoData} from \code{pdata_path} annotation file with added columns:
    \itemize{
      \item \code{lib.size} library size from \code{\link[edgeR]{calcNormFactors}}.
      \item \code{norm.factors} library normalization factors from \code{\link[edgeR]{calcNormFactors}}.
    }
}
}
\description{
Load bulk RNA-Seq data into an ExpressionSet.
}
\examples{

data_dir <- 'data-raw/patient_data/sjia/bulk/mono'
eset <- load_seq(data_dir, load_saved = FALSE)

}
