% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{navbarUI}
\alias{navbarUI}
\title{UI for navbar}
\usage{
navbarUI(tabs, active, logout_url = NULL)
}
\arguments{
\item{tabs}{Character vector of tab names to display}

\item{active}{the active tab name}

\item{logout_url}{url to logout from. NULL results in no Logout link}
}
\value{
shiny.tag with html for navbar
}
\description{
UI for navbar
}
\examples{

# UI with only Single-Cell page
navbarUI('Single Cell', 'Single Cell')

}
