% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-utils.R
\name{navbarUI}
\alias{navbarUI}
\title{UI for navbar}
\usage{
navbarUI(tabs, active)
}
\arguments{
\item{tabs}{Character vector of tab names to display}

\item{active}{the active tab name}
}
\value{
shiny.tag with html for navbar
}
\description{
UI for navbar
}
