% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{pairwise_wilcox}
\alias{pairwise_wilcox}
\title{Run pairwise wilcox tests between single cell clusters}
\usage{
pairwise_wilcox(
  scseq,
  groups = scseq$cluster,
  direction = "up",
  block = NULL,
  restrict = NULL
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}
}
\value{
List of \code{data.frame}s, one for each cluster.
}
\description{
Run pairwise wilcox tests between single cell clusters
}
