% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-server.R
\name{pathForm}
\alias{pathForm}
\title{Logic for form in Pathways tab}
\usage{
pathForm(input, output, session, new_dataset, data_dir)
}
\description{
Logic for form in Pathways tab
}
\keyword{internal}
