% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_scseq_report}
\alias{plot_scseq_report}
\title{Plot grid with scRNA-seq cluster and gene markers together}
\usage{
plot_scseq_report(sce, markers, point_size = 3)
}
\arguments{
\item{point_size}{}
}
\description{
Plot grid with scRNA-seq cluster and gene markers together
}
