% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_cluster}
\alias{plot_tsne_cluster}
\title{Plot TSNE coloured by cluster}
\usage{
plot_tsne_cluster(scseq, legend_title = NULL, selected_groups = NULL,
  point_size = 3, assay.type = "logcounts")
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat}.}

\item{legend_title}{Title for the legend. The default is \code{scseq@metadata$colour_by} with title case.}

\item{selected_groups}{The groups in \code{scseq$orig.ident} to show cell for. The default \code{NULL} shows all cells.}

\item{point_size}{Numeric scalar, specifying the size of the points. Defaults to 3.}

\item{assay.type}{The assay data that gets used by \code{\link[scater]{plotTSNE}}. If \code{scseq} is a \code{Seurat}
object then the default ('logcounts') will be from the \code{data} slot of the \code{Seurat::DefaultAssay}.}
}
\value{
\code{ggplot}
}
\description{
The groups are either \code{scseq$cluster} (default) or \code{scseq@metadata$colour_by}.
}
