% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_cluster}
\alias{plot_tsne_cluster}
\title{Plot TSNE coloured by cluster}
\usage{
plot_tsne_cluster(sce, legend_title, selected_groups = NULL,
  point_size = 3, assay.type = "logcounts", remove_axes = "xaxis")
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}

\item{legend_title}{Title for the legend.}

\item{selected_groups}{The groups in \code{sce$orig.ident} to show cell for. The default \code{NULL} shows all cells.}

\item{assay.type}{The assay data that gets used by \code{\link[scater]{plotTSNE}}. If \code{scseq} is a \code{Seurat}
object then the default ('logcounts') will be from the \code{data} slot of the \code{Seurat::DefaultAssay}.}
}
\value{
ggplot
}
\description{
The groups are either \code{sce$cluster} (default) or \code{sce@metadata$colour_by}.
}
