% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_cluster}
\alias{plot_tsne_cluster}
\title{Plot TSNE coloured by cluster}
\usage{
plot_tsne_cluster(
  scseq,
  label.highlight = NULL,
  legend_title = "Cluster",
  cols = NULL
)
}
\value{
\code{ggplot}
}
\description{
Plot TSNE coloured by cluster
}
