% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_cluster}
\alias{plot_tsne_cluster}
\title{Plot UMAP coloured by cluster}
\usage{
plot_tsne_cluster(
  scseq,
  selected_clusters = levels(scseq$cluster),
  pt.size = 2,
  legend_title = "Cluster",
  cols = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to \code{Seurat::DimPlot}.}
}
\value{
\code{ggplot}
}
\description{
Plot UMAP coloured by cluster
}
