% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{plot_tsne_feature}
\alias{plot_tsne_feature}
\title{Plot UMAP coloured by gene or QC metric}
\usage{
plot_tsne_feature(
  scseq,
  feature,
  cols = c("lightgray", "blue"),
  reverse_scale = feature \%in\% c("ribo_percent", "log10_sum", "log10_detected"),
  title = paste0("Expression by Cell: ", feature)
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} object.}

\item{feature}{Character vector specifying feature to colour cells by.}

\item{pt.size}{Numeric scalar, specifying the size of the points. Defaults to 3.}
}
\value{
\code{ggplot}
}
\description{
Plot UMAP coloured by gene or QC metric
}
