% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{preprocess_scseq}
\alias{preprocess_scseq}
\title{Utility wrapper to run normalization and variance stabilization}
\usage{
preprocess_scseq(scseq, use_sctransform = TRUE)
}
\arguments{
\item{scseq}{\code{Seurat} or \code{SingleCellExperiment} object}
}
\value{
Normalized and variance stabilized \code{scseq}.
}
\description{
If \code{scseq} is a \code{SingleCellExperiment} object then runs the simpleSingleCell workflow.
If \code{scseq} is a \code{Seurat} object then uses \code{SCTransform}.
}
