% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{preprocess_scseq}
\alias{preprocess_scseq}
\title{Utility wrapper to run normalization and variance stabilization}
\usage{
preprocess_scseq(scseq)
}
\arguments{
\item{scseq}{}
}
\description{
If \code{scseq} is a \code{SingleCellExperiment} object then runs the simpleSingleCell workflow.
If \code{scseq} is a \code{Seurat} object then uses \code{SCTransform}.
}
