% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{process_raw_scseq}
\alias{process_raw_scseq}
\title{Process Count Data for App}
\usage{
process_raw_scseq(
  scseq,
  dataset_name,
  sc_dir,
  cluster_alg = "leiden",
  npcs = 30,
  resoln = 1,
  hvgs = NULL,
  azimuth_ref = NULL,
  founder = NULL,
  progress = NULL,
  value = 0
)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment}}

\item{dataset_name}{Name of dataset to save}

\item{sc_dir}{Directory to save dataset to}

\item{founder}{Name of original founding dataset (may be ancestor of \code{from_dataset}).}

\item{progress}{Shiny progress object. Default (\code{NULL}) prints to stdout.}

\item{value}{Initial value of progress.}
}
\description{
Performs the following:
- normalization
- adds HVGs
- dimensionality reduction
- clustering
}
\details{
Then calls run_post_cluster
}
