% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-utils.R
\name{readRDS.safe}
\alias{readRDS.safe}
\title{Read RDS file safely}
\usage{
readRDS.safe(file, .nofile = NULL, .nullfile = NULL)
}
\arguments{
\item{file}{a \link[base]{connection} or the name of the file where the \R object
    is saved to or read from.}

\item{.nofile}{return value when file doesn't exist}

\item{.nullfile}{return value when file value is \code{NULL}}
}
\value{
For \code{readRDS}, an \R object.

  For \code{saveRDS}, \code{NULL} invisibly.

  For \code{infoRDS}, an \R list with elements \code{version} (version
  number, currently 2 or 3), \code{writer_version} (version of \R that
  produced the serialization), \code{min_reader_version} (minimum version of
  \R that can read the serialization), \code{format} (data representation)
  and \code{native_encoding} (native encoding of the session that produced
  the serialization, available since version 3).  The data representation is
  given as \code{"xdr"} for big-endian binary representation, \code{"ascii"}
  for ASCII representation (produced via \code{ascii = TRUE} or \code{ascii
  = NA}) or \code{"binary"} (binary representation with native
  \sQuote{endianness} which can be produced by \code{\link[base]{serialize}}).
}
\description{
Read RDS file safely
}
\keyword{internal}
