% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{reduce_dims}
\alias{reduce_dims}
\title{Perform PCA and TSNE dimensionality reduction}
\usage{
reduce_dims(sce)
}
\arguments{
\item{sce}{}
}
\value{

}
\description{
Perform PCA and TSNE dimensionality reduction
}
