% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-ds-utils.R
\name{remove_dataset_files}
\alias{remove_dataset_files}
\title{Delete stale dataset files after changing sample groups or running sva}
\usage{
remove_dataset_files(
  data_dir,
  patterns = c("^vsd.rds$", "^adjusted_\\\\d+svs.rds$", "^iqr_keep_\\\\d+svs.rds$",
    "^svobj.rds$", "^adjusted.rds$", "^numsv.rds$", "^lm_fit_\\\\d+svs.rds$",
    "cmap_res_.+.rds$", "l1000_drugs_res_.+.rds$", "l1000_genes_res_.+.rds$",
    "diff_expr_symbol_.+.rds$")
)
}
\arguments{
\item{data_dir}{Path to folder with dataset files}

\item{patterns}{patterns to remove. Default is all.}
}
\description{
Delete stale dataset files after changing sample groups or running sva
}
\keyword{internal}
