% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sva.R
\name{remove_sva}
\alias{remove_sva}
\title{Run surrogate variable analysis and factor out surrogate variables}
\usage{
remove_sva(eset)
}
\arguments{
\item{eset}{ExpressionSet with group column in \code{pData(eset)}}
}
\value{
matrix of vsd (if RNA-seq) RMA (microarray) normalized expression with
 surrogate variables factored out
}
\description{
Returned result usefull for heatmaps, MDS plots, machine learning, etc
}
