% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_alevin.R
\name{run_alevin}
\alias{run_alevin}
\title{Runs salmon alevin single-cell quantification.}
\usage{
run_alevin(data_dir, species = "human", overwrite = FALSE)
}
\arguments{
\item{data_dir}{Directory with raw 10X single-cell fastq.gz RNA-Seq files.}

\item{species}{Species name. Default is \code{human}.
Used to determine transcriptome index to use.}

\item{overwrite}{Do you want to overwrite results of previous run? Default is \code{FALSE}.}
}
\description{
Runs salmon alevin single-cell quantification.
}
\examples{

# first place data in data-raw/single-cell/example-data
data_dir <- 'data-raw/single-cell/example-data/Run2644-10X-Lung/10X_FID12518_Normal_3hg'
run_alevin(data_dir)

# another sample
data_dir <- 'data-raw/single-cell/example-data/Run2643-10X-Lung/10X_FID12518_Diseased_3hg'
run_alevin(data_dir)

}
