% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{run_drugseqr}
\alias{run_drugseqr}
\title{Run drugseqr app}
\usage{
run_drugseqr(
  app_name,
  data_dir = "/srv/drugseqr",
  app_dir = system.file("app", package = "drugseqr", mustWork = TRUE),
  pert_query_dir = file.path(data_dir, "pert_query_dir"),
  pert_signature_dir = file.path(data_dir, "pert_signature_dir"),
  indices_dir = file.path(data_dir, "indices"),
  tabs = c("Single Cell", "Bulk Data", "Drugs"),
  test = FALSE,
  test_data = FALSE,
  host = "0.0.0.0",
  port = 3838
)
}
\arguments{
\item{data_dir}{Directory containing drugseqr shiny app files.}

\item{pert_query_dir}{Path to directory where pert query results (using CMAP02/L1000 as query signature) will be downloaded as requested.}

\item{pert_signature_dir}{Path to directory where pert signatures for CMAP02/L1000 will be downloaded as requested.}

\item{indices_dir}{Path to directory containing \code{kallisto} indices and whitelists.}

\item{tabs}{Character vector of tabs to include in order desired. Must be subset of 'Single Cell', 'Bulk Data', and 'Drugs'.}

\item{test}{Boolean indicating if \code{shinytest} should be run (default is \code{FALSE}).
If \code{TRUE} test data will be used.}

\item{test_data}{Boolean indicating if test data should be used. Default is \code{TRUE}}
}
\description{
Used for local development
}
\examples{



# override default data_dir etc for development
app_name <- 'sjia'
app_dir <- 'inst/app'
data_dir <- '~/patient_data'
pert_query_dir <- '~/pert_query_dir'
pert_signature_dir <- '~/pert_signature_dir'
indices_dir <- '/srv/drugseqr/indices'

drugseqr::run_drugseqr(app_name, data_dir, app_dir, pert_query_dir, pert_signature_dir, indices_dir, port = 3839)

}
