% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{run_drugseqr}
\alias{run_drugseqr}
\title{Run drugseqr app}
\usage{
run_drugseqr(
  data_dir,
  app_dir = system.file("app", package = "drugseqr", mustWork = TRUE),
  pert_query_dir = "/srv/drugseqr/pert_query_dir",
  pert_signature_dir = "/srv/drugseqr/pert_signature_dir",
  indices_dir = "/srv/drugseqr/indices",
  test = FALSE,
  test_data = FALSE,
  host = "0.0.0.0",
  port = 3838
)
}
\arguments{
\item{data_dir}{Directory containing folders \code{'bulk'}, \code{'single-cell'}, and \code{'custom_queries'}.
Ignored if \code{test_data} is \code{TRUE}.}

\item{app_dir}{Directory containing drugseqr shiny app files.}

\item{pert_query_dir}{Path to directory where pert query results (using CMAP02/L1000 as query signature) will be downloaded as requested.}

\item{test}{Boolean indicating if \code{shinytest} should be run (default is \code{FALSE}).
If \code{TRUE} test data will be used.}

\item{test_data}{Boolean indicating if test data should be used. Default is \code{TRUE}}
}
\description{
Used for local development
}
\examples{

library(drugseqr)

# override default app_dir etc for development
app_dir <- 'inst/app'
data_dir <- 'data-raw/patient_data/sjia'
pert_query_dir <- 'data-raw/drug_gene_queries/data'
pert_signature_dir <- 'data-raw/drug_es/signatures'

run_drugseqr(data_dir, app_dir, pert_query_dir, pert_signature_dir, test_data = FALSE)
run_drugseqr(data_dir, app_dir, pert_query_dir, pert_signature_dir, test_data = TRUE)
run_drugseqr(data_dir, app_dir, pert_query_dir, pert_signature_dir, test = TRUE)

}
