% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drugseqr.R
\name{run_drugseqr}
\alias{run_drugseqr}
\title{Run drugseqr app}
\usage{
run_drugseqr(base_dir, test = FALSE, test_data = TRUE)
}
\arguments{
\item{base_dir}{Directory containing folders \code{'bulk'} and \code{'single-cell'}.
Ignored if \code{test_data} is \code{TRUE}.}

\item{test}{Boolean indicating if \code{shinytest} should be run (default is \code{FALSE}).
If \code{TRUE} test data will be used.}

\item{test_data}{Boolean indicating if test data should be used. Default is \code{TRUE}}
}
\description{
Used for local development
}
\examples{

base_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/IBD_20190712'
base_dir <- '~/Documents/Batcave/zaklab/drugseqr/data-raw/IBD'

run_drugseqr(base_dir, test_data = FALSE)

}
