% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dseqr.R
\name{run_dseqr}
\alias{run_dseqr}
\title{Run dseqr app}
\usage{
run_dseqr(
  app_name,
  data_dir,
  tabs = c("Single Cell", "Bulk Data", "Drugs"),
  pert_query_dir = file.path(data_dir, ".pert_query_dir"),
  pert_signature_dir = file.path(data_dir, ".pert_signature_dir"),
  gs_dir = file.path(data_dir, ".gs_dir"),
  indices_dir = file.path(data_dir, ".indices_dir"),
  tx2gene_dir = file.path(data_dir, ".tx2gene_dir"),
  app_dir = system.file("app", package = "dseqr", mustWork = TRUE),
  host = "0.0.0.0",
  port = 3838,
  logout_url = NULL,
  is_local = is.null(logout_url),
  test = FALSE,
  is_example = FALSE
)
}
\arguments{
\item{app_name}{Name of user folder in \code{data_dir}. Will be created if doesn't exist.}

\item{data_dir}{Directory containing app folders. By default also will contain folders
\code{.pert_query_dir}, \code{.pert_signature_dir}, and \code{.indices_dir}.}

\item{tabs}{Character vector of tabs to include in order desired. Must be subset of 'Single Cell', 'Bulk Data', and 'Drugs'.}

\item{pert_query_dir}{Path to directory where pert query results (using CMAP02/L1000 as query signature) will be downloaded as requested.}

\item{pert_signature_dir}{Path to directory where pert signatures for CMAP02/L1000 will be downloaded as requested.}

\item{gs_dir}{Path to directory where gene to Gene Ontology maps produced by \link{get_genego} are saved.}

\item{indices_dir}{Path to directory containing \code{kallisto} indices and whitelists.}

\item{tx2gene_dir}{Path to directory containing transcript to gene maps
produced by \link[dseqr.data]{load_tx2gene}.}

\item{app_dir}{Directory containing shiny app files. Default is to use 'app' directory of
installed dseqr package. Can be 'inst/app' if working from source code.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port between 3000:8000, excluding ports that are blocked
by Google Chrome for being considered unsafe: 3659, 4045, 5060,
5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random
ports will be tried.}

\item{logout_url}{URL used to log users out. Default is \code{NULL} for local use.}

\item{is_local}{Is dseqr running locally? If \code{FALSE}, uses CDNs for
some dependencies and sets up other dseqr.com specific tags. Default is \code{TRUE}
if \code{logout_url} is \code{NULL}, otherwise \code{FALSE}.}

\item{test}{Create a test with shinytest2? Default is \code{FALSE}.}

\item{is_example}{Is the app for demonstration purposes? If \code{TRUE}, various features
are disabled to prevent dataset changes.}
}
\value{
Runs dseqr app
}
\description{
Run dseqr application to explore single-cell and bulk RNA-seq datasets.
}
\examples{

if (interactive()) {

  data_dir <- tempdir()
  app_name <- 'example'
  run_dseqr(app_name, data_dir)
}

}
