% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_integrate_saved_scseqs}
\alias{run_integrate_saved_scseqs}
\title{Integration Utility Function for Background Process}
\usage{
run_integrate_saved_scseqs(
  sc_dir,
  dataset_names,
  integration_name,
  integration_types,
  ref_name
)
}
\arguments{
\item{sc_dir}{Directory with saved single-cell datasets.}

\item{integration_name}{Name for new integrated analysis.}

\item{integration_types}{Character vector of integration types to run.}
}
\value{
TRUE is successful, otherwise FALSE
}
\description{
Used as background process via \link{callr::r_bg}
}
\keyword{internal}
