% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_bulk.R
\name{run_kallisto_bulk}
\alias{run_kallisto_bulk}
\title{Runs kallisto quantification for bulk samples.}
\usage{
run_kallisto_bulk(
  indices_dir,
  data_dir,
  paired = NULL,
  pdata = NULL,
  species = "homo_sapiens",
  release = "94",
  fl.mean = NULL,
  fl.sd = NULL,
  updateProgress = NULL
)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices. See \code{\link{build_kallisto_index}}.}

\item{data_dir}{Directory with raw fastq.gz RNA-Seq files.}

\item{pdata}{Previous result of call to \code{run_kallisto_bulk} or \code{\link{select_pairs}}. Used to bypass another call to \code{select_pairs}.}

\item{species}{Species name. Default is \code{homo_sapiens}.
Used to determine transcriptome index to use.}

\item{fl.mean}{Estimated average fragment length (only relevant for single-end reads).}

\item{fl.sd}{Estimated standard deviation of fragment length (only relevant for single-end reads).}
}
\description{
For pair-ended experiments, reads for each pair should be in a seperate file.
}
\examples{

# first place IBD data in data-raw/example-data
indices_dir <- '/srv/drugseqr/indices'
data_dir <- file.path('data-raw', 'example-data')
run_kallisto_bulk(indices_dir, data_dir)

}
