% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{run_kallisto_scseq}
\alias{run_kallisto_scseq}
\title{Run kallisto/bustools for quantifying 10X scRNA-seq data}
\usage{
run_kallisto_scseq(
  indices_dir,
  data_dir,
  out_dir = file.path(data_dir, "bus_output"),
  bus_args = "-t 4",
  species = "homo_sapiens",
  release = "94",
  recount = TRUE,
  threads = 1
)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices built with \code{build_kallisto_index}.}

\item{data_dir}{Path to folder with 10X fastq.gz scRNA-seq files}

\item{out_dir}{Path to save output to.}

\item{bus_args}{Character vector of arguments to bustools.}

\item{species}{The species. Default is \code{'homo_sapiens'}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for
AnnotationHub - needs to match with
\code{\link[dseqr.data]{build_ensdb}}).}

\item{recount}{Overwrite previous output? Default is \code{TRUE}.}

\item{threads}{Number of threads to use. Default is 1.}
}
\value{
called for side effects.
}
\description{
Run kallisto/bustools for quantifying 10X scRNA-seq data
}
\seealso{
\link[rkal]{build_kallisto_index}
}
