% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{run_kallisto_scseq}
\alias{run_kallisto_scseq}
\title{Run kallisto/bustools for quantifying 10X scRNA-seq data}
\usage{
run_kallisto_scseq(indices_dir, data_dir, bus_args = c("-x 10xv2",
  "-t 4"), species = "homo_sapiens", release = "94")
}
\arguments{
\item{indices_dir}{Directory with kallisto indices. See \code{\link{build_kallisto_index}}.}

\item{data_dir}{Path to folder with 10X fastq.gz scRNA-seq files}

\item{bus_args}{Character vector of arguments to bustools.}

\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for AnnotationHub -
needs to match with \code{\link{build_ensdb}}).}
}
\description{
Run kallisto/bustools for quantifying 10X scRNA-seq data
}
\examples{

indices_dir <- 'data-raw/indices/kallisto'
data_dir <- 'data-raw/single-cell/example-data/Run2643-10X-Lung/10X_FID12518_Diseased_3hg'

run_kallisto_scseq(indices_dir, data_dir)

}
