% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kallisto_scseq.R
\name{run_kallisto_scseq}
\alias{run_kallisto_scseq}
\title{Run kallisto/bustools for quantifying 10X scRNA-seq data}
\usage{
run_kallisto_scseq(
  indices_dir,
  data_dir,
  bus_args = "-t 4",
  species = "homo_sapiens",
  release = "94",
  recount = TRUE
)
}
\arguments{
\item{indices_dir}{Directory with kallisto indices.}

\item{data_dir}{Path to folder with 10X fastq.gz scRNA-seq files}

\item{bus_args}{Character vector of arguments to bustools.}

\item{species}{The species. Default is \code{homo_sapiens}.}

\item{release}{ensembl release. Default is \code{94} (latest in release for
AnnotationHub - needs to match with
\code{\link[drugseqr.data]{build_ensdb}}).}

\item{recount}{Overwrite previous quantification?}
}
\description{
Run kallisto/bustools for quantifying 10X scRNA-seq data
}
\examples{

indices_dir <- 'data-raw/indices'
data_dir <- 'data-raw/single-cell/example-data/Run2643-10X-Lung/10X_FID12518_Diseased_3hg'

run_kallisto_scseq(indices_dir, data_dir)

}
\keyword{internal}
