% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{run_limma_pbulk}
\alias{run_limma_pbulk}
\title{Run pseudo bulk limma fit}
\usage{
run_limma_pbulk(scseq, dataset_dir, clusters_name)
}
\arguments{
\item{scseq}{\code{Seurat} object}

\item{dataset_dir}{Path to folder to save fit result in.}

\item{clusters_name}{String with comma seperated selected clusters.}
}
\description{
Run pseudo bulk limma fit
}
