% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{run_limma_scseq}
\alias{run_limma_scseq}
\title{Run limma fit for clusters in single cell RNA-Seq dataset}
\usage{
run_limma_scseq(obj, annot = "gene_name")
}
\arguments{
\item{obj}{\code{SingleCellExperiment} or list of \code{ExpressionSet} objects for pseudobulk.}

\item{annot}{Annotation to use. Default \code{'gene_name'} works for both mouse and human.}
}
\value{
Named list with slots: \itemize{
 \item fit result of \link[limma]{lmFit}.
 \item mod \code{model.matrix} used for \code{fit}.
}
}
\description{
Run limma fit for clusters in single cell RNA-Seq dataset
}
\keyword{internal}
