% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R
\name{run_limma_scseq}
\alias{run_limma_scseq}
\title{Run limma fit for clusters in single cell RNA-Seq dataset}
\usage{
run_limma_scseq(obj)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} object}

\item{dataset_dir}{Path to folder to save results and pseudobulk eset to}

\item{is_summed}{Boolean indicating if \code{obj} is a pseudobulk \code{SingleCellExperiment} object.}
}
\value{
Named list with slots: \itemize{
 \item fit result of \link[limma]{lmFit}.
 \item mod \code{model.matrix} used for \code{fit}.
}
}
\description{
Run limma fit for clusters in single cell RNA-Seq dataset
}
\keyword{internal}
