% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-utils.R, R/modules-sc-utils.R
\name{run_limma_scseq}
\alias{run_limma_scseq}
\title{Fit limma eBayes on single cell RNA-seq dataset}
\usage{
run_limma_scseq(obj, selected_clusters, dataset_dir)

run_limma_scseq(obj, selected_clusters, dataset_dir)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} or \code{ExpressionSet} for pseudobulk.}

\item{selected_clusters}{Character vector of selected clusters to run comparison for.}

\item{scseq}{\code{Seurat} object.}

\item{ident.1}{String with control group name in \code{Idents(scseq)}.}

\item{ident.2}{String with test group name in \code{Idents(scseq)}.}

\item{sc_dir}{Path to folder with single cell analyses.}

\item{anal_name}{Folder name in \code{sc_dir} that contains single cell analysis.}

\item{res_paths}{List of paths to saved results for markers of cells in \code{selected_clusters} and
differential expression analysis comparing test to control cells in \code{selected_clusters}.
Paths are stored in \code{'markers'} and \code{'anal'} slots.}
}
\value{
result of call to \code{\link[limma]{eBayes}}

List with slots \code{'markers'} and \code{'anal'} containing results for markers of
 cells in \code{selected_clusters} and differential expression analysis comparing test to control
 cells in \code{selected_clusters}.
}
\description{
Used by \code{run_comparison} for differential expression analysis.
}
\keyword{internal}
