% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_limma_scseq}
\alias{run_limma_scseq}
\title{Run single cell RNA-Seq test vs ctrl cluster comparison}
\usage{
run_limma_scseq(scseq, selected_clusters, dataset_dir)
}
\arguments{
\item{scseq}{\code{Suerat} object}

\item{selected_clusters}{Character vector of selected clusters to run comparison for.}

\item{sc_dir}{Path to folder with single cell analyses.}

\item{anal_name}{Folder name in \code{sc_dir} that contains single cell analysis.}

\item{res_paths}{List of paths to saved results for markers of cells in \code{selected_clusters} and
differential expression analysis comparing test to control cells in \code{selected_clusters}.
Paths are stored in \code{'markers'} and \code{'anal'} slots.}
}
\value{
List with slots \code{'markers'} and \code{'anal'} containing results for markers of
 cells in \code{selected_clusters} and differential expression analysis comparing test to control
 cells in \code{selected_clusters}.
}
\description{
Used by \code{run_comparison} for differential expression analysis.
}
