% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scseq.R
\name{run_pca}
\alias{run_pca}
\title{Cluster SingleCellExperiment}
\usage{
run_pca(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment}}
}
\value{
\code{sce} with column \code{cluster} in colData and \code{'npcs'} in metadata
}
\description{
Cluster SingleCellExperiment
}
