% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{run_post_cluster}
\alias{run_post_cluster}
\title{Post-Cluster Processing of SingleCellExperiment}
\usage{
run_post_cluster(
  scseq,
  dataset_name,
  sc_dir,
  resoln,
  progress = NULL,
  value = 0,
  reset_annot = TRUE
)
}
\arguments{
\item{scseq}{SingleCellExperiment}

\item{resoln}{resolution parameter}

\item{progress}{progress}

\item{value}{value}

\item{integrated}{is \code{scseq} integrated?}
}
\description{
Post-Cluster Processing of SingleCellExperiment
}
