% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_reduction}
\alias{run_reduction}
\title{Run TSNE or UMAP}
\usage{
run_reduction(sce, type = c("auto", "TSNE", "UMAP"), dimred = "PCA")
}
\arguments{
\item{sce}{\code{DietSCE}}

\item{dimred}{reducedDim to run TSNE or UMAP on}
}
\value{
\code{sce} with \code{'TSNE'} or \code{'UMAP'} \code{reducedDim}
}
\description{
Run TSNE or UMAP
}
