% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{run_salmon}
\alias{run_salmon}
\title{Runs salmon quantification.}
\usage{
run_salmon(data_dir, species = "homo_sapiens")
}
\arguments{
\item{data_dir}{Directory with raw RNA-Seq files.}

\item{species}{Species name. Default is \code{homo_sapiens}.
Used to determine transcriptome index to use.}
}
\description{
Runs salmon quantification.
}
\examples{

# first place IBD data in data-raw/example-data
data_dir <- file.path('data-raw', 'example-data')
run_salmon(data_dir)

}
