% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{run_salmon}
\alias{run_salmon}
\title{Runs salmon quantification.}
\usage{
run_salmon(data_dir, species = "homo_sapiens")
}
\arguments{
\item{data_dir}{Directory with .fastq.gz RNA-Seq files.}

\item{species}{Species name. Default is \code{homo_sapiens}.
Used to determine transcriptome index to use.}
}
\description{
Runs salmon quantification.
}
