% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scseq.R
\name{run_tsne}
\alias{run_tsne}
\title{Run TSNE for visualizing single cell data.}
\usage{
run_tsne(scseq)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object.}
}
\value{
\code{scseq} with TSNE results.
}
\description{
If \code{scseq} is a \code{SingleCellExperiment} object then uses \code{scater::runTSNE}.
If \code{scseq} is a \code{Seurat} object then uses \code{Seurat::RunTSNE}.
}
