% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{save_scseq_data}
\alias{save_scseq_data}
\title{Save Single Cell RNA-seq data for app}
\usage{
save_scseq_data(scseq_data, dataset_name, sc_dir, integrated = FALSE)
}
\arguments{
\item{scseq_data}{Named list with \code{scseq}, \code{markers}, and/or \code{annot}}

\item{dataset_name}{The analysis name.}

\item{sc_dir}{Path to directory with single-cell datasets.}

\item{integrated}{is the analysis integration. Default is \code{FALSE}}
}
\description{
Save Single Cell RNA-seq data for app
}
\keyword{internal}
