% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scseq_clusters.R
\name{save_scseq_data}
\alias{save_scseq_data}
\title{Save Single Cell RNA-seq data for app}
\usage{
save_scseq_data(scseq_data, anal_name, data_dir, integrated = FALSE)
}
\arguments{
\item{scseq_data}{Named list with \code{scseq}, \code{markers}, and/or \code{annot}}

\item{anal_name}{The analysis name.}

\item{data_dir}{Path to directory to save in}

\item{integrated}{is the analysis integration. Default is \code{FALSE}}
}
\description{
Save Single Cell RNA-seq data for app
}
