% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_reports.R
\name{save_scseq_reports}
\alias{save_scseq_reports}
\title{Save PDF of scRNA-seq reports}
\usage{
save_scseq_reports(scseq, markers, fname, point_size = 3)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat} object}

\item{markers}{Named list of \code{data.frame}s where \code{row.names} are the marker genes. One list per cluster in \code{scseq}
with the same name as the cluster.}

\item{fname}{String giving the name of the PDF file to save.}

\item{point_size}{Numeric scalar, specifying the size of the points. Defaults to 3.}
}
\value{
Saves report to \code{fname}.
}
\description{
Save PDF of scRNA-seq reports
}
