% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_reports.R
\name{save_scseq_reports}
\alias{save_scseq_reports}
\title{Save PDF of scRNA-seq reports}
\usage{
save_scseq_reports(scseq, markers, fname, pt.size = 3)
}
\arguments{
\item{scseq}{\code{SingleCellExperiment} or \code{Seurat}.}

\item{markers}{Named list of character vectors specifying the marker genes to plot for each cluster. The names are the
clusters that reports will be generated for in order.}

\item{fname}{String giving the name of the PDF file to save.}

\item{pt.size}{Numeric scalar, specifying the size of the points. Defaults to 3.}
}
\value{
Saves report to \code{fname}.
}
\description{
Save PDF of scRNA-seq reports
}
