% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_names.R
\name{scale_fill_names}
\alias{scale_fill_names}
\title{Get a fill scale with named values}
\usage{
scale_fill_names(levs)
}
\arguments{
\item{levs}{Character vector of levels to get a fill scale for.}
}
\value{
ggplot2 \code{scale_fill_manual} with values names from \code{levs}
}
\description{
Get a fill scale with named values
}
\keyword{internal}
