% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scClusterPlot}
\alias{scClusterPlot}
\title{Logic for cluster plots}
\usage{
scClusterPlot(
  input,
  output,
  session,
  scseq,
  cached_plot,
  fname_fun = function() { },
  downloadable = FALSE
)
}
\description{
Logic for cluster plots
}
\keyword{internal}
