% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scMarkerPlot}
\alias{scMarkerPlot}
\title{Logic for marker gene plots}
\usage{
scMarkerPlot(
  input,
  output,
  session,
  scseq,
  selected_gene,
  selected_group,
  cached_plot,
  fname_fun = function() { },
  downloadable = TRUE
)
}
\description{
Logic for marker gene plots
}
\keyword{internal}
