% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scMarkerPlot}
\alias{scMarkerPlot}
\title{Logic for marker feature plots}
\usage{
scMarkerPlot(
  input,
  output,
  session,
  scseq,
  selected_feature,
  dataset_name,
  downloadable = TRUE,
  custom_metrics = function() NULL
)
}
\description{
Logic for marker feature plots
}
\keyword{internal}
