% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scPageUI}
\alias{scPageUI}
\title{UI for Single Cell Exploration page}
\usage{
scPageUI(id, tab, active)

scPageUI(id, tab, active)
}
\description{
UI for Single Cell Exploration page

UI for Single Cell Exploration page
}
\keyword{internal}
