% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-path-server.R
\name{scSampleComparison}
\alias{scSampleComparison}
\title{Logic for single cell clusters selector for pathForm}
\usage{
scSampleComparison(
  input,
  output,
  session,
  data_dir,
  anal,
  is_sc,
  input_ids,
  with_path = FALSE,
  with_drugs = FALSE
)
}
\description{
Logic for single cell clusters selector for pathForm
}
\keyword{internal}
