% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{scSampleComparison}
\alias{scSampleComparison}
\title{Logic for single cell cluster analyses for Single Cell, Drugs, and Pathways tabs}
\usage{
scSampleComparison(
  input,
  output,
  session,
  dataset_dir,
  dataset_name,
  is.integrated = function() TRUE,
  input_scseq = function() NULL,
  is_sc = function() TRUE,
  exclude_ambient = function() FALSE,
  comparison_type = function() "samples"
)
}
\description{
Logic for single cell cluster analyses for Single Cell, Drugs, and Pathways tabs
}
\keyword{internal}
