% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scSampleComparisonInput}
\alias{scSampleComparisonInput}
\title{Input for Single Cell analysis}
\usage{
scSampleComparisonInput(id, with_dl = FALSE)
}
\description{
Used in Single Cell and Drugs tab
}
\keyword{internal}
