% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{scSampleComparisonInput}
\alias{scSampleComparisonInput}
\title{Input for Single Cell analysis}
\usage{
scSampleComparisonInput(id)

scSampleComparisonInput(id)
}
\description{
Used in Single Cell, Drugs and Pathways tab

Used in Single Cell, Drugs and Pathways tab
}
\keyword{internal}
