% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_contrast.R
\name{select_contrast}
\alias{select_contrast}
\title{Select contrast for an experiment.}
\usage{
select_contrast(eset)
}
\arguments{
\item{eset}{\code{ExpressionSet}}
}
\value{
\code{eset} with selected samples retained and column \code{group} with values \code{'control'} and \code{'test'}.
}
\description{
Used to select a contrast (control and test group samples) for an experiment.
}
\examples{

eset_path <- system.file('extdata', 'IBD', 'eset.rds', package='drugseqr')
eset <- readRDS(eset_path)

select_contrast(eset)


}
