% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pairs.R
\name{select_pairs}
\alias{select_pairs}
\title{Select pairs and replicates for an experiment.}
\usage{
select_pairs(data_dir, pdata_path)
}
\arguments{
\item{data_dir}{Directory with raw and quantified RNA-Seq files.}

\item{pdata_path}{Path to text file with sample annotations. Must be readable by \code{\link[data.table]{fread}}.
The first column should contain sample ids that match a single raw rna-seq data file name.}
}
\value{
data.frame of sample annotations with integer columns \code{Pair} (if paired-end) and \code{Replicate} (if added).
}
\description{
Used to select a pairs for paired-end experiments and replicate samples. Please follow prompt, ensuring correct file name matching and
end-type experiment identification.
}
\examples{

data_dir <- system.file('extdata', 'IBD', package='drugseqr')
pdata_path <- file.path(data_dir, 'Phenotypes.csv')

pdata <- select_pairs(data_dir, pdata_path)


}
