% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-server.R
\name{selectedAnal}
\alias{selectedAnal}
\title{Logic for selected dataset/analysis in Drugs tab}
\usage{
selectedAnal(
  input,
  output,
  session,
  data_dir,
  choices,
  new_custom,
  pert_query_dir = NULL
)
}
\description{
Logic for selected dataset/analysis in Drugs tab
}
\keyword{internal}
