% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{selectedAnal}
\alias{selectedAnal}
\title{Logic for selected analysis part of scForm}
\usage{
selectedAnal(input, output, session, sc_dir, new_anal)
}
\description{
Logic for selected analysis part of scForm
}
\keyword{internal}
