% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-ui.R
\name{selectedAnalInput}
\alias{selectedAnalInput}
\title{Input form/associated buttons for selecting single cell analysis}
\usage{
selectedAnalInput(id)
}
\description{
Input form/associated buttons for selecting single cell analysis
}
