% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-drugs-ui.R
\name{selectedAnalInput}
\alias{selectedAnalInput}
\title{Analysis input for Drugs page}
\usage{
selectedAnalInput(id, label = "Select a dataset or query signature:")
}
\description{
Analysis input for Drugs page
}
\keyword{internal}
