% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{selectedGene}
\alias{selectedGene}
\title{Logic for selected gene to show plots for}
\usage{
selectedGene(input, output, session, selected_anal, scseq,
  selected_markers, cluster_markers, selected_cluster, comparison_type)
}
\description{
Logic for selected gene to show plots for
}
\keyword{internal}
