% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-server.R
\name{selectedGene}
\alias{selectedGene}
\title{Logic for selected gene to show plots for}
\usage{
selectedGene(
  input,
  output,
  session,
  dataset_name,
  selected_markers,
  selected_cluster,
  comparison_type,
  ambient = function() NULL
)
}
\description{
Logic for selected gene to show plots for
}
\keyword{internal}
