% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{setup_fdata}
\alias{setup_fdata}
\title{Setup feature annotation data}
\usage{
setup_fdata(species = "Homo sapiens", release = "94")
}
\value{
\code{data.table} with columns \code{SYMBOL} and \code{ENTREZID_HS} corresponding to
  HGNC symbols and human entrez ids respectively.
}
\description{
Setup feature annotation data
}
\keyword{internal}
