% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_scseq.R
\name{strain_scseq}
\alias{strain_scseq}
\title{Remove ambient gene expression from sc-RNAseq counts}
\usage{
strain_scseq(counts, empty)
}
\arguments{
\item{counts}{sparse dgTMatrix of counts.}

\item{empty}{Boolean indicating columns in \code{counts} that are empty droplets.}
}
\value{
\code{counts} corrected for ambient contamination.
}
\description{
Uses \code{SoupX} to estimate and adjust counts for ambient contamination.
Currently only works with kallisto quantification results (and needs some work).
}
