% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{subset_saved_scseq}
\alias{subset_saved_scseq}
\title{Subset SingleCellExperiment}
\usage{
subset_saved_scseq(
  sc_dir,
  founder,
  from_dataset,
  dataset_name,
  exclude_clusters = NULL,
  subset_metrics = NULL,
  is_integrated = FALSE,
  is_include = FALSE,
  progress = NULL,
  hvgs = NULL,
  ref_name = NULL
)
}
\arguments{
\item{sc_dir}{Directory with saved single-cell datasets.}

\item{founder}{Name of original founding dataset (may be ancestor of \code{from_dataset}).}

\item{from_dataset}{Name of parent dataset.}

\item{dataset_name}{Name for new subset dataset.}

\item{exclude_clusters}{Charactor vector of clusters for excluding cells.}

\item{subset_metrics}{Character vector of metrics for excluding cells.}

\item{progress}{Optional shiny progress object.}

\item{exclude_cells}{Character vector of cells to exclude. Used for integrated datasets to exclude clusters.}
}
\value{
TRUE is successful, otherwise FALSE
}
\description{
Subset SingleCellExperiment
}
\keyword{internal}
