% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scseq_reports.R
\name{theme_dimgray}
\alias{theme_dimgray}
\title{Make ggplot axes and text dimgray}
\usage{
theme_dimgray(with_nums = TRUE)
}
\arguments{
\item{with_nums}{Include axis ticks/text? Default is TRUE.}
}
\value{
\code{theme}
}
\description{
Make ggplot axes and text dimgray
}
\keyword{internal}
