% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scseq.R
\name{theme_dimgray}
\alias{theme_dimgray}
\title{Make ggplot axes and text dimgray}
\usage{
theme_dimgray(plot)
}
\arguments{
\item{plot}{ggplot}
}
\value{
ggplot
}
\description{
Make ggplot axes and text dimgray
}
