% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-sc-utils.R
\name{validate_integration}
\alias{validate_integration}
\title{Validate dataset selection for integration}
\usage{
validate_integration(test, ctrl, anal_name, anal_options, pairs)
}
\arguments{
\item{test}{Character vector of test dataset names}

\item{ctrl}{Character vector of control dataset names}
}
\value{
\code{NULL} if valid, otherwise an error message
}
\description{
Validate dataset selection for integration
}
\keyword{internal}
