% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zakquery.R
\name{zakquery}
\alias{zakquery}
\title{Run Drug Query with Zak Specifications}
\usage{
zakquery(
  query_res,
  drug_study = c("CMAP02", "L1000 Drugs", "L1000 Genes"),
  cell_types = c("all", "haematopoietic", "monocytic", "lung"),
  min_signatures = 1,
  interactive = TRUE
)
}
\arguments{
\item{query_res}{Numeric vector of correlation values with names as HGNC symbols}

\item{drug_study}{Queried database giving rise to \code{query_res}.
One of \code{'CMAP02'},\code{'L1000 Genetic'}, or \code{'L1000 Drugs'}.}

\item{min_signatures}{Number of independent perturbagen signatures below which a perturbation is excluded. Default is \code{3}.}

\item{interactive}{if \code{TRUE} (default) then starts shiny app with first \code{cell_types}. Otherwise returns
data.frames with results for all \code{cell_types}.}
}
\value{
List of data.frames or result of run_drugseqr_table
}
\description{
Only clinical drugs, cells types: all, haematopoietic, monocytic, and lung
}
\keyword{internal}
